@@ -12,61 +12,61 @@
   },
   "author": "Moeen Basra",
   "license": "ISC",
-  "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.7.4",
-    "@babel/preset-env": "^7.7.4",
-    "@babel/preset-react": "^7.7.4",
-    "babel-eslint": "^10.0.2",
-    "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "browser-sync": "^2.26.7",
-    "browser-sync-webpack-plugin": "^2.2.2",
-    "cross-env": "^5.2.0",
-    "eslint": "^6.7.1",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-friendly-formatter": "^4.0.1",
-    "eslint-loader": "^2.1.2",
-    "eslint-plugin-html": "^6.0.0",
-    "eslint-plugin-import": "^2.18.0",
-    "eslint-plugin-node": "^9.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-standard": "^4.0.0",
-    "immutable": "^4.0.0-rc.12",
-    "laravel-mix": "^4.0.16",
-    "redux-logger": "^3.0.6",
-    "resolve-url-loader": "^3.1.1",
-    "sass": "^1.23.7",
-    "sass-loader": "^7.1.0"
-  },
-  "dependencies": {
-    "axios": "^0.19.0",
-    "bootstrap": "^4.4.0",
-    "classnames": "^2.2.6",
-    "draft-js": "^0.10.5",
-    "draftjs-to-html": "^0.8.4",
-    "font-awesome": "^4.7.0",
-    "history": "^4.9.0",
-    "jquery": "^3.4.1",
-    "lodash": "^4.17.11",
-    "moment": "^2.24.0",
-    "popper.js": "^1.16.0",
-    "prop-types": "^15.7.2",
-    "react": "^16.12.0",
-    "react-document-title": "^2.0.3",
-    "react-dom": "^16.12.0",
-    "react-draft-wysiwyg": "^1.13.2",
-    "react-loadable": "^5.5.0",
-    "react-redux": "^7.1.3",
-    "react-router-dom": "^5.1.2",
-    "reactstrap": "^8.1.1",
-    "redux": "^4.0.1",
-    "redux-thunk": "^2.3.0",
-    "ree-validate": "^3.0.2"
-  },
   "browserslist": [
     ">0.2%",
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ]
+  ],
+  "devDependencies": {
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/preset-env": "^7.8.7",
+    "@babel/preset-react": "^7.8.3",
+    "babel-eslint": "^10.1.0",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "browser-sync": "^2.26.7",
+    "browser-sync-webpack-plugin": "^2.2.2",
+    "cross-env": "^7.0.2",
+    "eslint": "^6.8.0",
+    "eslint-config-standard": "^14.1.0",
+    "eslint-friendly-formatter": "^4.0.1",
+    "eslint-loader": "^3.0.3",
+    "eslint-plugin-html": "^6.0.0",
+    "eslint-plugin-import": "^2.20.1",
+    "eslint-plugin-node": "^11.0.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.18.3",
+    "eslint-plugin-standard": "^4.0.1",
+    "immutable": "^4.0.0-rc.12",
+    "laravel-mix": "^5.0.1",
+    "redux-logger": "^3.0.6",
+    "resolve-url-loader": "^3.1.1",
+    "sass": "^1.26.2",
+    "sass-loader": "^8.0.2"
+  },
+  "dependencies": {
+    "axios": "^0.19.2",
+    "bootstrap": "^4.4.1",
+    "classnames": "^2.2.6",
+    "draft-js": "^0.11.4",
+    "draftjs-to-html": "^0.9.1",
+    "font-awesome": "^4.7.0",
+    "history": "^4.10.1",
+    "jquery": "^3.4.1",
+    "lodash": "^4.17.15",
+    "moment": "^2.24.0",
+    "popper.js": "^1.16.1",
+    "prop-types": "^15.7.2",
+    "react": "^16.13.0",
+    "react-document-title": "^2.0.3",
+    "react-dom": "^16.13.0",
+    "react-draft-wysiwyg": "^1.14.4",
+    "react-loadable": "^5.5.0",
+    "react-redux": "^7.2.0",
+    "react-router-dom": "^5.1.2",
+    "reactstrap": "^8.4.1",
+    "redux": "^4.0.5",
+    "redux-thunk": "^2.3.0",
+    "ree-validate": "^3.0.2"
+  }
 }
