@@ -8,7 +8,7 @@
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\");\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var jquery__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! jquery */ \"./node_modules/jquery/dist/jquery.js\");\n/* harmony import */ var jquery__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(jquery__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var lodash__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! lodash */ \"./node_modules/lodash/lodash.js\");\n/* harmony import */ var lodash__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(lodash__WEBPACK_IMPORTED_MODULE_3__);\n/* harmony import */ var react_router_dom__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! react-router-dom */ \"./node_modules/react-router-dom/esm/react-router-dom.js\");\n/* harmony import */ var _service__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ../../service */ \"./resources/js/modules/auth/service.js\");\n/* harmony import */ var ree_validate__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ree-validate */ \"./node_modules/ree-validate/dist/ree-validate.esm.js\");\n/* harmony import */ var _components_Form__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./components/Form */ \"./resources/js/modules/auth/pages/register/components/Form.js\");\nfunction _typeof(obj) { if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }\n\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(source, true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(source).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\n//import libs\n\n\n\n\n\n\n // import components\n\n // initialize component\n\nvar Page =\n/*#__PURE__*/\nfunction (_Component) {\n  _inherits(Page, _Component);\n\n  function Page(props) {\n    var _this;\n\n    _classCallCheck(this, Page);\n\n    _this = _possibleConstructorReturn(this, _getPrototypeOf(Page).call(this, props));\n    _this.validator = new ree_validate__WEBPACK_IMPORTED_MODULE_6__[\"default\"]({\n      name: 'required|min:6',\n      email: 'required|email',\n      password: 'required|min:6',\n      passwordConfirmation: 'required|min:6'\n    });\n    _this.state = {\n      credentials: {\n        name: '',\n        email: '',\n        password: '',\n        passwordConfirmation: ''\n      },\n      errors: _this.validator.errors,\n      fields: _this.validator.fields\n    };\n    _this.handleChange = _this.handleChange.bind(_assertThisInitialized(_this));\n    _this.handleSubmit = _this.handleSubmit.bind(_assertThisInitialized(_this));\n    return _this;\n  }\n\n  _createClass(Page, [{\n    key: \"componentDidMount\",\n    value: function componentDidMount() {\n      jquery__WEBPACK_IMPORTED_MODULE_2___default()('body').attr('style', 'background-color: #eee');\n    }\n  }, {\n    key: \"componentWillUnmount\",\n    value: function componentWillUnmount() {\n      jquery__WEBPACK_IMPORTED_MODULE_2___default()('body').removeAttr('style');\n    } // event to handle input change\n\n  }, {\n    key: \"handleChange\",\n    value: function handleChange(name, value) {\n      var _this2 = this;\n\n      var errors = this.validator.errors;\n      this.setState({\n        credentials: _objectSpread({}, this.state.credentials, _defineProperty({}, name, value))\n      });\n      errors.remove(name);\n      this.validator.validate(name, value).then(function () {\n        _this2.setState({\n          errors: errors\n        });\n      });\n    }\n  }, {\n    key: \"handleSubmit\",\n    value: function handleSubmit(e) {\n      var _this3 = this;\n\n      e.preventDefault();\n      var credentials = this.state.credentials;\n      var errors = this.validator.errors;\n      this.validator.validateAll(credentials).then(function (success) {\n        if (success) {\n          _this3.submit(credentials);\n        } else {\n          _this3.setState({\n            errors: errors\n          });\n        }\n      });\n    }\n  }, {\n    key: \"submit\",\n    value: function submit(credentials) {\n      var _this4 = this;\n\n      this.props.dispatch(Object(_service__WEBPACK_IMPORTED_MODULE_5__[\"register\"])(credentials)).catch(function (_ref) {\n        var error = _ref.error,\n            statusCode = _ref.statusCode;\n        var errors = _this4.validator.errors;\n\n        if (statusCode === 422) {\n          lodash__WEBPACK_IMPORTED_MODULE_3___default.a.forOwn(error, function (message, field) {\n            errors.add(field, message);\n          });\n        } else if (statusCode === 401) {\n          errors.add('password', error);\n        }\n\n        _this4.setState({\n          errors: errors\n        });\n      });\n    }\n  }, {\n    key: \"render\",\n    value: function render() {\n      // check if user is authenticated then redirect him to home page\n      if (this.props.isAuthenticated) {\n        return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(react_router_dom__WEBPACK_IMPORTED_MODULE_4__[\"Redirect\"], {\n          to: \"/\"\n        });\n      }\n\n      var _this$state$credentia = this.state.credentials,\n          name = _this$state$credentia.name,\n          email = _this$state$credentia.email,\n          password = _this$state$credentia.password,\n          passwordConfirmation = _this$state$credentia.passwordConfirmation;\n      var props = {\n        name: name,\n        email: email,\n        password: password,\n        passwordConfirmation: passwordConfirmation,\n        errors: this.state.errors,\n        handleChange: this.handleChange,\n        handleSubmit: this.handleSubmit\n      };\n      return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n        className: \"container py-5\"\n      }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n        className: \"row\"\n      }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n        className: \"col-md-12\"\n      }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n        className: \"row\"\n      }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n        className: \"mx-auto\"\n      }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"span\", {\n        className: \"anchor\"\n      }), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n        className: \"card has-shadow\"\n      }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n        className: \"card-body\"\n      }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_components_Form__WEBPACK_IMPORTED_MODULE_7__[\"default\"], props))))))));\n    }\n  }]);\n\n  return Page;\n}(react__WEBPACK_IMPORTED_MODULE_0__[\"Component\"]);\n\n_defineProperty(Page, \"displayName\", 'RegisterPage');\n\n_defineProperty(Page, \"propTypes\", {\n  isAuthenticated: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.bool.isRequired,\n  dispatch: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.func.isRequired\n});\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Page);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hdXRoL3BhZ2VzL3JlZ2lzdGVyL1BhZ2UuanM/NmRkNSJdLCJuYW1lcyI6WyJQYWdlIiwicHJvcHMiLCJ2YWxpZGF0b3IiLCJSZWVWYWxpZGF0ZSIsIm5hbWUiLCJlbWFpbCIsInBhc3N3b3JkIiwicGFzc3dvcmRDb25maXJtYXRpb24iLCJzdGF0ZSIsImNyZWRlbnRpYWxzIiwiZXJyb3JzIiwiZmllbGRzIiwiaGFuZGxlQ2hhbmdlIiwiYmluZCIsImhhbmRsZVN1Ym1pdCIsIiQiLCJhdHRyIiwicmVtb3ZlQXR0ciIsInZhbHVlIiwic2V0U3RhdGUiLCJyZW1vdmUiLCJ2YWxpZGF0ZSIsInRoZW4iLCJlIiwicHJldmVudERlZmF1bHQiLCJ2YWxpZGF0ZUFsbCIsInN1Y2Nlc3MiLCJzdWJtaXQiLCJkaXNwYXRjaCIsInJlZ2lzdGVyIiwiY2F0Y2giLCJlcnJvciIsInN0YXR1c0NvZGUiLCJfIiwiZm9yT3duIiwibWVzc2FnZSIsImZpZWxkIiwiYWRkIiwiaXNBdXRoZW50aWNhdGVkIiwiQ29tcG9uZW50IiwiUHJvcFR5cGVzIiwiYm9vbCIsImlzUmVxdWlyZWQiLCJmdW5jIl0sIm1hcHBpbmdzIjoiOzs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7O0FBQUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7Q0FHQTs7Q0FHQTs7SUFDTUEsSTs7Ozs7QUFPSixnQkFBWUMsS0FBWixFQUFtQjtBQUFBOztBQUFBOztBQUNqQiw4RUFBTUEsS0FBTjtBQUVBLFVBQUtDLFNBQUwsR0FBaUIsSUFBSUMsb0RBQUosQ0FBZ0I7QUFDL0JDLFVBQUksRUFBRSxnQkFEeUI7QUFFL0JDLFdBQUssRUFBRSxnQkFGd0I7QUFHL0JDLGNBQVEsRUFBRSxnQkFIcUI7QUFJL0JDLDBCQUFvQixFQUFFO0FBSlMsS0FBaEIsQ0FBakI7QUFPQSxVQUFLQyxLQUFMLEdBQWE7QUFDWEMsaUJBQVcsRUFBRTtBQUNYTCxZQUFJLEVBQUUsRUFESztBQUVYQyxhQUFLLEVBQUUsRUFGSTtBQUdYQyxnQkFBUSxFQUFFLEVBSEM7QUFJWEMsNEJBQW9CLEVBQUU7QUFKWCxPQURGO0FBT1hHLFlBQU0sRUFBRSxNQUFLUixTQUFMLENBQWVRLE1BUFo7QUFRWEMsWUFBTSxFQUFFLE1BQUtULFNBQUwsQ0FBZVM7QUFSWixLQUFiO0FBV0EsVUFBS0MsWUFBTCxHQUFvQixNQUFLQSxZQUFMLENBQWtCQyxJQUFsQiwrQkFBcEI7QUFDQSxVQUFLQyxZQUFMLEdBQW9CLE1BQUtBLFlBQUwsQ0FBa0JELElBQWxCLCtCQUFwQjtBQXRCaUI7QUF1QmxCOzs7O3dDQUVtQjtBQUNsQkUsbURBQUMsQ0FBQyxNQUFELENBQUQsQ0FBVUMsSUFBVixDQUFlLE9BQWYsRUFBd0Isd0JBQXhCO0FBQ0Q7OzsyQ0FFc0I7QUFDckJELG1EQUFDLENBQUMsTUFBRCxDQUFELENBQVVFLFVBQVYsQ0FBcUIsT0FBckI7QUFDRCxLLENBRUQ7Ozs7aUNBQ2FiLEksRUFBTWMsSyxFQUFPO0FBQUE7O0FBQUEsVUFDaEJSLE1BRGdCLEdBQ0wsS0FBS1IsU0FEQSxDQUNoQlEsTUFEZ0I7QUFHeEIsV0FBS1MsUUFBTCxDQUFjO0FBQUNWLG1CQUFXLG9CQUFPLEtBQUtELEtBQUwsQ0FBV0MsV0FBbEIsc0JBQWdDTCxJQUFoQyxFQUF1Q2MsS0FBdkM7QUFBWixPQUFkO0FBQ0FSLFlBQU0sQ0FBQ1UsTUFBUCxDQUFjaEIsSUFBZDtBQUVBLFdBQUtGLFNBQUwsQ0FBZW1CLFFBQWYsQ0FBd0JqQixJQUF4QixFQUE4QmMsS0FBOUIsRUFDR0ksSUFESCxDQUNRLFlBQU07QUFDVixjQUFJLENBQUNILFFBQUwsQ0FBYztBQUFFVCxnQkFBTSxFQUFOQTtBQUFGLFNBQWQ7QUFDRCxPQUhIO0FBSUQ7OztpQ0FFWWEsQyxFQUFHO0FBQUE7O0FBQ2RBLE9BQUMsQ0FBQ0MsY0FBRjtBQURjLFVBRU5mLFdBRk0sR0FFVSxLQUFLRCxLQUZmLENBRU5DLFdBRk07QUFBQSxVQUdOQyxNQUhNLEdBR0ssS0FBS1IsU0FIVixDQUdOUSxNQUhNO0FBS2QsV0FBS1IsU0FBTCxDQUFldUIsV0FBZixDQUEyQmhCLFdBQTNCLEVBQ0dhLElBREgsQ0FDUSxVQUFDSSxPQUFELEVBQWE7QUFDakIsWUFBSUEsT0FBSixFQUFhO0FBQ1gsZ0JBQUksQ0FBQ0MsTUFBTCxDQUFZbEIsV0FBWjtBQUNELFNBRkQsTUFFTztBQUNMLGdCQUFJLENBQUNVLFFBQUwsQ0FBYztBQUFFVCxrQkFBTSxFQUFOQTtBQUFGLFdBQWQ7QUFDRDtBQUNGLE9BUEg7QUFRRDs7OzJCQUVNRCxXLEVBQWE7QUFBQTs7QUFDbEIsV0FBS1IsS0FBTCxDQUFXMkIsUUFBWCxDQUFvQkMseURBQVEsQ0FBQ3BCLFdBQUQsQ0FBNUIsRUFDR3FCLEtBREgsQ0FDUyxnQkFBMkI7QUFBQSxZQUF4QkMsS0FBd0IsUUFBeEJBLEtBQXdCO0FBQUEsWUFBakJDLFVBQWlCLFFBQWpCQSxVQUFpQjtBQUFBLFlBQ3hCdEIsTUFEd0IsR0FDYixNQUFJLENBQUNSLFNBRFEsQ0FDeEJRLE1BRHdCOztBQUdoQyxZQUFJc0IsVUFBVSxLQUFLLEdBQW5CLEVBQXdCO0FBQ3RCQyx1REFBQyxDQUFDQyxNQUFGLENBQVNILEtBQVQsRUFBZ0IsVUFBQ0ksT0FBRCxFQUFVQyxLQUFWLEVBQW9CO0FBQ2xDMUIsa0JBQU0sQ0FBQzJCLEdBQVAsQ0FBV0QsS0FBWCxFQUFrQkQsT0FBbEI7QUFDRCxXQUZEO0FBR0QsU0FKRCxNQUlPLElBQUlILFVBQVUsS0FBSyxHQUFuQixFQUF3QjtBQUM3QnRCLGdCQUFNLENBQUMyQixHQUFQLENBQVcsVUFBWCxFQUF1Qk4sS0FBdkI7QUFDRDs7QUFFRCxjQUFJLENBQUNaLFFBQUwsQ0FBYztBQUFFVCxnQkFBTSxFQUFOQTtBQUFGLFNBQWQ7QUFDRCxPQWJIO0FBY0Q7Ozs2QkFFUTtBQUNQO0FBQ0EsVUFBSSxLQUFLVCxLQUFMLENBQVdxQyxlQUFmLEVBQWdDO0FBQzlCLGVBQU8sMkRBQUMseURBQUQ7QUFBVSxZQUFFLEVBQUM7QUFBYixVQUFQO0FBQ0Q7O0FBSk0sa0NBTWlELEtBQUs5QixLQUFMLENBQVdDLFdBTjVEO0FBQUEsVUFNQ0wsSUFORCx5QkFNQ0EsSUFORDtBQUFBLFVBTU9DLEtBTlAseUJBTU9BLEtBTlA7QUFBQSxVQU1jQyxRQU5kLHlCQU1jQSxRQU5kO0FBQUEsVUFNd0JDLG9CQU54Qix5QkFNd0JBLG9CQU54QjtBQU9QLFVBQU1OLEtBQUssR0FBRztBQUNaRyxZQUFJLEVBQUpBLElBRFk7QUFFWkMsYUFBSyxFQUFMQSxLQUZZO0FBR1pDLGdCQUFRLEVBQVJBLFFBSFk7QUFJWkMsNEJBQW9CLEVBQXBCQSxvQkFKWTtBQUtaRyxjQUFNLEVBQUUsS0FBS0YsS0FBTCxDQUFXRSxNQUxQO0FBTVpFLG9CQUFZLEVBQUUsS0FBS0EsWUFOUDtBQU9aRSxvQkFBWSxFQUFFLEtBQUtBO0FBUFAsT0FBZDtBQVVBLGFBQVE7QUFBSyxpQkFBUyxFQUFDO0FBQWYsU0FDTjtBQUFLLGlCQUFTLEVBQUM7QUFBZixTQUNFO0FBQUssaUJBQVMsRUFBQztBQUFmLFNBQ0U7QUFBSyxpQkFBUyxFQUFDO0FBQWYsU0FDRTtBQUFLLGlCQUFTLEVBQUM7QUFBZixTQUNFO0FBQU0saUJBQVMsRUFBQztBQUFoQixRQURGLEVBRUU7QUFBSyxpQkFBUyxFQUFDO0FBQWYsU0FDRTtBQUFLLGlCQUFTLEVBQUM7QUFBZixTQUNFLDJEQUFDLHdEQUFELEVBQVViLEtBQVYsQ0FERixDQURGLENBRkYsQ0FERixDQURGLENBREYsQ0FETSxDQUFSO0FBZ0JEOzs7O0VBdEhnQnNDLCtDOztnQkFBYnZDLEksaUJBQ2lCLGM7O2dCQURqQkEsSSxlQUVlO0FBQ2pCc0MsaUJBQWUsRUFBRUUsaURBQVMsQ0FBQ0MsSUFBVixDQUFlQyxVQURmO0FBRWpCZCxVQUFRLEVBQUVZLGlEQUFTLENBQUNHLElBQVYsQ0FBZUQ7QUFGUixDOztBQXVITjFDLG1FQUFmIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvYXV0aC9wYWdlcy9yZWdpc3Rlci9QYWdlLmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiLy9pbXBvcnQgbGlic1xuaW1wb3J0IFJlYWN0LCB7IENvbXBvbmVudCB9IGZyb20gJ3JlYWN0J1xuaW1wb3J0IFByb3BUeXBlcyBmcm9tICdwcm9wLXR5cGVzJ1xuaW1wb3J0ICQgZnJvbSAnanF1ZXJ5J1xuaW1wb3J0IF8gZnJvbSAnbG9kYXNoJ1xuaW1wb3J0IHsgUmVkaXJlY3QgfSBmcm9tICdyZWFjdC1yb3V0ZXItZG9tJ1xuaW1wb3J0IHsgcmVnaXN0ZXIgfSBmcm9tICcuLi8uLi9zZXJ2aWNlJ1xuaW1wb3J0IFJlZVZhbGlkYXRlIGZyb20gJ3JlZS12YWxpZGF0ZSdcblxuLy8gaW1wb3J0IGNvbXBvbmVudHNcbmltcG9ydCBGb3JtIGZyb20gJy4vY29tcG9uZW50cy9Gb3JtJ1xuXG4vLyBpbml0aWFsaXplIGNvbXBvbmVudFxuY2xhc3MgUGFnZSBleHRlbmRzIENvbXBvbmVudCB7XG4gIHN0YXRpYyBkaXNwbGF5TmFtZSA9ICdSZWdpc3RlclBhZ2UnXG4gIHN0YXRpYyBwcm9wVHlwZXMgPSB7XG4gICAgaXNBdXRoZW50aWNhdGVkOiBQcm9wVHlwZXMuYm9vbC5pc1JlcXVpcmVkLFxuICAgIGRpc3BhdGNoOiBQcm9wVHlwZXMuZnVuYy5pc1JlcXVpcmVkLFxuICB9XG4gIFxuICBjb25zdHJ1Y3Rvcihwcm9wcykge1xuICAgIHN1cGVyKHByb3BzKVxuICAgIFxuICAgIHRoaXMudmFsaWRhdG9yID0gbmV3IFJlZVZhbGlkYXRlKHtcbiAgICAgIG5hbWU6ICdyZXF1aXJlZHxtaW46NicsXG4gICAgICBlbWFpbDogJ3JlcXVpcmVkfGVtYWlsJyxcbiAgICAgIHBhc3N3b3JkOiAncmVxdWlyZWR8bWluOjYnLFxuICAgICAgcGFzc3dvcmRDb25maXJtYXRpb246ICdyZXF1aXJlZHxtaW46NidcbiAgICB9KVxuICAgIFxuICAgIHRoaXMuc3RhdGUgPSB7XG4gICAgICBjcmVkZW50aWFsczoge1xuICAgICAgICBuYW1lOiAnJyxcbiAgICAgICAgZW1haWw6ICcnLFxuICAgICAgICBwYXNzd29yZDogJycsXG4gICAgICAgIHBhc3N3b3JkQ29uZmlybWF0aW9uOiAnJyxcbiAgICAgIH0sXG4gICAgICBlcnJvcnM6IHRoaXMudmFsaWRhdG9yLmVycm9ycyxcbiAgICAgIGZpZWxkczogdGhpcy52YWxpZGF0b3IuZmllbGRzXG4gICAgfVxuICAgIFxuICAgIHRoaXMuaGFuZGxlQ2hhbmdlID0gdGhpcy5oYW5kbGVDaGFuZ2UuYmluZCh0aGlzKVxuICAgIHRoaXMuaGFuZGxlU3VibWl0ID0gdGhpcy5oYW5kbGVTdWJtaXQuYmluZCh0aGlzKVxuICB9XG4gIFxuICBjb21wb25lbnREaWRNb3VudCgpIHtcbiAgICAkKCdib2R5JykuYXR0cignc3R5bGUnLCAnYmFja2dyb3VuZC1jb2xvcjogI2VlZScpXG4gIH1cbiAgXG4gIGNvbXBvbmVudFdpbGxVbm1vdW50KCkge1xuICAgICQoJ2JvZHknKS5yZW1vdmVBdHRyKCdzdHlsZScpXG4gIH1cbiAgXG4gIC8vIGV2ZW50IHRvIGhhbmRsZSBpbnB1dCBjaGFuZ2VcbiAgaGFuZGxlQ2hhbmdlKG5hbWUsIHZhbHVlKSB7XG4gICAgY29uc3QgeyBlcnJvcnMgfSA9IHRoaXMudmFsaWRhdG9yXG5cbiAgICB0aGlzLnNldFN0YXRlKHtjcmVkZW50aWFsczogeyAuLi50aGlzLnN0YXRlLmNyZWRlbnRpYWxzLCBbbmFtZV06IHZhbHVlIH19KVxuICAgIGVycm9ycy5yZW1vdmUobmFtZSlcbiAgICBcbiAgICB0aGlzLnZhbGlkYXRvci52YWxpZGF0ZShuYW1lLCB2YWx1ZSlcbiAgICAgIC50aGVuKCgpID0+IHtcbiAgICAgICAgdGhpcy5zZXRTdGF0ZSh7IGVycm9ycyB9KVxuICAgICAgfSlcbiAgfVxuICBcbiAgaGFuZGxlU3VibWl0KGUpIHtcbiAgICBlLnByZXZlbnREZWZhdWx0KClcbiAgICBjb25zdCB7IGNyZWRlbnRpYWxzIH0gPSB0aGlzLnN0YXRlXG4gICAgY29uc3QgeyBlcnJvcnMgfSA9IHRoaXMudmFsaWRhdG9yXG4gIFxuICAgIHRoaXMudmFsaWRhdG9yLnZhbGlkYXRlQWxsKGNyZWRlbnRpYWxzKVxuICAgICAgLnRoZW4oKHN1Y2Nlc3MpID0+IHtcbiAgICAgICAgaWYgKHN1Y2Nlc3MpIHtcbiAgICAgICAgICB0aGlzLnN1Ym1pdChjcmVkZW50aWFscylcbiAgICAgICAgfSBlbHNlIHtcbiAgICAgICAgICB0aGlzLnNldFN0YXRlKHsgZXJyb3JzIH0pXG4gICAgICAgIH1cbiAgICAgIH0pXG4gIH1cbiAgXG4gIHN1Ym1pdChjcmVkZW50aWFscykge1xuICAgIHRoaXMucHJvcHMuZGlzcGF0Y2gocmVnaXN0ZXIoY3JlZGVudGlhbHMpKVxuICAgICAgLmNhdGNoKCh7IGVycm9yLCBzdGF0dXNDb2RlIH0pID0+IHtcbiAgICAgICAgY29uc3QgeyBlcnJvcnMgfSA9IHRoaXMudmFsaWRhdG9yXG4gICAgICAgIFxuICAgICAgICBpZiAoc3RhdHVzQ29kZSA9PT0gNDIyKSB7XG4gICAgICAgICAgXy5mb3JPd24oZXJyb3IsIChtZXNzYWdlLCBmaWVsZCkgPT4ge1xuICAgICAgICAgICAgZXJyb3JzLmFkZChmaWVsZCwgbWVzc2FnZSk7XG4gICAgICAgICAgfSk7XG4gICAgICAgIH0gZWxzZSBpZiAoc3RhdHVzQ29kZSA9PT0gNDAxKSB7XG4gICAgICAgICAgZXJyb3JzLmFkZCgncGFzc3dvcmQnLCBlcnJvcik7XG4gICAgICAgIH1cbiAgICAgICAgXG4gICAgICAgIHRoaXMuc2V0U3RhdGUoeyBlcnJvcnMgfSlcbiAgICAgIH0pXG4gIH1cbiAgXG4gIHJlbmRlcigpIHtcbiAgICAvLyBjaGVjayBpZiB1c2VyIGlzIGF1dGhlbnRpY2F0ZWQgdGhlbiByZWRpcmVjdCBoaW0gdG8gaG9tZSBwYWdlXG4gICAgaWYgKHRoaXMucHJvcHMuaXNBdXRoZW50aWNhdGVkKSB7XG4gICAgICByZXR1cm4gPFJlZGlyZWN0IHRvPVwiL1wiIC8+XG4gICAgfVxuICAgIFxuICAgIGNvbnN0IHsgbmFtZSwgZW1haWwsIHBhc3N3b3JkLCBwYXNzd29yZENvbmZpcm1hdGlvbiB9ID0gdGhpcy5zdGF0ZS5jcmVkZW50aWFsc1xuICAgIGNvbnN0IHByb3BzID0ge1xuICAgICAgbmFtZSxcbiAgICAgIGVtYWlsLFxuICAgICAgcGFzc3dvcmQsXG4gICAgICBwYXNzd29yZENvbmZpcm1hdGlvbixcbiAgICAgIGVycm9yczogdGhpcy5zdGF0ZS5lcnJvcnMsXG4gICAgICBoYW5kbGVDaGFuZ2U6IHRoaXMuaGFuZGxlQ2hhbmdlLFxuICAgICAgaGFuZGxlU3VibWl0OiB0aGlzLmhhbmRsZVN1Ym1pdCxcbiAgICB9XG4gICAgXG4gICAgcmV0dXJuICg8ZGl2IGNsYXNzTmFtZT1cImNvbnRhaW5lciBweS01XCI+XG4gICAgICA8ZGl2IGNsYXNzTmFtZT1cInJvd1wiPlxuICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cImNvbC1tZC0xMlwiPlxuICAgICAgICAgIDxkaXYgY2xhc3NOYW1lPVwicm93XCI+XG4gICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cIm14LWF1dG9cIj5cbiAgICAgICAgICAgICAgPHNwYW4gY2xhc3NOYW1lPVwiYW5jaG9yXCIvPlxuICAgICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cImNhcmQgaGFzLXNoYWRvd1wiPlxuICAgICAgICAgICAgICAgIDxkaXYgY2xhc3NOYW1lPVwiY2FyZC1ib2R5XCI+XG4gICAgICAgICAgICAgICAgICA8Rm9ybSB7Li4ucHJvcHN9ICAvPlxuICAgICAgICAgICAgICAgIDwvZGl2PlxuICAgICAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgICAgIDwvZGl2PlxuICAgICAgICAgIDwvZGl2PlxuICAgICAgICA8L2Rpdj5cbiAgICAgIDwvZGl2PlxuICAgIDwvZGl2PilcbiAgfVxufVxuXG5leHBvcnQgZGVmYXVsdCBQYWdlXG4iXSwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./resources/js/modules/auth/pages/register/Page.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\");\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var jquery__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! jquery */ \"./node_modules/jquery/dist/jquery.js\");\n/* harmony import */ var jquery__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(jquery__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var lodash__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! lodash */ \"./node_modules/lodash/lodash.js\");\n/* harmony import */ var lodash__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(lodash__WEBPACK_IMPORTED_MODULE_3__);\n/* harmony import */ var react_router_dom__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! react-router-dom */ \"./node_modules/react-router-dom/esm/react-router-dom.js\");\n/* harmony import */ var _service__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ../../service */ \"./resources/js/modules/auth/service.js\");\n/* harmony import */ var ree_validate__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ree-validate */ \"./node_modules/ree-validate/dist/ree-validate.esm.js\");\n/* harmony import */ var _components_Form__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./components/Form */ \"./resources/js/modules/auth/pages/register/components/Form.js\");\nfunction ownKeys(object, enumerableOnly) { var keys = Object.keys(object); if (Object.getOwnPropertySymbols) { var symbols = Object.getOwnPropertySymbols(object); if (enumerableOnly) symbols = symbols.filter(function (sym) { return Object.getOwnPropertyDescriptor(object, sym).enumerable; }); keys.push.apply(keys, symbols); } return keys; }\n\nfunction _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; if (i % 2) { ownKeys(Object(source), true).forEach(function (key) { _defineProperty(target, key, source[key]); }); } else if (Object.getOwnPropertyDescriptors) { Object.defineProperties(target, Object.getOwnPropertyDescriptors(source)); } else { ownKeys(Object(source)).forEach(function (key) { Object.defineProperty(target, key, Object.getOwnPropertyDescriptor(source, key)); }); } } return target; }\n\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\n//import libs\n\n\n\n\n\n\n // import components\n\n // initialize component\n\nclass Page extends react__WEBPACK_IMPORTED_MODULE_0__[\"Component\"] {\n  constructor(props) {\n    super(props);\n    this.validator = new ree_validate__WEBPACK_IMPORTED_MODULE_6__[\"default\"]({\n      name: 'required|min:6',\n      email: 'required|email',\n      password: 'required|min:6',\n      passwordConfirmation: 'required|min:6'\n    });\n    this.state = {\n      credentials: {\n        name: '',\n        email: '',\n        password: '',\n        passwordConfirmation: ''\n      },\n      errors: this.validator.errors,\n      fields: this.validator.fields\n    };\n    this.handleChange = this.handleChange.bind(this);\n    this.handleSubmit = this.handleSubmit.bind(this);\n  }\n\n  componentDidMount() {\n    jquery__WEBPACK_IMPORTED_MODULE_2___default()('body').attr('style', 'background-color: #eee');\n  }\n\n  componentWillUnmount() {\n    jquery__WEBPACK_IMPORTED_MODULE_2___default()('body').removeAttr('style');\n  } // event to handle input change\n\n\n  handleChange(name, value) {\n    const errors = this.validator.errors;\n    this.setState({\n      credentials: _objectSpread({}, this.state.credentials, {\n        [name]: value\n      })\n    });\n    errors.remove(name);\n    this.validator.validate(name, value).then(() => {\n      this.setState({\n        errors\n      });\n    });\n  }\n\n  handleSubmit(e) {\n    e.preventDefault();\n    const credentials = this.state.credentials;\n    const errors = this.validator.errors;\n    this.validator.validateAll(credentials).then(success => {\n      if (success) {\n        this.submit(credentials);\n      } else {\n        this.setState({\n          errors\n        });\n      }\n    });\n  }\n\n  submit(credentials) {\n    this.props.dispatch(Object(_service__WEBPACK_IMPORTED_MODULE_5__[\"register\"])(credentials)).catch(({\n      error,\n      statusCode\n    }) => {\n      const errors = this.validator.errors;\n\n      if (statusCode === 422) {\n        lodash__WEBPACK_IMPORTED_MODULE_3___default.a.forOwn(error, (message, field) => {\n          errors.add(field, message);\n        });\n      } else if (statusCode === 401) {\n        errors.add('password', error);\n      }\n\n      this.setState({\n        errors\n      });\n    });\n  }\n\n  render() {\n    // check if user is authenticated then redirect him to home page\n    if (this.props.isAuthenticated) {\n      return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(react_router_dom__WEBPACK_IMPORTED_MODULE_4__[\"Redirect\"], {\n        to: \"/\"\n      });\n    }\n\n    const _this$state$credentia = this.state.credentials,\n          name = _this$state$credentia.name,\n          email = _this$state$credentia.email,\n          password = _this$state$credentia.password,\n          passwordConfirmation = _this$state$credentia.passwordConfirmation;\n    const props = {\n      name,\n      email,\n      password,\n      passwordConfirmation,\n      errors: this.state.errors,\n      handleChange: this.handleChange,\n      handleSubmit: this.handleSubmit\n    };\n    return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"container py-5\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"row\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"col-md-12\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"row\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"mx-auto\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"span\", {\n      className: \"anchor\"\n    }), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"card has-shadow\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"card-body\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_components_Form__WEBPACK_IMPORTED_MODULE_7__[\"default\"], props))))))));\n  }\n\n}\n\n_defineProperty(Page, \"displayName\", 'RegisterPage');\n\n_defineProperty(Page, \"propTypes\", {\n  isAuthenticated: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.bool.isRequired,\n  dispatch: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.func.isRequired\n});\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Page);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hdXRoL3BhZ2VzL3JlZ2lzdGVyL1BhZ2UuanM/NmRkNSJdLCJuYW1lcyI6WyJQYWdlIiwiQ29tcG9uZW50IiwiY29uc3RydWN0b3IiLCJwcm9wcyIsInZhbGlkYXRvciIsIlJlZVZhbGlkYXRlIiwibmFtZSIsImVtYWlsIiwicGFzc3dvcmQiLCJwYXNzd29yZENvbmZpcm1hdGlvbiIsInN0YXRlIiwiY3JlZGVudGlhbHMiLCJlcnJvcnMiLCJmaWVsZHMiLCJoYW5kbGVDaGFuZ2UiLCJiaW5kIiwiaGFuZGxlU3VibWl0IiwiY29tcG9uZW50RGlkTW91bnQiLCIkIiwiYXR0ciIsImNvbXBvbmVudFdpbGxVbm1vdW50IiwicmVtb3ZlQXR0ciIsInZhbHVlIiwic2V0U3RhdGUiLCJyZW1vdmUiLCJ2YWxpZGF0ZSIsInRoZW4iLCJlIiwicHJldmVudERlZmF1bHQiLCJ2YWxpZGF0ZUFsbCIsInN1Y2Nlc3MiLCJzdWJtaXQiLCJkaXNwYXRjaCIsInJlZ2lzdGVyIiwiY2F0Y2giLCJlcnJvciIsInN0YXR1c0NvZGUiLCJfIiwiZm9yT3duIiwibWVzc2FnZSIsImZpZWxkIiwiYWRkIiwicmVuZGVyIiwiaXNBdXRoZW50aWNhdGVkIiwiUHJvcFR5cGVzIiwiYm9vbCIsImlzUmVxdWlyZWQiLCJmdW5jIl0sIm1hcHBpbmdzIjoiOzs7Ozs7Ozs7Ozs7Ozs7Ozs7O0FBQUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7Q0FHQTs7Q0FHQTs7QUFDQSxNQUFNQSxJQUFOLFNBQW1CQywrQ0FBbkIsQ0FBNkI7QUFPM0JDLGFBQVcsQ0FBQ0MsS0FBRCxFQUFRO0FBQ2pCLFVBQU1BLEtBQU47QUFFQSxTQUFLQyxTQUFMLEdBQWlCLElBQUlDLG9EQUFKLENBQWdCO0FBQy9CQyxVQUFJLEVBQUUsZ0JBRHlCO0FBRS9CQyxXQUFLLEVBQUUsZ0JBRndCO0FBRy9CQyxjQUFRLEVBQUUsZ0JBSHFCO0FBSS9CQywwQkFBb0IsRUFBRTtBQUpTLEtBQWhCLENBQWpCO0FBT0EsU0FBS0MsS0FBTCxHQUFhO0FBQ1hDLGlCQUFXLEVBQUU7QUFDWEwsWUFBSSxFQUFFLEVBREs7QUFFWEMsYUFBSyxFQUFFLEVBRkk7QUFHWEMsZ0JBQVEsRUFBRSxFQUhDO0FBSVhDLDRCQUFvQixFQUFFO0FBSlgsT0FERjtBQU9YRyxZQUFNLEVBQUUsS0FBS1IsU0FBTCxDQUFlUSxNQVBaO0FBUVhDLFlBQU0sRUFBRSxLQUFLVCxTQUFMLENBQWVTO0FBUlosS0FBYjtBQVdBLFNBQUtDLFlBQUwsR0FBb0IsS0FBS0EsWUFBTCxDQUFrQkMsSUFBbEIsQ0FBdUIsSUFBdkIsQ0FBcEI7QUFDQSxTQUFLQyxZQUFMLEdBQW9CLEtBQUtBLFlBQUwsQ0FBa0JELElBQWxCLENBQXVCLElBQXZCLENBQXBCO0FBQ0Q7O0FBRURFLG1CQUFpQixHQUFHO0FBQ2xCQyxpREFBQyxDQUFDLE1BQUQsQ0FBRCxDQUFVQyxJQUFWLENBQWUsT0FBZixFQUF3Qix3QkFBeEI7QUFDRDs7QUFFREMsc0JBQW9CLEdBQUc7QUFDckJGLGlEQUFDLENBQUMsTUFBRCxDQUFELENBQVVHLFVBQVYsQ0FBcUIsT0FBckI7QUFDRCxHQXRDMEIsQ0F3QzNCOzs7QUFDQVAsY0FBWSxDQUFDUixJQUFELEVBQU9nQixLQUFQLEVBQWM7QUFBQSxVQUNoQlYsTUFEZ0IsR0FDTCxLQUFLUixTQURBLENBQ2hCUSxNQURnQjtBQUd4QixTQUFLVyxRQUFMLENBQWM7QUFBQ1osaUJBQVcsb0JBQU8sS0FBS0QsS0FBTCxDQUFXQyxXQUFsQjtBQUErQixTQUFDTCxJQUFELEdBQVFnQjtBQUF2QztBQUFaLEtBQWQ7QUFDQVYsVUFBTSxDQUFDWSxNQUFQLENBQWNsQixJQUFkO0FBRUEsU0FBS0YsU0FBTCxDQUFlcUIsUUFBZixDQUF3Qm5CLElBQXhCLEVBQThCZ0IsS0FBOUIsRUFDR0ksSUFESCxDQUNRLE1BQU07QUFDVixXQUFLSCxRQUFMLENBQWM7QUFBRVg7QUFBRixPQUFkO0FBQ0QsS0FISDtBQUlEOztBQUVESSxjQUFZLENBQUNXLENBQUQsRUFBSTtBQUNkQSxLQUFDLENBQUNDLGNBQUY7QUFEYyxVQUVOakIsV0FGTSxHQUVVLEtBQUtELEtBRmYsQ0FFTkMsV0FGTTtBQUFBLFVBR05DLE1BSE0sR0FHSyxLQUFLUixTQUhWLENBR05RLE1BSE07QUFLZCxTQUFLUixTQUFMLENBQWV5QixXQUFmLENBQTJCbEIsV0FBM0IsRUFDR2UsSUFESCxDQUNTSSxPQUFELElBQWE7QUFDakIsVUFBSUEsT0FBSixFQUFhO0FBQ1gsYUFBS0MsTUFBTCxDQUFZcEIsV0FBWjtBQUNELE9BRkQsTUFFTztBQUNMLGFBQUtZLFFBQUwsQ0FBYztBQUFFWDtBQUFGLFNBQWQ7QUFDRDtBQUNGLEtBUEg7QUFRRDs7QUFFRG1CLFFBQU0sQ0FBQ3BCLFdBQUQsRUFBYztBQUNsQixTQUFLUixLQUFMLENBQVc2QixRQUFYLENBQW9CQyx5REFBUSxDQUFDdEIsV0FBRCxDQUE1QixFQUNHdUIsS0FESCxDQUNTLENBQUM7QUFBRUMsV0FBRjtBQUFTQztBQUFULEtBQUQsS0FBMkI7QUFBQSxZQUN4QnhCLE1BRHdCLEdBQ2IsS0FBS1IsU0FEUSxDQUN4QlEsTUFEd0I7O0FBR2hDLFVBQUl3QixVQUFVLEtBQUssR0FBbkIsRUFBd0I7QUFDdEJDLHFEQUFDLENBQUNDLE1BQUYsQ0FBU0gsS0FBVCxFQUFnQixDQUFDSSxPQUFELEVBQVVDLEtBQVYsS0FBb0I7QUFDbEM1QixnQkFBTSxDQUFDNkIsR0FBUCxDQUFXRCxLQUFYLEVBQWtCRCxPQUFsQjtBQUNELFNBRkQ7QUFHRCxPQUpELE1BSU8sSUFBSUgsVUFBVSxLQUFLLEdBQW5CLEVBQXdCO0FBQzdCeEIsY0FBTSxDQUFDNkIsR0FBUCxDQUFXLFVBQVgsRUFBdUJOLEtBQXZCO0FBQ0Q7O0FBRUQsV0FBS1osUUFBTCxDQUFjO0FBQUVYO0FBQUYsT0FBZDtBQUNELEtBYkg7QUFjRDs7QUFFRDhCLFFBQU0sR0FBRztBQUNQO0FBQ0EsUUFBSSxLQUFLdkMsS0FBTCxDQUFXd0MsZUFBZixFQUFnQztBQUM5QixhQUFPLDJEQUFDLHlEQUFEO0FBQVUsVUFBRSxFQUFDO0FBQWIsUUFBUDtBQUNEOztBQUpNLGtDQU1pRCxLQUFLakMsS0FBTCxDQUFXQyxXQU41RDtBQUFBLFVBTUNMLElBTkQseUJBTUNBLElBTkQ7QUFBQSxVQU1PQyxLQU5QLHlCQU1PQSxLQU5QO0FBQUEsVUFNY0MsUUFOZCx5QkFNY0EsUUFOZDtBQUFBLFVBTXdCQyxvQkFOeEIseUJBTXdCQSxvQkFOeEI7QUFPUCxVQUFNTixLQUFLLEdBQUc7QUFDWkcsVUFEWTtBQUVaQyxXQUZZO0FBR1pDLGNBSFk7QUFJWkMsMEJBSlk7QUFLWkcsWUFBTSxFQUFFLEtBQUtGLEtBQUwsQ0FBV0UsTUFMUDtBQU1aRSxrQkFBWSxFQUFFLEtBQUtBLFlBTlA7QUFPWkUsa0JBQVksRUFBRSxLQUFLQTtBQVBQLEtBQWQ7QUFVQSxXQUFRO0FBQUssZUFBUyxFQUFDO0FBQWYsT0FDTjtBQUFLLGVBQVMsRUFBQztBQUFmLE9BQ0U7QUFBSyxlQUFTLEVBQUM7QUFBZixPQUNFO0FBQUssZUFBUyxFQUFDO0FBQWYsT0FDRTtBQUFLLGVBQVMsRUFBQztBQUFmLE9BQ0U7QUFBTSxlQUFTLEVBQUM7QUFBaEIsTUFERixFQUVFO0FBQUssZUFBUyxFQUFDO0FBQWYsT0FDRTtBQUFLLGVBQVMsRUFBQztBQUFmLE9BQ0UsMkRBQUMsd0RBQUQsRUFBVWIsS0FBVixDQURGLENBREYsQ0FGRixDQURGLENBREYsQ0FERixDQURNLENBQVI7QUFnQkQ7O0FBdEgwQjs7Z0JBQXZCSCxJLGlCQUNpQixjOztnQkFEakJBLEksZUFFZTtBQUNqQjJDLGlCQUFlLEVBQUVDLGlEQUFTLENBQUNDLElBQVYsQ0FBZUMsVUFEZjtBQUVqQmQsVUFBUSxFQUFFWSxpREFBUyxDQUFDRyxJQUFWLENBQWVEO0FBRlIsQzs7QUF1SE45QyxtRUFBZiIsImZpbGUiOiIuL3Jlc291cmNlcy9qcy9tb2R1bGVzL2F1dGgvcGFnZXMvcmVnaXN0ZXIvUGFnZS5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbIi8vaW1wb3J0IGxpYnNcbmltcG9ydCBSZWFjdCwgeyBDb21wb25lbnQgfSBmcm9tICdyZWFjdCdcbmltcG9ydCBQcm9wVHlwZXMgZnJvbSAncHJvcC10eXBlcydcbmltcG9ydCAkIGZyb20gJ2pxdWVyeSdcbmltcG9ydCBfIGZyb20gJ2xvZGFzaCdcbmltcG9ydCB7IFJlZGlyZWN0IH0gZnJvbSAncmVhY3Qtcm91dGVyLWRvbSdcbmltcG9ydCB7IHJlZ2lzdGVyIH0gZnJvbSAnLi4vLi4vc2VydmljZSdcbmltcG9ydCBSZWVWYWxpZGF0ZSBmcm9tICdyZWUtdmFsaWRhdGUnXG5cbi8vIGltcG9ydCBjb21wb25lbnRzXG5pbXBvcnQgRm9ybSBmcm9tICcuL2NvbXBvbmVudHMvRm9ybSdcblxuLy8gaW5pdGlhbGl6ZSBjb21wb25lbnRcbmNsYXNzIFBhZ2UgZXh0ZW5kcyBDb21wb25lbnQge1xuICBzdGF0aWMgZGlzcGxheU5hbWUgPSAnUmVnaXN0ZXJQYWdlJ1xuICBzdGF0aWMgcHJvcFR5cGVzID0ge1xuICAgIGlzQXV0aGVudGljYXRlZDogUHJvcFR5cGVzLmJvb2wuaXNSZXF1aXJlZCxcbiAgICBkaXNwYXRjaDogUHJvcFR5cGVzLmZ1bmMuaXNSZXF1aXJlZCxcbiAgfVxuICBcbiAgY29uc3RydWN0b3IocHJvcHMpIHtcbiAgICBzdXBlcihwcm9wcylcbiAgICBcbiAgICB0aGlzLnZhbGlkYXRvciA9IG5ldyBSZWVWYWxpZGF0ZSh7XG4gICAgICBuYW1lOiAncmVxdWlyZWR8bWluOjYnLFxuICAgICAgZW1haWw6ICdyZXF1aXJlZHxlbWFpbCcsXG4gICAgICBwYXNzd29yZDogJ3JlcXVpcmVkfG1pbjo2JyxcbiAgICAgIHBhc3N3b3JkQ29uZmlybWF0aW9uOiAncmVxdWlyZWR8bWluOjYnXG4gICAgfSlcbiAgICBcbiAgICB0aGlzLnN0YXRlID0ge1xuICAgICAgY3JlZGVudGlhbHM6IHtcbiAgICAgICAgbmFtZTogJycsXG4gICAgICAgIGVtYWlsOiAnJyxcbiAgICAgICAgcGFzc3dvcmQ6ICcnLFxuICAgICAgICBwYXNzd29yZENvbmZpcm1hdGlvbjogJycsXG4gICAgICB9LFxuICAgICAgZXJyb3JzOiB0aGlzLnZhbGlkYXRvci5lcnJvcnMsXG4gICAgICBmaWVsZHM6IHRoaXMudmFsaWRhdG9yLmZpZWxkc1xuICAgIH1cbiAgICBcbiAgICB0aGlzLmhhbmRsZUNoYW5nZSA9IHRoaXMuaGFuZGxlQ2hhbmdlLmJpbmQodGhpcylcbiAgICB0aGlzLmhhbmRsZVN1Ym1pdCA9IHRoaXMuaGFuZGxlU3VibWl0LmJpbmQodGhpcylcbiAgfVxuICBcbiAgY29tcG9uZW50RGlkTW91bnQoKSB7XG4gICAgJCgnYm9keScpLmF0dHIoJ3N0eWxlJywgJ2JhY2tncm91bmQtY29sb3I6ICNlZWUnKVxuICB9XG4gIFxuICBjb21wb25lbnRXaWxsVW5tb3VudCgpIHtcbiAgICAkKCdib2R5JykucmVtb3ZlQXR0cignc3R5bGUnKVxuICB9XG4gIFxuICAvLyBldmVudCB0byBoYW5kbGUgaW5wdXQgY2hhbmdlXG4gIGhhbmRsZUNoYW5nZShuYW1lLCB2YWx1ZSkge1xuICAgIGNvbnN0IHsgZXJyb3JzIH0gPSB0aGlzLnZhbGlkYXRvclxuXG4gICAgdGhpcy5zZXRTdGF0ZSh7Y3JlZGVudGlhbHM6IHsgLi4udGhpcy5zdGF0ZS5jcmVkZW50aWFscywgW25hbWVdOiB2YWx1ZSB9fSlcbiAgICBlcnJvcnMucmVtb3ZlKG5hbWUpXG4gICAgXG4gICAgdGhpcy52YWxpZGF0b3IudmFsaWRhdGUobmFtZSwgdmFsdWUpXG4gICAgICAudGhlbigoKSA9PiB7XG4gICAgICAgIHRoaXMuc2V0U3RhdGUoeyBlcnJvcnMgfSlcbiAgICAgIH0pXG4gIH1cbiAgXG4gIGhhbmRsZVN1Ym1pdChlKSB7XG4gICAgZS5wcmV2ZW50RGVmYXVsdCgpXG4gICAgY29uc3QgeyBjcmVkZW50aWFscyB9ID0gdGhpcy5zdGF0ZVxuICAgIGNvbnN0IHsgZXJyb3JzIH0gPSB0aGlzLnZhbGlkYXRvclxuICBcbiAgICB0aGlzLnZhbGlkYXRvci52YWxpZGF0ZUFsbChjcmVkZW50aWFscylcbiAgICAgIC50aGVuKChzdWNjZXNzKSA9PiB7XG4gICAgICAgIGlmIChzdWNjZXNzKSB7XG4gICAgICAgICAgdGhpcy5zdWJtaXQoY3JlZGVudGlhbHMpXG4gICAgICAgIH0gZWxzZSB7XG4gICAgICAgICAgdGhpcy5zZXRTdGF0ZSh7IGVycm9ycyB9KVxuICAgICAgICB9XG4gICAgICB9KVxuICB9XG4gIFxuICBzdWJtaXQoY3JlZGVudGlhbHMpIHtcbiAgICB0aGlzLnByb3BzLmRpc3BhdGNoKHJlZ2lzdGVyKGNyZWRlbnRpYWxzKSlcbiAgICAgIC5jYXRjaCgoeyBlcnJvciwgc3RhdHVzQ29kZSB9KSA9PiB7XG4gICAgICAgIGNvbnN0IHsgZXJyb3JzIH0gPSB0aGlzLnZhbGlkYXRvclxuICAgICAgICBcbiAgICAgICAgaWYgKHN0YXR1c0NvZGUgPT09IDQyMikge1xuICAgICAgICAgIF8uZm9yT3duKGVycm9yLCAobWVzc2FnZSwgZmllbGQpID0+IHtcbiAgICAgICAgICAgIGVycm9ycy5hZGQoZmllbGQsIG1lc3NhZ2UpO1xuICAgICAgICAgIH0pO1xuICAgICAgICB9IGVsc2UgaWYgKHN0YXR1c0NvZGUgPT09IDQwMSkge1xuICAgICAgICAgIGVycm9ycy5hZGQoJ3Bhc3N3b3JkJywgZXJyb3IpO1xuICAgICAgICB9XG4gICAgICAgIFxuICAgICAgICB0aGlzLnNldFN0YXRlKHsgZXJyb3JzIH0pXG4gICAgICB9KVxuICB9XG4gIFxuICByZW5kZXIoKSB7XG4gICAgLy8gY2hlY2sgaWYgdXNlciBpcyBhdXRoZW50aWNhdGVkIHRoZW4gcmVkaXJlY3QgaGltIHRvIGhvbWUgcGFnZVxuICAgIGlmICh0aGlzLnByb3BzLmlzQXV0aGVudGljYXRlZCkge1xuICAgICAgcmV0dXJuIDxSZWRpcmVjdCB0bz1cIi9cIiAvPlxuICAgIH1cbiAgICBcbiAgICBjb25zdCB7IG5hbWUsIGVtYWlsLCBwYXNzd29yZCwgcGFzc3dvcmRDb25maXJtYXRpb24gfSA9IHRoaXMuc3RhdGUuY3JlZGVudGlhbHNcbiAgICBjb25zdCBwcm9wcyA9IHtcbiAgICAgIG5hbWUsXG4gICAgICBlbWFpbCxcbiAgICAgIHBhc3N3b3JkLFxuICAgICAgcGFzc3dvcmRDb25maXJtYXRpb24sXG4gICAgICBlcnJvcnM6IHRoaXMuc3RhdGUuZXJyb3JzLFxuICAgICAgaGFuZGxlQ2hhbmdlOiB0aGlzLmhhbmRsZUNoYW5nZSxcbiAgICAgIGhhbmRsZVN1Ym1pdDogdGhpcy5oYW5kbGVTdWJtaXQsXG4gICAgfVxuICAgIFxuICAgIHJldHVybiAoPGRpdiBjbGFzc05hbWU9XCJjb250YWluZXIgcHktNVwiPlxuICAgICAgPGRpdiBjbGFzc05hbWU9XCJyb3dcIj5cbiAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJjb2wtbWQtMTJcIj5cbiAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cInJvd1wiPlxuICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJteC1hdXRvXCI+XG4gICAgICAgICAgICAgIDxzcGFuIGNsYXNzTmFtZT1cImFuY2hvclwiLz5cbiAgICAgICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJjYXJkIGhhcy1zaGFkb3dcIj5cbiAgICAgICAgICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cImNhcmQtYm9keVwiPlxuICAgICAgICAgICAgICAgICAgPEZvcm0gey4uLnByb3BzfSAgLz5cbiAgICAgICAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgICAgICAgPC9kaXY+XG4gICAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgICA8L2Rpdj5cbiAgICAgICAgPC9kaXY+XG4gICAgICA8L2Rpdj5cbiAgICA8L2Rpdj4pXG4gIH1cbn1cblxuZXhwb3J0IGRlZmF1bHQgUGFnZVxuIl0sInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./resources/js/modules/auth/pages/register/Page.js\n");
 
 /***/ }),
 
@@ -20,7 +20,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var reac
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\");\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var react_router_dom__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! react-router-dom */ \"./node_modules/react-router-dom/esm/react-router-dom.js\");\n\n\n\nvar displayName = 'RegisterFrom';\nvar propTypes = {\n  name: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.string.isRequired,\n  email: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.string.isRequired,\n  password: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.string.isRequired,\n  passwordConfirmation: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.string.isRequired,\n  errors: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.object.isRequired,\n  handleChange: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.func.isRequired,\n  handleSubmit: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.func.isRequired\n};\n\nvar Form = function Form(_ref) {\n  var name = _ref.name,\n      email = _ref.email,\n      password = _ref.password,\n      passwordConfirmation = _ref.passwordConfirmation,\n      errors = _ref.errors,\n      handleChange = _ref.handleChange,\n      handleSubmit = _ref.handleSubmit;\n  return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"form\", {\n    className: \"form\",\n    role: \"form\",\n    onSubmit: handleSubmit,\n    noValidate: true\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"h2\", {\n    className: \"card-title\"\n  }, \"Sign up\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"form-group\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"label\", {\n    htmlFor: \"name\",\n    className: \"sr-only\"\n  }, \"Email\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"input\", {\n    type: \"text\",\n    className: \"form-control form-control-lg rounded-0 \".concat(errors.has('name') && 'is-invalid'),\n    name: \"name\",\n    id: \"name\",\n    placeholder: \"Full Name\",\n    value: name || '',\n    onChange: function onChange(e) {\n      return handleChange(e.target.name, e.target.value);\n    },\n    required: true,\n    autoFocus: true\n  }), errors.has('name') && react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"invalid-feedback\"\n  }, errors.first('name'))), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"form-group\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"label\", {\n    htmlFor: \"email\",\n    className: \"sr-only\"\n  }, \"Email\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"input\", {\n    type: \"email\",\n    className: \"form-control form-control-lg rounded-0 \".concat(errors.has('email') && 'is-invalid'),\n    name: \"email\",\n    id: \"email\",\n    placeholder: \"Email address\",\n    value: email || '',\n    onChange: function onChange(e) {\n      return handleChange(e.target.name, e.target.value);\n    },\n    required: true,\n    autoFocus: true\n  }), errors.has('email') && react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"invalid-feedback\"\n  }, errors.first('email'))), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"form-group\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"label\", {\n    htmlFor: \"password\",\n    className: \"sr-only\"\n  }, \"Password\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"input\", {\n    type: \"password\",\n    className: \"form-control form-control-lg rounded-0 \".concat(errors.has('password') && 'is-invalid'),\n    id: \"password\",\n    name: \"password\",\n    placeholder: \"Password\",\n    value: password || '',\n    onChange: function onChange(e) {\n      return handleChange(e.target.name, e.target.value);\n    },\n    required: true\n  }), errors.has('password') && react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"invalid-feedback\"\n  }, errors.first('password'))), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"form-group\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"label\", {\n    htmlFor: \"passwordConfirmation\",\n    className: \"sr-only\"\n  }, \"Confirm Password\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"input\", {\n    type: \"password\",\n    className: \"form-control form-control-lg rounded-0 \".concat(errors.has('passwordConfirmation') && 'is-invalid'),\n    id: \"passwordConfirmation\",\n    name: \"passwordConfirmation\",\n    placeholder: \"Confirm Password\",\n    value: passwordConfirmation || '',\n    onChange: function onChange(e) {\n      return handleChange(e.target.name, e.target.value);\n    },\n    required: true\n  }), errors.has('passwordConfirmation') && react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"invalid-feedback\"\n  }, errors.first('passwordConfirmation'))), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"button\", {\n    className: \"btn btn-lg btn-primary btn-block\",\n    type: \"submit\",\n    disabled: errors.any()\n  }, \"Sign Up\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"p\", null, \"Back to\", react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(react_router_dom__WEBPACK_IMPORTED_MODULE_2__[\"Link\"], {\n    to: \"/login\",\n    className: \"btn btn-link\"\n  }, \"Login\")));\n};\n\nForm.displayName = displayName;\nForm.propTypes = propTypes;\n/* harmony default export */ __webpack_exports__[\"default\"] = (Form);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hdXRoL3BhZ2VzL3JlZ2lzdGVyL2NvbXBvbmVudHMvRm9ybS5qcz82ZDY4Il0sIm5hbWVzIjpbImRpc3BsYXlOYW1lIiwicHJvcFR5cGVzIiwibmFtZSIsIlByb3BUeXBlcyIsInN0cmluZyIsImlzUmVxdWlyZWQiLCJlbWFpbCIsInBhc3N3b3JkIiwicGFzc3dvcmRDb25maXJtYXRpb24iLCJlcnJvcnMiLCJvYmplY3QiLCJoYW5kbGVDaGFuZ2UiLCJmdW5jIiwiaGFuZGxlU3VibWl0IiwiRm9ybSIsImhhcyIsImUiLCJ0YXJnZXQiLCJ2YWx1ZSIsImZpcnN0IiwiYW55Il0sIm1hcHBpbmdzIjoiQUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUNBO0FBQ0E7QUFFQSxJQUFNQSxXQUFXLEdBQUcsY0FBcEI7QUFFQSxJQUFNQyxTQUFTLEdBQUc7QUFDaEJDLE1BQUksRUFBRUMsaURBQVMsQ0FBQ0MsTUFBVixDQUFpQkMsVUFEUDtBQUVoQkMsT0FBSyxFQUFFSCxpREFBUyxDQUFDQyxNQUFWLENBQWlCQyxVQUZSO0FBR2hCRSxVQUFRLEVBQUVKLGlEQUFTLENBQUNDLE1BQVYsQ0FBaUJDLFVBSFg7QUFJaEJHLHNCQUFvQixFQUFFTCxpREFBUyxDQUFDQyxNQUFWLENBQWlCQyxVQUp2QjtBQUtoQkksUUFBTSxFQUFFTixpREFBUyxDQUFDTyxNQUFWLENBQWlCTCxVQUxUO0FBTWhCTSxjQUFZLEVBQUVSLGlEQUFTLENBQUNTLElBQVYsQ0FBZVAsVUFOYjtBQU9oQlEsY0FBWSxFQUFFVixpREFBUyxDQUFDUyxJQUFWLENBQWVQO0FBUGIsQ0FBbEI7O0FBVUEsSUFBTVMsSUFBSSxHQUFHLFNBQVBBLElBQU8sT0FBeUY7QUFBQSxNQUF0RlosSUFBc0YsUUFBdEZBLElBQXNGO0FBQUEsTUFBaEZJLEtBQWdGLFFBQWhGQSxLQUFnRjtBQUFBLE1BQXpFQyxRQUF5RSxRQUF6RUEsUUFBeUU7QUFBQSxNQUEvREMsb0JBQStELFFBQS9EQSxvQkFBK0Q7QUFBQSxNQUF6Q0MsTUFBeUMsUUFBekNBLE1BQXlDO0FBQUEsTUFBakNFLFlBQWlDLFFBQWpDQSxZQUFpQztBQUFBLE1BQW5CRSxZQUFtQixRQUFuQkEsWUFBbUI7QUFDcEcsU0FBUTtBQUFNLGFBQVMsRUFBQyxNQUFoQjtBQUF1QixRQUFJLEVBQUMsTUFBNUI7QUFBbUMsWUFBUSxFQUFFQSxZQUE3QztBQUEyRCxjQUFVO0FBQXJFLEtBQ047QUFBSSxhQUFTLEVBQUM7QUFBZCxlQURNLEVBRU47QUFBSyxhQUFTLEVBQUM7QUFBZixLQUNFO0FBQU8sV0FBTyxFQUFDLE1BQWY7QUFBc0IsYUFBUyxFQUFDO0FBQWhDLGFBREYsRUFFRTtBQUFPLFFBQUksRUFBQyxNQUFaO0FBQ08sYUFBUyxtREFBNENKLE1BQU0sQ0FBQ00sR0FBUCxDQUFXLE1BQVgsS0FBc0IsWUFBbEUsQ0FEaEI7QUFFTyxRQUFJLEVBQUMsTUFGWjtBQUdPLE1BQUUsRUFBQyxNQUhWO0FBSU8sZUFBVyxFQUFDLFdBSm5CO0FBS08sU0FBSyxFQUFFYixJQUFJLElBQUksRUFMdEI7QUFNTyxZQUFRLEVBQUUsa0JBQUFjLENBQUM7QUFBQSxhQUFJTCxZQUFZLENBQUNLLENBQUMsQ0FBQ0MsTUFBRixDQUFTZixJQUFWLEVBQWdCYyxDQUFDLENBQUNDLE1BQUYsQ0FBU0MsS0FBekIsQ0FBaEI7QUFBQSxLQU5sQjtBQU9PLFlBQVEsTUFQZjtBQVFPLGFBQVM7QUFSaEIsSUFGRixFQVdHVCxNQUFNLENBQUNNLEdBQVAsQ0FBVyxNQUFYLEtBQXNCO0FBQUssYUFBUyxFQUFDO0FBQWYsS0FBbUNOLE1BQU0sQ0FBQ1UsS0FBUCxDQUFhLE1BQWIsQ0FBbkMsQ0FYekIsQ0FGTSxFQWVOO0FBQUssYUFBUyxFQUFDO0FBQWYsS0FDRTtBQUFPLFdBQU8sRUFBQyxPQUFmO0FBQXVCLGFBQVMsRUFBQztBQUFqQyxhQURGLEVBRUU7QUFBTyxRQUFJLEVBQUMsT0FBWjtBQUNPLGFBQVMsbURBQTRDVixNQUFNLENBQUNNLEdBQVAsQ0FBVyxPQUFYLEtBQXVCLFlBQW5FLENBRGhCO0FBRU8sUUFBSSxFQUFDLE9BRlo7QUFHTyxNQUFFLEVBQUMsT0FIVjtBQUlPLGVBQVcsRUFBQyxlQUpuQjtBQUtPLFNBQUssRUFBRVQsS0FBSyxJQUFJLEVBTHZCO0FBTU8sWUFBUSxFQUFFLGtCQUFBVSxDQUFDO0FBQUEsYUFBSUwsWUFBWSxDQUFDSyxDQUFDLENBQUNDLE1BQUYsQ0FBU2YsSUFBVixFQUFnQmMsQ0FBQyxDQUFDQyxNQUFGLENBQVNDLEtBQXpCLENBQWhCO0FBQUEsS0FObEI7QUFPTyxZQUFRLE1BUGY7QUFRTyxhQUFTO0FBUmhCLElBRkYsRUFXR1QsTUFBTSxDQUFDTSxHQUFQLENBQVcsT0FBWCxLQUF1QjtBQUFLLGFBQVMsRUFBQztBQUFmLEtBQW1DTixNQUFNLENBQUNVLEtBQVAsQ0FBYSxPQUFiLENBQW5DLENBWDFCLENBZk0sRUE0Qk47QUFBSyxhQUFTLEVBQUM7QUFBZixLQUNFO0FBQU8sV0FBTyxFQUFDLFVBQWY7QUFBMEIsYUFBUyxFQUFDO0FBQXBDLGdCQURGLEVBRUU7QUFBTyxRQUFJLEVBQUMsVUFBWjtBQUNPLGFBQVMsbURBQTRDVixNQUFNLENBQUNNLEdBQVAsQ0FBVyxVQUFYLEtBQTBCLFlBQXRFLENBRGhCO0FBRU8sTUFBRSxFQUFDLFVBRlY7QUFHTyxRQUFJLEVBQUMsVUFIWjtBQUlPLGVBQVcsRUFBQyxVQUpuQjtBQUtPLFNBQUssRUFBRVIsUUFBUSxJQUFJLEVBTDFCO0FBTU8sWUFBUSxFQUFFLGtCQUFBUyxDQUFDO0FBQUEsYUFBSUwsWUFBWSxDQUFDSyxDQUFDLENBQUNDLE1BQUYsQ0FBU2YsSUFBVixFQUFnQmMsQ0FBQyxDQUFDQyxNQUFGLENBQVNDLEtBQXpCLENBQWhCO0FBQUEsS0FObEI7QUFPTyxZQUFRO0FBUGYsSUFGRixFQVVHVCxNQUFNLENBQUNNLEdBQVAsQ0FBVyxVQUFYLEtBQTBCO0FBQUssYUFBUyxFQUFDO0FBQWYsS0FBbUNOLE1BQU0sQ0FBQ1UsS0FBUCxDQUFhLFVBQWIsQ0FBbkMsQ0FWN0IsQ0E1Qk0sRUF3Q047QUFBSyxhQUFTLEVBQUM7QUFBZixLQUNFO0FBQU8sV0FBTyxFQUFDLHNCQUFmO0FBQXNDLGFBQVMsRUFBQztBQUFoRCx3QkFERixFQUVFO0FBQU8sUUFBSSxFQUFDLFVBQVo7QUFDTyxhQUFTLG1EQUE0Q1YsTUFBTSxDQUFDTSxHQUFQLENBQVcsc0JBQVgsS0FBc0MsWUFBbEYsQ0FEaEI7QUFFTyxNQUFFLEVBQUMsc0JBRlY7QUFHTyxRQUFJLEVBQUMsc0JBSFo7QUFJTyxlQUFXLEVBQUMsa0JBSm5CO0FBS08sU0FBSyxFQUFFUCxvQkFBb0IsSUFBSSxFQUx0QztBQU1PLFlBQVEsRUFBRSxrQkFBQVEsQ0FBQztBQUFBLGFBQUlMLFlBQVksQ0FBQ0ssQ0FBQyxDQUFDQyxNQUFGLENBQVNmLElBQVYsRUFBZ0JjLENBQUMsQ0FBQ0MsTUFBRixDQUFTQyxLQUF6QixDQUFoQjtBQUFBLEtBTmxCO0FBT08sWUFBUTtBQVBmLElBRkYsRUFVR1QsTUFBTSxDQUFDTSxHQUFQLENBQVcsc0JBQVgsS0FBc0M7QUFBSyxhQUFTLEVBQUM7QUFBZixLQUFtQ04sTUFBTSxDQUFDVSxLQUFQLENBQWEsc0JBQWIsQ0FBbkMsQ0FWekMsQ0F4Q00sRUFvRE47QUFBUSxhQUFTLEVBQUMsa0NBQWxCO0FBQ1EsUUFBSSxFQUFDLFFBRGI7QUFFUSxZQUFRLEVBQUVWLE1BQU0sQ0FBQ1csR0FBUDtBQUZsQixlQXBETSxFQXVETixpRkFBVSwyREFBQyxxREFBRDtBQUFNLE1BQUUsRUFBQyxRQUFUO0FBQWtCLGFBQVMsRUFBQztBQUE1QixhQUFWLENBdkRNLENBQVI7QUF5REQsQ0ExREQ7O0FBNERBTixJQUFJLENBQUNkLFdBQUwsR0FBbUJBLFdBQW5CO0FBQ0FjLElBQUksQ0FBQ2IsU0FBTCxHQUFpQkEsU0FBakI7QUFFZWEsbUVBQWYiLCJmaWxlIjoiLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hdXRoL3BhZ2VzL3JlZ2lzdGVyL2NvbXBvbmVudHMvRm9ybS5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBSZWFjdCBmcm9tICdyZWFjdCdcbmltcG9ydCBQcm9wVHlwZXMgZnJvbSAncHJvcC10eXBlcydcbmltcG9ydCB7IExpbmsgfSBmcm9tICdyZWFjdC1yb3V0ZXItZG9tJ1xuXG5jb25zdCBkaXNwbGF5TmFtZSA9ICdSZWdpc3RlckZyb20nXG5cbmNvbnN0IHByb3BUeXBlcyA9IHtcbiAgbmFtZTogUHJvcFR5cGVzLnN0cmluZy5pc1JlcXVpcmVkLFxuICBlbWFpbDogUHJvcFR5cGVzLnN0cmluZy5pc1JlcXVpcmVkLFxuICBwYXNzd29yZDogUHJvcFR5cGVzLnN0cmluZy5pc1JlcXVpcmVkLFxuICBwYXNzd29yZENvbmZpcm1hdGlvbjogUHJvcFR5cGVzLnN0cmluZy5pc1JlcXVpcmVkLFxuICBlcnJvcnM6IFByb3BUeXBlcy5vYmplY3QuaXNSZXF1aXJlZCxcbiAgaGFuZGxlQ2hhbmdlOiBQcm9wVHlwZXMuZnVuYy5pc1JlcXVpcmVkLFxuICBoYW5kbGVTdWJtaXQ6IFByb3BUeXBlcy5mdW5jLmlzUmVxdWlyZWQsXG59XG5cbmNvbnN0IEZvcm0gPSAoeyBuYW1lLCBlbWFpbCwgcGFzc3dvcmQsIHBhc3N3b3JkQ29uZmlybWF0aW9uLCBlcnJvcnMsIGhhbmRsZUNoYW5nZSwgaGFuZGxlU3VibWl0IH0pID0+IHtcbiAgcmV0dXJuICg8Zm9ybSBjbGFzc05hbWU9XCJmb3JtXCIgcm9sZT1cImZvcm1cIiBvblN1Ym1pdD17aGFuZGxlU3VibWl0fSBub1ZhbGlkYXRlPlxuICAgIDxoMiBjbGFzc05hbWU9XCJjYXJkLXRpdGxlXCI+U2lnbiB1cDwvaDI+XG4gICAgPGRpdiBjbGFzc05hbWU9XCJmb3JtLWdyb3VwXCI+XG4gICAgICA8bGFiZWwgaHRtbEZvcj1cIm5hbWVcIiBjbGFzc05hbWU9XCJzci1vbmx5XCI+RW1haWw8L2xhYmVsPlxuICAgICAgPGlucHV0IHR5cGU9XCJ0ZXh0XCJcbiAgICAgICAgICAgICBjbGFzc05hbWU9e2Bmb3JtLWNvbnRyb2wgZm9ybS1jb250cm9sLWxnIHJvdW5kZWQtMCAke2Vycm9ycy5oYXMoJ25hbWUnKSAmJiAnaXMtaW52YWxpZCd9YH1cbiAgICAgICAgICAgICBuYW1lPVwibmFtZVwiXG4gICAgICAgICAgICAgaWQ9XCJuYW1lXCJcbiAgICAgICAgICAgICBwbGFjZWhvbGRlcj1cIkZ1bGwgTmFtZVwiXG4gICAgICAgICAgICAgdmFsdWU9e25hbWUgfHwgJyd9XG4gICAgICAgICAgICAgb25DaGFuZ2U9e2UgPT4gaGFuZGxlQ2hhbmdlKGUudGFyZ2V0Lm5hbWUsIGUudGFyZ2V0LnZhbHVlKX1cbiAgICAgICAgICAgICByZXF1aXJlZFxuICAgICAgICAgICAgIGF1dG9Gb2N1cy8+XG4gICAgICB7ZXJyb3JzLmhhcygnbmFtZScpICYmIDxkaXYgY2xhc3NOYW1lPVwiaW52YWxpZC1mZWVkYmFja1wiPntlcnJvcnMuZmlyc3QoJ25hbWUnKX08L2Rpdj59XG4gICAgPC9kaXY+XG4gICAgPGRpdiBjbGFzc05hbWU9XCJmb3JtLWdyb3VwXCI+XG4gICAgICA8bGFiZWwgaHRtbEZvcj1cImVtYWlsXCIgY2xhc3NOYW1lPVwic3Itb25seVwiPkVtYWlsPC9sYWJlbD5cbiAgICAgIDxpbnB1dCB0eXBlPVwiZW1haWxcIlxuICAgICAgICAgICAgIGNsYXNzTmFtZT17YGZvcm0tY29udHJvbCBmb3JtLWNvbnRyb2wtbGcgcm91bmRlZC0wICR7ZXJyb3JzLmhhcygnZW1haWwnKSAmJiAnaXMtaW52YWxpZCd9YH1cbiAgICAgICAgICAgICBuYW1lPVwiZW1haWxcIlxuICAgICAgICAgICAgIGlkPVwiZW1haWxcIlxuICAgICAgICAgICAgIHBsYWNlaG9sZGVyPVwiRW1haWwgYWRkcmVzc1wiXG4gICAgICAgICAgICAgdmFsdWU9e2VtYWlsIHx8ICcnfVxuICAgICAgICAgICAgIG9uQ2hhbmdlPXtlID0+IGhhbmRsZUNoYW5nZShlLnRhcmdldC5uYW1lLCBlLnRhcmdldC52YWx1ZSl9XG4gICAgICAgICAgICAgcmVxdWlyZWRcbiAgICAgICAgICAgICBhdXRvRm9jdXMvPlxuICAgICAge2Vycm9ycy5oYXMoJ2VtYWlsJykgJiYgPGRpdiBjbGFzc05hbWU9XCJpbnZhbGlkLWZlZWRiYWNrXCI+e2Vycm9ycy5maXJzdCgnZW1haWwnKX08L2Rpdj59XG4gICAgPC9kaXY+XG4gICAgPGRpdiBjbGFzc05hbWU9XCJmb3JtLWdyb3VwXCI+XG4gICAgICA8bGFiZWwgaHRtbEZvcj1cInBhc3N3b3JkXCIgY2xhc3NOYW1lPVwic3Itb25seVwiPlBhc3N3b3JkPC9sYWJlbD5cbiAgICAgIDxpbnB1dCB0eXBlPVwicGFzc3dvcmRcIlxuICAgICAgICAgICAgIGNsYXNzTmFtZT17YGZvcm0tY29udHJvbCBmb3JtLWNvbnRyb2wtbGcgcm91bmRlZC0wICR7ZXJyb3JzLmhhcygncGFzc3dvcmQnKSAmJiAnaXMtaW52YWxpZCd9YH1cbiAgICAgICAgICAgICBpZD1cInBhc3N3b3JkXCJcbiAgICAgICAgICAgICBuYW1lPVwicGFzc3dvcmRcIlxuICAgICAgICAgICAgIHBsYWNlaG9sZGVyPVwiUGFzc3dvcmRcIlxuICAgICAgICAgICAgIHZhbHVlPXtwYXNzd29yZCB8fCAnJ31cbiAgICAgICAgICAgICBvbkNoYW5nZT17ZSA9PiBoYW5kbGVDaGFuZ2UoZS50YXJnZXQubmFtZSwgZS50YXJnZXQudmFsdWUpfVxuICAgICAgICAgICAgIHJlcXVpcmVkLz5cbiAgICAgIHtlcnJvcnMuaGFzKCdwYXNzd29yZCcpICYmIDxkaXYgY2xhc3NOYW1lPVwiaW52YWxpZC1mZWVkYmFja1wiPntlcnJvcnMuZmlyc3QoJ3Bhc3N3b3JkJyl9PC9kaXY+fVxuICAgIDwvZGl2PlxuICAgIDxkaXYgY2xhc3NOYW1lPVwiZm9ybS1ncm91cFwiPlxuICAgICAgPGxhYmVsIGh0bWxGb3I9XCJwYXNzd29yZENvbmZpcm1hdGlvblwiIGNsYXNzTmFtZT1cInNyLW9ubHlcIj5Db25maXJtIFBhc3N3b3JkPC9sYWJlbD5cbiAgICAgIDxpbnB1dCB0eXBlPVwicGFzc3dvcmRcIlxuICAgICAgICAgICAgIGNsYXNzTmFtZT17YGZvcm0tY29udHJvbCBmb3JtLWNvbnRyb2wtbGcgcm91bmRlZC0wICR7ZXJyb3JzLmhhcygncGFzc3dvcmRDb25maXJtYXRpb24nKSAmJiAnaXMtaW52YWxpZCd9YH1cbiAgICAgICAgICAgICBpZD1cInBhc3N3b3JkQ29uZmlybWF0aW9uXCJcbiAgICAgICAgICAgICBuYW1lPVwicGFzc3dvcmRDb25maXJtYXRpb25cIlxuICAgICAgICAgICAgIHBsYWNlaG9sZGVyPVwiQ29uZmlybSBQYXNzd29yZFwiXG4gICAgICAgICAgICAgdmFsdWU9e3Bhc3N3b3JkQ29uZmlybWF0aW9uIHx8ICcnfVxuICAgICAgICAgICAgIG9uQ2hhbmdlPXtlID0+IGhhbmRsZUNoYW5nZShlLnRhcmdldC5uYW1lLCBlLnRhcmdldC52YWx1ZSl9XG4gICAgICAgICAgICAgcmVxdWlyZWQvPlxuICAgICAge2Vycm9ycy5oYXMoJ3Bhc3N3b3JkQ29uZmlybWF0aW9uJykgJiYgPGRpdiBjbGFzc05hbWU9XCJpbnZhbGlkLWZlZWRiYWNrXCI+e2Vycm9ycy5maXJzdCgncGFzc3dvcmRDb25maXJtYXRpb24nKX08L2Rpdj59XG4gICAgPC9kaXY+XG4gICAgPGJ1dHRvbiBjbGFzc05hbWU9XCJidG4gYnRuLWxnIGJ0bi1wcmltYXJ5IGJ0bi1ibG9ja1wiXG4gICAgICAgICAgICB0eXBlPVwic3VibWl0XCJcbiAgICAgICAgICAgIGRpc2FibGVkPXtlcnJvcnMuYW55KCl9PlNpZ24gVXA8L2J1dHRvbj5cbiAgICA8cD5CYWNrIHRvPExpbmsgdG89XCIvbG9naW5cIiBjbGFzc05hbWU9XCJidG4gYnRuLWxpbmtcIj5Mb2dpbjwvTGluaz48L3A+XG4gIDwvZm9ybT4pXG59XG5cbkZvcm0uZGlzcGxheU5hbWUgPSBkaXNwbGF5TmFtZVxuRm9ybS5wcm9wVHlwZXMgPSBwcm9wVHlwZXNcblxuZXhwb3J0IGRlZmF1bHQgRm9ybVxuIl0sInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./resources/js/modules/auth/pages/register/components/Form.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\");\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var react_router_dom__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! react-router-dom */ \"./node_modules/react-router-dom/esm/react-router-dom.js\");\n\n\n\nconst displayName = 'RegisterFrom';\nconst propTypes = {\n  name: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.string.isRequired,\n  email: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.string.isRequired,\n  password: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.string.isRequired,\n  passwordConfirmation: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.string.isRequired,\n  errors: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.object.isRequired,\n  handleChange: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.func.isRequired,\n  handleSubmit: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.func.isRequired\n};\n\nconst Form = ({\n  name,\n  email,\n  password,\n  passwordConfirmation,\n  errors,\n  handleChange,\n  handleSubmit\n}) => {\n  return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"form\", {\n    className: \"form\",\n    role: \"form\",\n    onSubmit: handleSubmit,\n    noValidate: true\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"h2\", {\n    className: \"card-title\"\n  }, \"Sign up\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"form-group\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"label\", {\n    htmlFor: \"name\",\n    className: \"sr-only\"\n  }, \"Email\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"input\", {\n    type: \"text\",\n    className: \"form-control form-control-lg rounded-0 \".concat(errors.has('name') && 'is-invalid'),\n    name: \"name\",\n    id: \"name\",\n    placeholder: \"Full Name\",\n    value: name || '',\n    onChange: e => handleChange(e.target.name, e.target.value),\n    required: true,\n    autoFocus: true\n  }), errors.has('name') && react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"invalid-feedback\"\n  }, errors.first('name'))), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"form-group\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"label\", {\n    htmlFor: \"email\",\n    className: \"sr-only\"\n  }, \"Email\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"input\", {\n    type: \"email\",\n    className: \"form-control form-control-lg rounded-0 \".concat(errors.has('email') && 'is-invalid'),\n    name: \"email\",\n    id: \"email\",\n    placeholder: \"Email address\",\n    value: email || '',\n    onChange: e => handleChange(e.target.name, e.target.value),\n    required: true,\n    autoFocus: true\n  }), errors.has('email') && react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"invalid-feedback\"\n  }, errors.first('email'))), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"form-group\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"label\", {\n    htmlFor: \"password\",\n    className: \"sr-only\"\n  }, \"Password\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"input\", {\n    type: \"password\",\n    className: \"form-control form-control-lg rounded-0 \".concat(errors.has('password') && 'is-invalid'),\n    id: \"password\",\n    name: \"password\",\n    placeholder: \"Password\",\n    value: password || '',\n    onChange: e => handleChange(e.target.name, e.target.value),\n    required: true\n  }), errors.has('password') && react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"invalid-feedback\"\n  }, errors.first('password'))), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"form-group\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"label\", {\n    htmlFor: \"passwordConfirmation\",\n    className: \"sr-only\"\n  }, \"Confirm Password\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"input\", {\n    type: \"password\",\n    className: \"form-control form-control-lg rounded-0 \".concat(errors.has('passwordConfirmation') && 'is-invalid'),\n    id: \"passwordConfirmation\",\n    name: \"passwordConfirmation\",\n    placeholder: \"Confirm Password\",\n    value: passwordConfirmation || '',\n    onChange: e => handleChange(e.target.name, e.target.value),\n    required: true\n  }), errors.has('passwordConfirmation') && react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"invalid-feedback\"\n  }, errors.first('passwordConfirmation'))), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"button\", {\n    className: \"btn btn-lg btn-primary btn-block\",\n    type: \"submit\",\n    disabled: errors.any()\n  }, \"Sign Up\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"p\", null, \"Back to\", react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(react_router_dom__WEBPACK_IMPORTED_MODULE_2__[\"Link\"], {\n    to: \"/login\",\n    className: \"btn btn-link\"\n  }, \"Login\")));\n};\n\nForm.displayName = displayName;\nForm.propTypes = propTypes;\n/* harmony default export */ __webpack_exports__[\"default\"] = (Form);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hdXRoL3BhZ2VzL3JlZ2lzdGVyL2NvbXBvbmVudHMvRm9ybS5qcz82ZDY4Il0sIm5hbWVzIjpbImRpc3BsYXlOYW1lIiwicHJvcFR5cGVzIiwibmFtZSIsIlByb3BUeXBlcyIsInN0cmluZyIsImlzUmVxdWlyZWQiLCJlbWFpbCIsInBhc3N3b3JkIiwicGFzc3dvcmRDb25maXJtYXRpb24iLCJlcnJvcnMiLCJvYmplY3QiLCJoYW5kbGVDaGFuZ2UiLCJmdW5jIiwiaGFuZGxlU3VibWl0IiwiRm9ybSIsImhhcyIsImUiLCJ0YXJnZXQiLCJ2YWx1ZSIsImZpcnN0IiwiYW55Il0sIm1hcHBpbmdzIjoiQUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUNBO0FBQ0E7QUFFQSxNQUFNQSxXQUFXLEdBQUcsY0FBcEI7QUFFQSxNQUFNQyxTQUFTLEdBQUc7QUFDaEJDLE1BQUksRUFBRUMsaURBQVMsQ0FBQ0MsTUFBVixDQUFpQkMsVUFEUDtBQUVoQkMsT0FBSyxFQUFFSCxpREFBUyxDQUFDQyxNQUFWLENBQWlCQyxVQUZSO0FBR2hCRSxVQUFRLEVBQUVKLGlEQUFTLENBQUNDLE1BQVYsQ0FBaUJDLFVBSFg7QUFJaEJHLHNCQUFvQixFQUFFTCxpREFBUyxDQUFDQyxNQUFWLENBQWlCQyxVQUp2QjtBQUtoQkksUUFBTSxFQUFFTixpREFBUyxDQUFDTyxNQUFWLENBQWlCTCxVQUxUO0FBTWhCTSxjQUFZLEVBQUVSLGlEQUFTLENBQUNTLElBQVYsQ0FBZVAsVUFOYjtBQU9oQlEsY0FBWSxFQUFFVixpREFBUyxDQUFDUyxJQUFWLENBQWVQO0FBUGIsQ0FBbEI7O0FBVUEsTUFBTVMsSUFBSSxHQUFHLENBQUM7QUFBRVosTUFBRjtBQUFRSSxPQUFSO0FBQWVDLFVBQWY7QUFBeUJDLHNCQUF6QjtBQUErQ0MsUUFBL0M7QUFBdURFLGNBQXZEO0FBQXFFRTtBQUFyRSxDQUFELEtBQXlGO0FBQ3BHLFNBQVE7QUFBTSxhQUFTLEVBQUMsTUFBaEI7QUFBdUIsUUFBSSxFQUFDLE1BQTVCO0FBQW1DLFlBQVEsRUFBRUEsWUFBN0M7QUFBMkQsY0FBVTtBQUFyRSxLQUNOO0FBQUksYUFBUyxFQUFDO0FBQWQsZUFETSxFQUVOO0FBQUssYUFBUyxFQUFDO0FBQWYsS0FDRTtBQUFPLFdBQU8sRUFBQyxNQUFmO0FBQXNCLGFBQVMsRUFBQztBQUFoQyxhQURGLEVBRUU7QUFBTyxRQUFJLEVBQUMsTUFBWjtBQUNPLGFBQVMsbURBQTRDSixNQUFNLENBQUNNLEdBQVAsQ0FBVyxNQUFYLEtBQXNCLFlBQWxFLENBRGhCO0FBRU8sUUFBSSxFQUFDLE1BRlo7QUFHTyxNQUFFLEVBQUMsTUFIVjtBQUlPLGVBQVcsRUFBQyxXQUpuQjtBQUtPLFNBQUssRUFBRWIsSUFBSSxJQUFJLEVBTHRCO0FBTU8sWUFBUSxFQUFFYyxDQUFDLElBQUlMLFlBQVksQ0FBQ0ssQ0FBQyxDQUFDQyxNQUFGLENBQVNmLElBQVYsRUFBZ0JjLENBQUMsQ0FBQ0MsTUFBRixDQUFTQyxLQUF6QixDQU5sQztBQU9PLFlBQVEsTUFQZjtBQVFPLGFBQVM7QUFSaEIsSUFGRixFQVdHVCxNQUFNLENBQUNNLEdBQVAsQ0FBVyxNQUFYLEtBQXNCO0FBQUssYUFBUyxFQUFDO0FBQWYsS0FBbUNOLE1BQU0sQ0FBQ1UsS0FBUCxDQUFhLE1BQWIsQ0FBbkMsQ0FYekIsQ0FGTSxFQWVOO0FBQUssYUFBUyxFQUFDO0FBQWYsS0FDRTtBQUFPLFdBQU8sRUFBQyxPQUFmO0FBQXVCLGFBQVMsRUFBQztBQUFqQyxhQURGLEVBRUU7QUFBTyxRQUFJLEVBQUMsT0FBWjtBQUNPLGFBQVMsbURBQTRDVixNQUFNLENBQUNNLEdBQVAsQ0FBVyxPQUFYLEtBQXVCLFlBQW5FLENBRGhCO0FBRU8sUUFBSSxFQUFDLE9BRlo7QUFHTyxNQUFFLEVBQUMsT0FIVjtBQUlPLGVBQVcsRUFBQyxlQUpuQjtBQUtPLFNBQUssRUFBRVQsS0FBSyxJQUFJLEVBTHZCO0FBTU8sWUFBUSxFQUFFVSxDQUFDLElBQUlMLFlBQVksQ0FBQ0ssQ0FBQyxDQUFDQyxNQUFGLENBQVNmLElBQVYsRUFBZ0JjLENBQUMsQ0FBQ0MsTUFBRixDQUFTQyxLQUF6QixDQU5sQztBQU9PLFlBQVEsTUFQZjtBQVFPLGFBQVM7QUFSaEIsSUFGRixFQVdHVCxNQUFNLENBQUNNLEdBQVAsQ0FBVyxPQUFYLEtBQXVCO0FBQUssYUFBUyxFQUFDO0FBQWYsS0FBbUNOLE1BQU0sQ0FBQ1UsS0FBUCxDQUFhLE9BQWIsQ0FBbkMsQ0FYMUIsQ0FmTSxFQTRCTjtBQUFLLGFBQVMsRUFBQztBQUFmLEtBQ0U7QUFBTyxXQUFPLEVBQUMsVUFBZjtBQUEwQixhQUFTLEVBQUM7QUFBcEMsZ0JBREYsRUFFRTtBQUFPLFFBQUksRUFBQyxVQUFaO0FBQ08sYUFBUyxtREFBNENWLE1BQU0sQ0FBQ00sR0FBUCxDQUFXLFVBQVgsS0FBMEIsWUFBdEUsQ0FEaEI7QUFFTyxNQUFFLEVBQUMsVUFGVjtBQUdPLFFBQUksRUFBQyxVQUhaO0FBSU8sZUFBVyxFQUFDLFVBSm5CO0FBS08sU0FBSyxFQUFFUixRQUFRLElBQUksRUFMMUI7QUFNTyxZQUFRLEVBQUVTLENBQUMsSUFBSUwsWUFBWSxDQUFDSyxDQUFDLENBQUNDLE1BQUYsQ0FBU2YsSUFBVixFQUFnQmMsQ0FBQyxDQUFDQyxNQUFGLENBQVNDLEtBQXpCLENBTmxDO0FBT08sWUFBUTtBQVBmLElBRkYsRUFVR1QsTUFBTSxDQUFDTSxHQUFQLENBQVcsVUFBWCxLQUEwQjtBQUFLLGFBQVMsRUFBQztBQUFmLEtBQW1DTixNQUFNLENBQUNVLEtBQVAsQ0FBYSxVQUFiLENBQW5DLENBVjdCLENBNUJNLEVBd0NOO0FBQUssYUFBUyxFQUFDO0FBQWYsS0FDRTtBQUFPLFdBQU8sRUFBQyxzQkFBZjtBQUFzQyxhQUFTLEVBQUM7QUFBaEQsd0JBREYsRUFFRTtBQUFPLFFBQUksRUFBQyxVQUFaO0FBQ08sYUFBUyxtREFBNENWLE1BQU0sQ0FBQ00sR0FBUCxDQUFXLHNCQUFYLEtBQXNDLFlBQWxGLENBRGhCO0FBRU8sTUFBRSxFQUFDLHNCQUZWO0FBR08sUUFBSSxFQUFDLHNCQUhaO0FBSU8sZUFBVyxFQUFDLGtCQUpuQjtBQUtPLFNBQUssRUFBRVAsb0JBQW9CLElBQUksRUFMdEM7QUFNTyxZQUFRLEVBQUVRLENBQUMsSUFBSUwsWUFBWSxDQUFDSyxDQUFDLENBQUNDLE1BQUYsQ0FBU2YsSUFBVixFQUFnQmMsQ0FBQyxDQUFDQyxNQUFGLENBQVNDLEtBQXpCLENBTmxDO0FBT08sWUFBUTtBQVBmLElBRkYsRUFVR1QsTUFBTSxDQUFDTSxHQUFQLENBQVcsc0JBQVgsS0FBc0M7QUFBSyxhQUFTLEVBQUM7QUFBZixLQUFtQ04sTUFBTSxDQUFDVSxLQUFQLENBQWEsc0JBQWIsQ0FBbkMsQ0FWekMsQ0F4Q00sRUFvRE47QUFBUSxhQUFTLEVBQUMsa0NBQWxCO0FBQ1EsUUFBSSxFQUFDLFFBRGI7QUFFUSxZQUFRLEVBQUVWLE1BQU0sQ0FBQ1csR0FBUDtBQUZsQixlQXBETSxFQXVETixpRkFBVSwyREFBQyxxREFBRDtBQUFNLE1BQUUsRUFBQyxRQUFUO0FBQWtCLGFBQVMsRUFBQztBQUE1QixhQUFWLENBdkRNLENBQVI7QUF5REQsQ0ExREQ7O0FBNERBTixJQUFJLENBQUNkLFdBQUwsR0FBbUJBLFdBQW5CO0FBQ0FjLElBQUksQ0FBQ2IsU0FBTCxHQUFpQkEsU0FBakI7QUFFZWEsbUVBQWYiLCJmaWxlIjoiLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hdXRoL3BhZ2VzL3JlZ2lzdGVyL2NvbXBvbmVudHMvRm9ybS5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBSZWFjdCBmcm9tICdyZWFjdCdcbmltcG9ydCBQcm9wVHlwZXMgZnJvbSAncHJvcC10eXBlcydcbmltcG9ydCB7IExpbmsgfSBmcm9tICdyZWFjdC1yb3V0ZXItZG9tJ1xuXG5jb25zdCBkaXNwbGF5TmFtZSA9ICdSZWdpc3RlckZyb20nXG5cbmNvbnN0IHByb3BUeXBlcyA9IHtcbiAgbmFtZTogUHJvcFR5cGVzLnN0cmluZy5pc1JlcXVpcmVkLFxuICBlbWFpbDogUHJvcFR5cGVzLnN0cmluZy5pc1JlcXVpcmVkLFxuICBwYXNzd29yZDogUHJvcFR5cGVzLnN0cmluZy5pc1JlcXVpcmVkLFxuICBwYXNzd29yZENvbmZpcm1hdGlvbjogUHJvcFR5cGVzLnN0cmluZy5pc1JlcXVpcmVkLFxuICBlcnJvcnM6IFByb3BUeXBlcy5vYmplY3QuaXNSZXF1aXJlZCxcbiAgaGFuZGxlQ2hhbmdlOiBQcm9wVHlwZXMuZnVuYy5pc1JlcXVpcmVkLFxuICBoYW5kbGVTdWJtaXQ6IFByb3BUeXBlcy5mdW5jLmlzUmVxdWlyZWQsXG59XG5cbmNvbnN0IEZvcm0gPSAoeyBuYW1lLCBlbWFpbCwgcGFzc3dvcmQsIHBhc3N3b3JkQ29uZmlybWF0aW9uLCBlcnJvcnMsIGhhbmRsZUNoYW5nZSwgaGFuZGxlU3VibWl0IH0pID0+IHtcbiAgcmV0dXJuICg8Zm9ybSBjbGFzc05hbWU9XCJmb3JtXCIgcm9sZT1cImZvcm1cIiBvblN1Ym1pdD17aGFuZGxlU3VibWl0fSBub1ZhbGlkYXRlPlxuICAgIDxoMiBjbGFzc05hbWU9XCJjYXJkLXRpdGxlXCI+U2lnbiB1cDwvaDI+XG4gICAgPGRpdiBjbGFzc05hbWU9XCJmb3JtLWdyb3VwXCI+XG4gICAgICA8bGFiZWwgaHRtbEZvcj1cIm5hbWVcIiBjbGFzc05hbWU9XCJzci1vbmx5XCI+RW1haWw8L2xhYmVsPlxuICAgICAgPGlucHV0IHR5cGU9XCJ0ZXh0XCJcbiAgICAgICAgICAgICBjbGFzc05hbWU9e2Bmb3JtLWNvbnRyb2wgZm9ybS1jb250cm9sLWxnIHJvdW5kZWQtMCAke2Vycm9ycy5oYXMoJ25hbWUnKSAmJiAnaXMtaW52YWxpZCd9YH1cbiAgICAgICAgICAgICBuYW1lPVwibmFtZVwiXG4gICAgICAgICAgICAgaWQ9XCJuYW1lXCJcbiAgICAgICAgICAgICBwbGFjZWhvbGRlcj1cIkZ1bGwgTmFtZVwiXG4gICAgICAgICAgICAgdmFsdWU9e25hbWUgfHwgJyd9XG4gICAgICAgICAgICAgb25DaGFuZ2U9e2UgPT4gaGFuZGxlQ2hhbmdlKGUudGFyZ2V0Lm5hbWUsIGUudGFyZ2V0LnZhbHVlKX1cbiAgICAgICAgICAgICByZXF1aXJlZFxuICAgICAgICAgICAgIGF1dG9Gb2N1cy8+XG4gICAgICB7ZXJyb3JzLmhhcygnbmFtZScpICYmIDxkaXYgY2xhc3NOYW1lPVwiaW52YWxpZC1mZWVkYmFja1wiPntlcnJvcnMuZmlyc3QoJ25hbWUnKX08L2Rpdj59XG4gICAgPC9kaXY+XG4gICAgPGRpdiBjbGFzc05hbWU9XCJmb3JtLWdyb3VwXCI+XG4gICAgICA8bGFiZWwgaHRtbEZvcj1cImVtYWlsXCIgY2xhc3NOYW1lPVwic3Itb25seVwiPkVtYWlsPC9sYWJlbD5cbiAgICAgIDxpbnB1dCB0eXBlPVwiZW1haWxcIlxuICAgICAgICAgICAgIGNsYXNzTmFtZT17YGZvcm0tY29udHJvbCBmb3JtLWNvbnRyb2wtbGcgcm91bmRlZC0wICR7ZXJyb3JzLmhhcygnZW1haWwnKSAmJiAnaXMtaW52YWxpZCd9YH1cbiAgICAgICAgICAgICBuYW1lPVwiZW1haWxcIlxuICAgICAgICAgICAgIGlkPVwiZW1haWxcIlxuICAgICAgICAgICAgIHBsYWNlaG9sZGVyPVwiRW1haWwgYWRkcmVzc1wiXG4gICAgICAgICAgICAgdmFsdWU9e2VtYWlsIHx8ICcnfVxuICAgICAgICAgICAgIG9uQ2hhbmdlPXtlID0+IGhhbmRsZUNoYW5nZShlLnRhcmdldC5uYW1lLCBlLnRhcmdldC52YWx1ZSl9XG4gICAgICAgICAgICAgcmVxdWlyZWRcbiAgICAgICAgICAgICBhdXRvRm9jdXMvPlxuICAgICAge2Vycm9ycy5oYXMoJ2VtYWlsJykgJiYgPGRpdiBjbGFzc05hbWU9XCJpbnZhbGlkLWZlZWRiYWNrXCI+e2Vycm9ycy5maXJzdCgnZW1haWwnKX08L2Rpdj59XG4gICAgPC9kaXY+XG4gICAgPGRpdiBjbGFzc05hbWU9XCJmb3JtLWdyb3VwXCI+XG4gICAgICA8bGFiZWwgaHRtbEZvcj1cInBhc3N3b3JkXCIgY2xhc3NOYW1lPVwic3Itb25seVwiPlBhc3N3b3JkPC9sYWJlbD5cbiAgICAgIDxpbnB1dCB0eXBlPVwicGFzc3dvcmRcIlxuICAgICAgICAgICAgIGNsYXNzTmFtZT17YGZvcm0tY29udHJvbCBmb3JtLWNvbnRyb2wtbGcgcm91bmRlZC0wICR7ZXJyb3JzLmhhcygncGFzc3dvcmQnKSAmJiAnaXMtaW52YWxpZCd9YH1cbiAgICAgICAgICAgICBpZD1cInBhc3N3b3JkXCJcbiAgICAgICAgICAgICBuYW1lPVwicGFzc3dvcmRcIlxuICAgICAgICAgICAgIHBsYWNlaG9sZGVyPVwiUGFzc3dvcmRcIlxuICAgICAgICAgICAgIHZhbHVlPXtwYXNzd29yZCB8fCAnJ31cbiAgICAgICAgICAgICBvbkNoYW5nZT17ZSA9PiBoYW5kbGVDaGFuZ2UoZS50YXJnZXQubmFtZSwgZS50YXJnZXQudmFsdWUpfVxuICAgICAgICAgICAgIHJlcXVpcmVkLz5cbiAgICAgIHtlcnJvcnMuaGFzKCdwYXNzd29yZCcpICYmIDxkaXYgY2xhc3NOYW1lPVwiaW52YWxpZC1mZWVkYmFja1wiPntlcnJvcnMuZmlyc3QoJ3Bhc3N3b3JkJyl9PC9kaXY+fVxuICAgIDwvZGl2PlxuICAgIDxkaXYgY2xhc3NOYW1lPVwiZm9ybS1ncm91cFwiPlxuICAgICAgPGxhYmVsIGh0bWxGb3I9XCJwYXNzd29yZENvbmZpcm1hdGlvblwiIGNsYXNzTmFtZT1cInNyLW9ubHlcIj5Db25maXJtIFBhc3N3b3JkPC9sYWJlbD5cbiAgICAgIDxpbnB1dCB0eXBlPVwicGFzc3dvcmRcIlxuICAgICAgICAgICAgIGNsYXNzTmFtZT17YGZvcm0tY29udHJvbCBmb3JtLWNvbnRyb2wtbGcgcm91bmRlZC0wICR7ZXJyb3JzLmhhcygncGFzc3dvcmRDb25maXJtYXRpb24nKSAmJiAnaXMtaW52YWxpZCd9YH1cbiAgICAgICAgICAgICBpZD1cInBhc3N3b3JkQ29uZmlybWF0aW9uXCJcbiAgICAgICAgICAgICBuYW1lPVwicGFzc3dvcmRDb25maXJtYXRpb25cIlxuICAgICAgICAgICAgIHBsYWNlaG9sZGVyPVwiQ29uZmlybSBQYXNzd29yZFwiXG4gICAgICAgICAgICAgdmFsdWU9e3Bhc3N3b3JkQ29uZmlybWF0aW9uIHx8ICcnfVxuICAgICAgICAgICAgIG9uQ2hhbmdlPXtlID0+IGhhbmRsZUNoYW5nZShlLnRhcmdldC5uYW1lLCBlLnRhcmdldC52YWx1ZSl9XG4gICAgICAgICAgICAgcmVxdWlyZWQvPlxuICAgICAge2Vycm9ycy5oYXMoJ3Bhc3N3b3JkQ29uZmlybWF0aW9uJykgJiYgPGRpdiBjbGFzc05hbWU9XCJpbnZhbGlkLWZlZWRiYWNrXCI+e2Vycm9ycy5maXJzdCgncGFzc3dvcmRDb25maXJtYXRpb24nKX08L2Rpdj59XG4gICAgPC9kaXY+XG4gICAgPGJ1dHRvbiBjbGFzc05hbWU9XCJidG4gYnRuLWxnIGJ0bi1wcmltYXJ5IGJ0bi1ibG9ja1wiXG4gICAgICAgICAgICB0eXBlPVwic3VibWl0XCJcbiAgICAgICAgICAgIGRpc2FibGVkPXtlcnJvcnMuYW55KCl9PlNpZ24gVXA8L2J1dHRvbj5cbiAgICA8cD5CYWNrIHRvPExpbmsgdG89XCIvbG9naW5cIiBjbGFzc05hbWU9XCJidG4gYnRuLWxpbmtcIj5Mb2dpbjwvTGluaz48L3A+XG4gIDwvZm9ybT4pXG59XG5cbkZvcm0uZGlzcGxheU5hbWUgPSBkaXNwbGF5TmFtZVxuRm9ybS5wcm9wVHlwZXMgPSBwcm9wVHlwZXNcblxuZXhwb3J0IGRlZmF1bHQgRm9ybVxuIl0sInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./resources/js/modules/auth/pages/register/components/Form.js\n");
 
 /***/ }),
 
@@ -32,7 +32,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var reac
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react_redux__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n/* harmony import */ var _Page__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./Page */ \"./resources/js/modules/auth/pages/register/Page.js\");\n// import libs\n // import components\n\n\n\nvar mapStateToProps = function mapStateToProps(state) {\n  return {\n    isAuthenticated: state.auth.isAuthenticated\n  };\n};\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Object(react_redux__WEBPACK_IMPORTED_MODULE_0__[\"connect\"])(mapStateToProps)(_Page__WEBPACK_IMPORTED_MODULE_1__[\"default\"]));//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hdXRoL3BhZ2VzL3JlZ2lzdGVyL2luZGV4LmpzP2I2MTEiXSwibmFtZXMiOlsibWFwU3RhdGVUb1Byb3BzIiwic3RhdGUiLCJpc0F1dGhlbnRpY2F0ZWQiLCJhdXRoIiwiY29ubmVjdCIsIlBhZ2UiXSwibWFwcGluZ3MiOiJBQUFBO0FBQUE7QUFBQTtBQUFBO0NBR0E7O0FBQ0E7O0FBRUEsSUFBTUEsZUFBZSxHQUFHLFNBQWxCQSxlQUFrQixDQUFBQyxLQUFLLEVBQUk7QUFDL0IsU0FBTztBQUNMQyxtQkFBZSxFQUFFRCxLQUFLLENBQUNFLElBQU4sQ0FBV0Q7QUFEdkIsR0FBUDtBQUdELENBSkQ7O0FBTWVFLDBIQUFPLENBQUNKLGVBQUQsQ0FBUCxDQUF5QkssNkNBQXpCLENBQWYiLCJmaWxlIjoiLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hdXRoL3BhZ2VzL3JlZ2lzdGVyL2luZGV4LmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiLy8gaW1wb3J0IGxpYnNcbmltcG9ydCB7IGNvbm5lY3QgfSBmcm9tICdyZWFjdC1yZWR1eCdcblxuLy8gaW1wb3J0IGNvbXBvbmVudHNcbmltcG9ydCBQYWdlIGZyb20gJy4vUGFnZSdcblxuY29uc3QgbWFwU3RhdGVUb1Byb3BzID0gc3RhdGUgPT4ge1xuICByZXR1cm4ge1xuICAgIGlzQXV0aGVudGljYXRlZDogc3RhdGUuYXV0aC5pc0F1dGhlbnRpY2F0ZWQsXG4gIH1cbn1cblxuZXhwb3J0IGRlZmF1bHQgY29ubmVjdChtYXBTdGF0ZVRvUHJvcHMpKFBhZ2UpXG4iXSwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./resources/js/modules/auth/pages/register/index.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react_redux__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n/* harmony import */ var _Page__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./Page */ \"./resources/js/modules/auth/pages/register/Page.js\");\n// import libs\n // import components\n\n\n\nconst mapStateToProps = state => {\n  return {\n    isAuthenticated: state.auth.isAuthenticated\n  };\n};\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Object(react_redux__WEBPACK_IMPORTED_MODULE_0__[\"connect\"])(mapStateToProps)(_Page__WEBPACK_IMPORTED_MODULE_1__[\"default\"]));//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hdXRoL3BhZ2VzL3JlZ2lzdGVyL2luZGV4LmpzP2I2MTEiXSwibmFtZXMiOlsibWFwU3RhdGVUb1Byb3BzIiwic3RhdGUiLCJpc0F1dGhlbnRpY2F0ZWQiLCJhdXRoIiwiY29ubmVjdCIsIlBhZ2UiXSwibWFwcGluZ3MiOiJBQUFBO0FBQUE7QUFBQTtBQUFBO0NBR0E7O0FBQ0E7O0FBRUEsTUFBTUEsZUFBZSxHQUFHQyxLQUFLLElBQUk7QUFDL0IsU0FBTztBQUNMQyxtQkFBZSxFQUFFRCxLQUFLLENBQUNFLElBQU4sQ0FBV0Q7QUFEdkIsR0FBUDtBQUdELENBSkQ7O0FBTWVFLDBIQUFPLENBQUNKLGVBQUQsQ0FBUCxDQUF5QkssNkNBQXpCLENBQWYiLCJmaWxlIjoiLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hdXRoL3BhZ2VzL3JlZ2lzdGVyL2luZGV4LmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiLy8gaW1wb3J0IGxpYnNcbmltcG9ydCB7IGNvbm5lY3QgfSBmcm9tICdyZWFjdC1yZWR1eCdcblxuLy8gaW1wb3J0IGNvbXBvbmVudHNcbmltcG9ydCBQYWdlIGZyb20gJy4vUGFnZSdcblxuY29uc3QgbWFwU3RhdGVUb1Byb3BzID0gc3RhdGUgPT4ge1xuICByZXR1cm4ge1xuICAgIGlzQXV0aGVudGljYXRlZDogc3RhdGUuYXV0aC5pc0F1dGhlbnRpY2F0ZWQsXG4gIH1cbn1cblxuZXhwb3J0IGRlZmF1bHQgY29ubmVjdChtYXBTdGF0ZVRvUHJvcHMpKFBhZ2UpXG4iXSwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./resources/js/modules/auth/pages/register/index.js\n");
 
 /***/ })
 
