@@ -8,7 +8,7 @@
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\");\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var react_router_dom__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! react-router-dom */ \"./node_modules/react-router-dom/esm/react-router-dom.js\");\n// import libs\n\n // import components\n\n\nvar displayName = 'ArticleComponent';\nvar propTypes = {\n  index: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.number.isRequired,\n  article: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.object.isRequired\n}; // const renderAuthor = (article) => {\n//   return article.user && `By ${ article.user.name }`\n// }\n\nvar renderPublishedAt = function renderPublishedAt(article) {\n  return article.publishedAt && \"at \".concat(article.publishedAt.format('MMMM D, YYYY'));\n};\n\nfunction render(_ref) {\n  var article = _ref.article;\n  return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"col-12 col-sm-9 mb-5 mx-auto\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"card\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"card-body\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"h4\", {\n    className: \"card-title\"\n  }, article.title), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"h6\", {\n    className: \"card-subtitle mb-2 text-muted\"\n  }, renderPublishedAt(article)), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"p\", {\n    className: \"card-text\"\n  }, article.description), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(react_router_dom__WEBPACK_IMPORTED_MODULE_2__[\"Link\"], {\n    to: \"blog/\".concat(article.slug),\n    className: \"card-link\"\n  }, \"Read More\"))));\n}\n\nrender.displayName = displayName;\nrender.propTypes = propTypes;\n/* harmony default export */ __webpack_exports__[\"default\"] = (render);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvY29tbW9uL2FydGljbGVzL2xpc3RpbmcvY29tcG9uZW50cy9BcnRpY2xlLmpzPzFkNjYiXSwibmFtZXMiOlsiZGlzcGxheU5hbWUiLCJwcm9wVHlwZXMiLCJpbmRleCIsIlByb3BUeXBlcyIsIm51bWJlciIsImlzUmVxdWlyZWQiLCJhcnRpY2xlIiwib2JqZWN0IiwicmVuZGVyUHVibGlzaGVkQXQiLCJwdWJsaXNoZWRBdCIsImZvcm1hdCIsInJlbmRlciIsInRpdGxlIiwiZGVzY3JpcHRpb24iLCJzbHVnIl0sIm1hcHBpbmdzIjoiQUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUNBO0NBR0E7O0FBQ0E7QUFFQSxJQUFNQSxXQUFXLEdBQUcsa0JBQXBCO0FBQ0EsSUFBTUMsU0FBUyxHQUFHO0FBQ2hCQyxPQUFLLEVBQUVDLGlEQUFTLENBQUNDLE1BQVYsQ0FBaUJDLFVBRFI7QUFFaEJDLFNBQU8sRUFBRUgsaURBQVMsQ0FBQ0ksTUFBVixDQUFpQkY7QUFGVixDQUFsQixDLENBS0E7QUFDQTtBQUNBOztBQUVBLElBQU1HLGlCQUFpQixHQUFHLFNBQXBCQSxpQkFBb0IsQ0FBQ0YsT0FBRCxFQUFhO0FBQ3JDLFNBQU9BLE9BQU8sQ0FBQ0csV0FBUixpQkFBNkJILE9BQU8sQ0FBQ0csV0FBUixDQUFvQkMsTUFBcEIsQ0FBMkIsY0FBM0IsQ0FBN0IsQ0FBUDtBQUNELENBRkQ7O0FBSUEsU0FBU0MsTUFBVCxPQUE4QjtBQUFBLE1BQVhMLE9BQVcsUUFBWEEsT0FBVztBQUM1QixTQUFPO0FBQUssYUFBUyxFQUFDO0FBQWYsS0FDTDtBQUFLLGFBQVMsRUFBQztBQUFmLEtBQ0U7QUFBSyxhQUFTLEVBQUM7QUFBZixLQUNFO0FBQUksYUFBUyxFQUFDO0FBQWQsS0FBNEJBLE9BQU8sQ0FBQ00sS0FBcEMsQ0FERixFQUVFO0FBQUksYUFBUyxFQUFDO0FBQWQsS0FBK0NKLGlCQUFpQixDQUFDRixPQUFELENBQWhFLENBRkYsRUFHRTtBQUFHLGFBQVMsRUFBQztBQUFiLEtBQTJCQSxPQUFPLENBQUNPLFdBQW5DLENBSEYsRUFJRSwyREFBQyxxREFBRDtBQUFNLE1BQUUsaUJBQVVQLE9BQU8sQ0FBQ1EsSUFBbEIsQ0FBUjtBQUFrQyxhQUFTLEVBQUM7QUFBNUMsaUJBSkYsQ0FERixDQURLLENBQVA7QUFVRDs7QUFFREgsTUFBTSxDQUFDWCxXQUFQLEdBQXFCQSxXQUFyQjtBQUNBVyxNQUFNLENBQUNWLFNBQVAsR0FBbUJBLFNBQW5CO0FBRWVVLHFFQUFmIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL2NvbW1vbi9hcnRpY2xlcy9saXN0aW5nL2NvbXBvbmVudHMvQXJ0aWNsZS5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbIi8vIGltcG9ydCBsaWJzXG5pbXBvcnQgUmVhY3QgZnJvbSAncmVhY3QnXG5pbXBvcnQgUHJvcFR5cGVzIGZyb20gJ3Byb3AtdHlwZXMnXG5cbi8vIGltcG9ydCBjb21wb25lbnRzXG5pbXBvcnQgeyBMaW5rIH0gZnJvbSAncmVhY3Qtcm91dGVyLWRvbSdcblxuY29uc3QgZGlzcGxheU5hbWUgPSAnQXJ0aWNsZUNvbXBvbmVudCdcbmNvbnN0IHByb3BUeXBlcyA9IHtcbiAgaW5kZXg6IFByb3BUeXBlcy5udW1iZXIuaXNSZXF1aXJlZCxcbiAgYXJ0aWNsZTogUHJvcFR5cGVzLm9iamVjdC5pc1JlcXVpcmVkLFxufVxuXG4vLyBjb25zdCByZW5kZXJBdXRob3IgPSAoYXJ0aWNsZSkgPT4ge1xuLy8gICByZXR1cm4gYXJ0aWNsZS51c2VyICYmIGBCeSAkeyBhcnRpY2xlLnVzZXIubmFtZSB9YFxuLy8gfVxuXG5jb25zdCByZW5kZXJQdWJsaXNoZWRBdCA9IChhcnRpY2xlKSA9PiB7XG4gIHJldHVybiBhcnRpY2xlLnB1Ymxpc2hlZEF0ICYmIGBhdCAke2FydGljbGUucHVibGlzaGVkQXQuZm9ybWF0KCdNTU1NIEQsIFlZWVknKX1gXG59XG5cbmZ1bmN0aW9uIHJlbmRlciAoeyBhcnRpY2xlIH0pIHtcbiAgcmV0dXJuIDxkaXYgY2xhc3NOYW1lPVwiY29sLTEyIGNvbC1zbS05IG1iLTUgbXgtYXV0b1wiPlxuICAgIDxkaXYgY2xhc3NOYW1lPVwiY2FyZFwiPlxuICAgICAgPGRpdiBjbGFzc05hbWU9XCJjYXJkLWJvZHlcIj5cbiAgICAgICAgPGg0IGNsYXNzTmFtZT1cImNhcmQtdGl0bGVcIj57YXJ0aWNsZS50aXRsZX08L2g0PlxuICAgICAgICA8aDYgY2xhc3NOYW1lPVwiY2FyZC1zdWJ0aXRsZSBtYi0yIHRleHQtbXV0ZWRcIj57cmVuZGVyUHVibGlzaGVkQXQoYXJ0aWNsZSl9PC9oNj5cbiAgICAgICAgPHAgY2xhc3NOYW1lPVwiY2FyZC10ZXh0XCI+eyBhcnRpY2xlLmRlc2NyaXB0aW9uIH08L3A+XG4gICAgICAgIDxMaW5rIHRvPXtgYmxvZy8ke2FydGljbGUuc2x1Z31gfSBjbGFzc05hbWU9XCJjYXJkLWxpbmtcIj5SZWFkIE1vcmU8L0xpbms+XG4gICAgICA8L2Rpdj5cbiAgICA8L2Rpdj5cbiAgPC9kaXY+XG59XG5cbnJlbmRlci5kaXNwbGF5TmFtZSA9IGRpc3BsYXlOYW1lXG5yZW5kZXIucHJvcFR5cGVzID0gcHJvcFR5cGVzXG5cbmV4cG9ydCBkZWZhdWx0IHJlbmRlclxuIl0sInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./resources/js/common/articles/listing/components/Article.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\");\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var react_router_dom__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! react-router-dom */ \"./node_modules/react-router-dom/esm/react-router-dom.js\");\n// import libs\n\n // import components\n\n\nconst displayName = 'ArticleComponent';\nconst propTypes = {\n  index: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.number.isRequired,\n  article: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.object.isRequired\n}; // const renderAuthor = (article) => {\n//   return article.user && `By ${ article.user.name }`\n// }\n\nconst renderPublishedAt = article => {\n  return article.publishedAt && \"at \".concat(article.publishedAt.format('MMMM D, YYYY'));\n};\n\nfunction render({\n  article\n}) {\n  return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"col-12 col-sm-9 mb-5 mx-auto\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"card\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"card-body\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"h4\", {\n    className: \"card-title\"\n  }, article.title), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"h6\", {\n    className: \"card-subtitle mb-2 text-muted\"\n  }, renderPublishedAt(article)), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"p\", {\n    className: \"card-text\"\n  }, article.description), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(react_router_dom__WEBPACK_IMPORTED_MODULE_2__[\"Link\"], {\n    to: \"blog/\".concat(article.slug),\n    className: \"card-link\"\n  }, \"Read More\"))));\n}\n\nrender.displayName = displayName;\nrender.propTypes = propTypes;\n/* harmony default export */ __webpack_exports__[\"default\"] = (render);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvY29tbW9uL2FydGljbGVzL2xpc3RpbmcvY29tcG9uZW50cy9BcnRpY2xlLmpzPzFkNjYiXSwibmFtZXMiOlsiZGlzcGxheU5hbWUiLCJwcm9wVHlwZXMiLCJpbmRleCIsIlByb3BUeXBlcyIsIm51bWJlciIsImlzUmVxdWlyZWQiLCJhcnRpY2xlIiwib2JqZWN0IiwicmVuZGVyUHVibGlzaGVkQXQiLCJwdWJsaXNoZWRBdCIsImZvcm1hdCIsInJlbmRlciIsInRpdGxlIiwiZGVzY3JpcHRpb24iLCJzbHVnIl0sIm1hcHBpbmdzIjoiQUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUNBO0NBR0E7O0FBQ0E7QUFFQSxNQUFNQSxXQUFXLEdBQUcsa0JBQXBCO0FBQ0EsTUFBTUMsU0FBUyxHQUFHO0FBQ2hCQyxPQUFLLEVBQUVDLGlEQUFTLENBQUNDLE1BQVYsQ0FBaUJDLFVBRFI7QUFFaEJDLFNBQU8sRUFBRUgsaURBQVMsQ0FBQ0ksTUFBVixDQUFpQkY7QUFGVixDQUFsQixDLENBS0E7QUFDQTtBQUNBOztBQUVBLE1BQU1HLGlCQUFpQixHQUFJRixPQUFELElBQWE7QUFDckMsU0FBT0EsT0FBTyxDQUFDRyxXQUFSLGlCQUE2QkgsT0FBTyxDQUFDRyxXQUFSLENBQW9CQyxNQUFwQixDQUEyQixjQUEzQixDQUE3QixDQUFQO0FBQ0QsQ0FGRDs7QUFJQSxTQUFTQyxNQUFULENBQWlCO0FBQUVMO0FBQUYsQ0FBakIsRUFBOEI7QUFDNUIsU0FBTztBQUFLLGFBQVMsRUFBQztBQUFmLEtBQ0w7QUFBSyxhQUFTLEVBQUM7QUFBZixLQUNFO0FBQUssYUFBUyxFQUFDO0FBQWYsS0FDRTtBQUFJLGFBQVMsRUFBQztBQUFkLEtBQTRCQSxPQUFPLENBQUNNLEtBQXBDLENBREYsRUFFRTtBQUFJLGFBQVMsRUFBQztBQUFkLEtBQStDSixpQkFBaUIsQ0FBQ0YsT0FBRCxDQUFoRSxDQUZGLEVBR0U7QUFBRyxhQUFTLEVBQUM7QUFBYixLQUEyQkEsT0FBTyxDQUFDTyxXQUFuQyxDQUhGLEVBSUUsMkRBQUMscURBQUQ7QUFBTSxNQUFFLGlCQUFVUCxPQUFPLENBQUNRLElBQWxCLENBQVI7QUFBa0MsYUFBUyxFQUFDO0FBQTVDLGlCQUpGLENBREYsQ0FESyxDQUFQO0FBVUQ7O0FBRURILE1BQU0sQ0FBQ1gsV0FBUCxHQUFxQkEsV0FBckI7QUFDQVcsTUFBTSxDQUFDVixTQUFQLEdBQW1CQSxTQUFuQjtBQUVlVSxxRUFBZiIsImZpbGUiOiIuL3Jlc291cmNlcy9qcy9jb21tb24vYXJ0aWNsZXMvbGlzdGluZy9jb21wb25lbnRzL0FydGljbGUuanMuanMiLCJzb3VyY2VzQ29udGVudCI6WyIvLyBpbXBvcnQgbGlic1xuaW1wb3J0IFJlYWN0IGZyb20gJ3JlYWN0J1xuaW1wb3J0IFByb3BUeXBlcyBmcm9tICdwcm9wLXR5cGVzJ1xuXG4vLyBpbXBvcnQgY29tcG9uZW50c1xuaW1wb3J0IHsgTGluayB9IGZyb20gJ3JlYWN0LXJvdXRlci1kb20nXG5cbmNvbnN0IGRpc3BsYXlOYW1lID0gJ0FydGljbGVDb21wb25lbnQnXG5jb25zdCBwcm9wVHlwZXMgPSB7XG4gIGluZGV4OiBQcm9wVHlwZXMubnVtYmVyLmlzUmVxdWlyZWQsXG4gIGFydGljbGU6IFByb3BUeXBlcy5vYmplY3QuaXNSZXF1aXJlZCxcbn1cblxuLy8gY29uc3QgcmVuZGVyQXV0aG9yID0gKGFydGljbGUpID0+IHtcbi8vICAgcmV0dXJuIGFydGljbGUudXNlciAmJiBgQnkgJHsgYXJ0aWNsZS51c2VyLm5hbWUgfWBcbi8vIH1cblxuY29uc3QgcmVuZGVyUHVibGlzaGVkQXQgPSAoYXJ0aWNsZSkgPT4ge1xuICByZXR1cm4gYXJ0aWNsZS5wdWJsaXNoZWRBdCAmJiBgYXQgJHthcnRpY2xlLnB1Ymxpc2hlZEF0LmZvcm1hdCgnTU1NTSBELCBZWVlZJyl9YFxufVxuXG5mdW5jdGlvbiByZW5kZXIgKHsgYXJ0aWNsZSB9KSB7XG4gIHJldHVybiA8ZGl2IGNsYXNzTmFtZT1cImNvbC0xMiBjb2wtc20tOSBtYi01IG14LWF1dG9cIj5cbiAgICA8ZGl2IGNsYXNzTmFtZT1cImNhcmRcIj5cbiAgICAgIDxkaXYgY2xhc3NOYW1lPVwiY2FyZC1ib2R5XCI+XG4gICAgICAgIDxoNCBjbGFzc05hbWU9XCJjYXJkLXRpdGxlXCI+e2FydGljbGUudGl0bGV9PC9oND5cbiAgICAgICAgPGg2IGNsYXNzTmFtZT1cImNhcmQtc3VidGl0bGUgbWItMiB0ZXh0LW11dGVkXCI+e3JlbmRlclB1Ymxpc2hlZEF0KGFydGljbGUpfTwvaDY+XG4gICAgICAgIDxwIGNsYXNzTmFtZT1cImNhcmQtdGV4dFwiPnsgYXJ0aWNsZS5kZXNjcmlwdGlvbiB9PC9wPlxuICAgICAgICA8TGluayB0bz17YGJsb2cvJHthcnRpY2xlLnNsdWd9YH0gY2xhc3NOYW1lPVwiY2FyZC1saW5rXCI+UmVhZCBNb3JlPC9MaW5rPlxuICAgICAgPC9kaXY+XG4gICAgPC9kaXY+XG4gIDwvZGl2PlxufVxuXG5yZW5kZXIuZGlzcGxheU5hbWUgPSBkaXNwbGF5TmFtZVxucmVuZGVyLnByb3BUeXBlcyA9IHByb3BUeXBlc1xuXG5leHBvcnQgZGVmYXVsdCByZW5kZXJcbiJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/common/articles/listing/components/Article.js\n");
 
 /***/ }),
 
@@ -20,7 +20,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var reac
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\");\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var _Article__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Article */ \"./resources/js/common/articles/listing/components/Article.js\");\nfunction _typeof(obj) { if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\n\n\n\n\nvar Articles =\n/*#__PURE__*/\nfunction (_Component) {\n  _inherits(Articles, _Component);\n\n  function Articles(props) {\n    var _this;\n\n    _classCallCheck(this, Articles);\n\n    _this = _possibleConstructorReturn(this, _getPrototypeOf(Articles).call(this, props));\n    _this.state = {//\n    };\n    return _this;\n  }\n\n  _createClass(Articles, [{\n    key: \"renderArticles\",\n    value: function renderArticles() {\n      return this.props.articles.map(function (article, index) {\n        return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_Article__WEBPACK_IMPORTED_MODULE_2__[\"default\"], {\n          key: \"article-\".concat(index),\n          index: index,\n          article: article\n        });\n      });\n    }\n  }, {\n    key: \"render\",\n    value: function render() {\n      return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"section\", {\n        id: \"components-articles\"\n      }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n        className: \"container\"\n      }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n        className: \"row\"\n      }, this.props.articles && this.renderArticles())));\n    }\n  }]);\n\n  return Articles;\n}(react__WEBPACK_IMPORTED_MODULE_0__[\"Component\"]);\n\n_defineProperty(Articles, \"displayName\", 'Articles');\n\n_defineProperty(Articles, \"propTypes\", {\n  articles: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.array.isRequired,\n  dispatch: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.func.isRequired\n});\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Articles);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvY29tbW9uL2FydGljbGVzL2xpc3RpbmcvY29tcG9uZW50cy9BcnRpY2xlcy5qcz85NDU2Il0sIm5hbWVzIjpbIkFydGljbGVzIiwicHJvcHMiLCJzdGF0ZSIsImFydGljbGVzIiwibWFwIiwiYXJ0aWNsZSIsImluZGV4IiwicmVuZGVyQXJ0aWNsZXMiLCJDb21wb25lbnQiLCJQcm9wVHlwZXMiLCJhcnJheSIsImlzUmVxdWlyZWQiLCJkaXNwYXRjaCIsImZ1bmMiXSwibWFwcGluZ3MiOiI7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7O0FBQUE7QUFDQTtBQUNBOztJQUVNQSxROzs7OztBQU9KLG9CQUFZQyxLQUFaLEVBQW1CO0FBQUE7O0FBQUE7O0FBQ2pCLGtGQUFNQSxLQUFOO0FBRUEsVUFBS0MsS0FBTCxHQUFhLENBQ1g7QUFEVyxLQUFiO0FBSGlCO0FBTWxCOzs7O3FDQUVnQjtBQUNmLGFBQU8sS0FBS0QsS0FBTCxDQUFXRSxRQUFYLENBQW9CQyxHQUFwQixDQUF3QixVQUFDQyxPQUFELEVBQVVDLEtBQVYsRUFBb0I7QUFDakQsZUFBTywyREFBQyxnREFBRDtBQUFTLGFBQUcsb0JBQWFBLEtBQWIsQ0FBWjtBQUNTLGVBQUssRUFBRUEsS0FEaEI7QUFFUyxpQkFBTyxFQUFFRDtBQUZsQixVQUFQO0FBR0QsT0FKTSxDQUFQO0FBS0Q7Ozs2QkFFUTtBQUNQLGFBQVE7QUFBUyxVQUFFLEVBQUM7QUFBWixTQUNOO0FBQUssaUJBQVMsRUFBQztBQUFmLFNBQ0U7QUFBSyxpQkFBUyxFQUFDO0FBQWYsU0FDSSxLQUFLSixLQUFMLENBQVdFLFFBQVgsSUFBdUIsS0FBS0ksY0FBTCxFQUQzQixDQURGLENBRE0sQ0FBUjtBQU9EOzs7O0VBL0JvQkMsK0M7O2dCQUFqQlIsUSxpQkFDaUIsVTs7Z0JBRGpCQSxRLGVBRWU7QUFDakJHLFVBQVEsRUFBRU0saURBQVMsQ0FBQ0MsS0FBVixDQUFnQkMsVUFEVDtBQUVqQkMsVUFBUSxFQUFFSCxpREFBUyxDQUFDSSxJQUFWLENBQWVGO0FBRlIsQzs7QUFnQ05YLHVFQUFmIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL2NvbW1vbi9hcnRpY2xlcy9saXN0aW5nL2NvbXBvbmVudHMvQXJ0aWNsZXMuanMuanMiLCJzb3VyY2VzQ29udGVudCI6WyJpbXBvcnQgUmVhY3QsIHtDb21wb25lbnR9IGZyb20gJ3JlYWN0J1xuaW1wb3J0IFByb3BUeXBlcyBmcm9tICdwcm9wLXR5cGVzJ1xuaW1wb3J0IEFydGljbGUgZnJvbSAnLi9BcnRpY2xlJ1xuXG5jbGFzcyBBcnRpY2xlcyBleHRlbmRzIENvbXBvbmVudCB7XG4gIHN0YXRpYyBkaXNwbGF5TmFtZSA9ICdBcnRpY2xlcydcbiAgc3RhdGljIHByb3BUeXBlcyA9IHtcbiAgICBhcnRpY2xlczogUHJvcFR5cGVzLmFycmF5LmlzUmVxdWlyZWQsXG4gICAgZGlzcGF0Y2g6IFByb3BUeXBlcy5mdW5jLmlzUmVxdWlyZWQsXG4gIH1cbiAgXG4gIGNvbnN0cnVjdG9yKHByb3BzKSB7XG4gICAgc3VwZXIocHJvcHMpXG4gICAgXG4gICAgdGhpcy5zdGF0ZSA9IHtcbiAgICAgIC8vXG4gICAgfVxuICB9XG4gIFxuICByZW5kZXJBcnRpY2xlcygpIHtcbiAgICByZXR1cm4gdGhpcy5wcm9wcy5hcnRpY2xlcy5tYXAoKGFydGljbGUsIGluZGV4KSA9PiB7XG4gICAgICByZXR1cm4gPEFydGljbGUga2V5PXtgYXJ0aWNsZS0ke2luZGV4fWB9XG4gICAgICAgICAgICAgICAgICAgICAgaW5kZXg9e2luZGV4fVxuICAgICAgICAgICAgICAgICAgICAgIGFydGljbGU9e2FydGljbGV9Lz5cbiAgICB9KVxuICB9XG4gIFxuICByZW5kZXIoKSB7XG4gICAgcmV0dXJuICg8c2VjdGlvbiBpZD1cImNvbXBvbmVudHMtYXJ0aWNsZXNcIj5cbiAgICAgIDxkaXYgY2xhc3NOYW1lPVwiY29udGFpbmVyXCI+XG4gICAgICAgIDxkaXYgY2xhc3NOYW1lPVwicm93XCI+XG4gICAgICAgICAgeyB0aGlzLnByb3BzLmFydGljbGVzICYmIHRoaXMucmVuZGVyQXJ0aWNsZXMoKSB9XG4gICAgICAgIDwvZGl2PlxuICAgICAgPC9kaXY+XG4gICAgPC9zZWN0aW9uPilcbiAgfVxufVxuXG5leHBvcnQgZGVmYXVsdCBBcnRpY2xlc1xuIl0sInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./resources/js/common/articles/listing/components/Articles.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\");\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var _Article__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Article */ \"./resources/js/common/articles/listing/components/Article.js\");\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\n\n\n\n\nclass Articles extends react__WEBPACK_IMPORTED_MODULE_0__[\"Component\"] {\n  constructor(props) {\n    super(props);\n    this.state = {//\n    };\n  }\n\n  renderArticles() {\n    return this.props.articles.map((article, index) => {\n      return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_Article__WEBPACK_IMPORTED_MODULE_2__[\"default\"], {\n        key: \"article-\".concat(index),\n        index: index,\n        article: article\n      });\n    });\n  }\n\n  render() {\n    return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"section\", {\n      id: \"components-articles\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"container\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"row\"\n    }, this.props.articles && this.renderArticles())));\n  }\n\n}\n\n_defineProperty(Articles, \"displayName\", 'Articles');\n\n_defineProperty(Articles, \"propTypes\", {\n  articles: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.array.isRequired,\n  dispatch: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.func.isRequired\n});\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Articles);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvY29tbW9uL2FydGljbGVzL2xpc3RpbmcvY29tcG9uZW50cy9BcnRpY2xlcy5qcz85NDU2Il0sIm5hbWVzIjpbIkFydGljbGVzIiwiQ29tcG9uZW50IiwiY29uc3RydWN0b3IiLCJwcm9wcyIsInN0YXRlIiwicmVuZGVyQXJ0aWNsZXMiLCJhcnRpY2xlcyIsIm1hcCIsImFydGljbGUiLCJpbmRleCIsInJlbmRlciIsIlByb3BUeXBlcyIsImFycmF5IiwiaXNSZXF1aXJlZCIsImRpc3BhdGNoIiwiZnVuYyJdLCJtYXBwaW5ncyI6Ijs7Ozs7Ozs7QUFBQTtBQUNBO0FBQ0E7O0FBRUEsTUFBTUEsUUFBTixTQUF1QkMsK0NBQXZCLENBQWlDO0FBTy9CQyxhQUFXLENBQUNDLEtBQUQsRUFBUTtBQUNqQixVQUFNQSxLQUFOO0FBRUEsU0FBS0MsS0FBTCxHQUFhLENBQ1g7QUFEVyxLQUFiO0FBR0Q7O0FBRURDLGdCQUFjLEdBQUc7QUFDZixXQUFPLEtBQUtGLEtBQUwsQ0FBV0csUUFBWCxDQUFvQkMsR0FBcEIsQ0FBd0IsQ0FBQ0MsT0FBRCxFQUFVQyxLQUFWLEtBQW9CO0FBQ2pELGFBQU8sMkRBQUMsZ0RBQUQ7QUFBUyxXQUFHLG9CQUFhQSxLQUFiLENBQVo7QUFDUyxhQUFLLEVBQUVBLEtBRGhCO0FBRVMsZUFBTyxFQUFFRDtBQUZsQixRQUFQO0FBR0QsS0FKTSxDQUFQO0FBS0Q7O0FBRURFLFFBQU0sR0FBRztBQUNQLFdBQVE7QUFBUyxRQUFFLEVBQUM7QUFBWixPQUNOO0FBQUssZUFBUyxFQUFDO0FBQWYsT0FDRTtBQUFLLGVBQVMsRUFBQztBQUFmLE9BQ0ksS0FBS1AsS0FBTCxDQUFXRyxRQUFYLElBQXVCLEtBQUtELGNBQUwsRUFEM0IsQ0FERixDQURNLENBQVI7QUFPRDs7QUEvQjhCOztnQkFBM0JMLFEsaUJBQ2lCLFU7O2dCQURqQkEsUSxlQUVlO0FBQ2pCTSxVQUFRLEVBQUVLLGlEQUFTLENBQUNDLEtBQVYsQ0FBZ0JDLFVBRFQ7QUFFakJDLFVBQVEsRUFBRUgsaURBQVMsQ0FBQ0ksSUFBVixDQUFlRjtBQUZSLEM7O0FBZ0NOYix1RUFBZiIsImZpbGUiOiIuL3Jlc291cmNlcy9qcy9jb21tb24vYXJ0aWNsZXMvbGlzdGluZy9jb21wb25lbnRzL0FydGljbGVzLmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiaW1wb3J0IFJlYWN0LCB7Q29tcG9uZW50fSBmcm9tICdyZWFjdCdcbmltcG9ydCBQcm9wVHlwZXMgZnJvbSAncHJvcC10eXBlcydcbmltcG9ydCBBcnRpY2xlIGZyb20gJy4vQXJ0aWNsZSdcblxuY2xhc3MgQXJ0aWNsZXMgZXh0ZW5kcyBDb21wb25lbnQge1xuICBzdGF0aWMgZGlzcGxheU5hbWUgPSAnQXJ0aWNsZXMnXG4gIHN0YXRpYyBwcm9wVHlwZXMgPSB7XG4gICAgYXJ0aWNsZXM6IFByb3BUeXBlcy5hcnJheS5pc1JlcXVpcmVkLFxuICAgIGRpc3BhdGNoOiBQcm9wVHlwZXMuZnVuYy5pc1JlcXVpcmVkLFxuICB9XG4gIFxuICBjb25zdHJ1Y3Rvcihwcm9wcykge1xuICAgIHN1cGVyKHByb3BzKVxuICAgIFxuICAgIHRoaXMuc3RhdGUgPSB7XG4gICAgICAvL1xuICAgIH1cbiAgfVxuICBcbiAgcmVuZGVyQXJ0aWNsZXMoKSB7XG4gICAgcmV0dXJuIHRoaXMucHJvcHMuYXJ0aWNsZXMubWFwKChhcnRpY2xlLCBpbmRleCkgPT4ge1xuICAgICAgcmV0dXJuIDxBcnRpY2xlIGtleT17YGFydGljbGUtJHtpbmRleH1gfVxuICAgICAgICAgICAgICAgICAgICAgIGluZGV4PXtpbmRleH1cbiAgICAgICAgICAgICAgICAgICAgICBhcnRpY2xlPXthcnRpY2xlfS8+XG4gICAgfSlcbiAgfVxuICBcbiAgcmVuZGVyKCkge1xuICAgIHJldHVybiAoPHNlY3Rpb24gaWQ9XCJjb21wb25lbnRzLWFydGljbGVzXCI+XG4gICAgICA8ZGl2IGNsYXNzTmFtZT1cImNvbnRhaW5lclwiPlxuICAgICAgICA8ZGl2IGNsYXNzTmFtZT1cInJvd1wiPlxuICAgICAgICAgIHsgdGhpcy5wcm9wcy5hcnRpY2xlcyAmJiB0aGlzLnJlbmRlckFydGljbGVzKCkgfVxuICAgICAgICA8L2Rpdj5cbiAgICAgIDwvZGl2PlxuICAgIDwvc2VjdGlvbj4pXG4gIH1cbn1cblxuZXhwb3J0IGRlZmF1bHQgQXJ0aWNsZXNcbiJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/common/articles/listing/components/Articles.js\n");
 
 /***/ }),
 
@@ -32,7 +32,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var reac
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react_redux__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n/* harmony import */ var _modules_article_Article__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../../modules/article/Article */ \"./resources/js/modules/article/Article.js\");\n/* harmony import */ var _components_Articles__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./components/Articles */ \"./resources/js/common/articles/listing/components/Articles.js\");\nfunction _objectWithoutProperties(source, excluded) { if (source == null) return {}; var target = _objectWithoutPropertiesLoose(source, excluded); var key, i; if (Object.getOwnPropertySymbols) { var sourceSymbolKeys = Object.getOwnPropertySymbols(source); for (i = 0; i < sourceSymbolKeys.length; i++) { key = sourceSymbolKeys[i]; if (excluded.indexOf(key) >= 0) continue; if (!Object.prototype.propertyIsEnumerable.call(source, key)) continue; target[key] = source[key]; } } return target; }\n\nfunction _objectWithoutPropertiesLoose(source, excluded) { if (source == null) return {}; var target = {}; var sourceKeys = Object.keys(source); var key, i; for (i = 0; i < sourceKeys.length; i++) { key = sourceKeys[i]; if (excluded.indexOf(key) >= 0) continue; target[key] = source[key]; } return target; }\n\n// libs\n\n // components\n\n\n\nvar mapStateToProps = function mapStateToProps(state) {\n  var _state$articles = state.articles,\n      data = _state$articles.data,\n      meta = _objectWithoutProperties(_state$articles, [\"data\"]);\n\n  return {\n    articles: data.map(function (article) {\n      return new _modules_article_Article__WEBPACK_IMPORTED_MODULE_1__[\"default\"](article);\n    }),\n    meta: Object.assign({}, meta)\n  };\n};\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Object(react_redux__WEBPACK_IMPORTED_MODULE_0__[\"connect\"])(mapStateToProps)(_components_Articles__WEBPACK_IMPORTED_MODULE_2__[\"default\"]));//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvY29tbW9uL2FydGljbGVzL2xpc3RpbmcvaW5kZXguanM/MmJmYSJdLCJuYW1lcyI6WyJtYXBTdGF0ZVRvUHJvcHMiLCJzdGF0ZSIsImFydGljbGVzIiwiZGF0YSIsIm1ldGEiLCJtYXAiLCJhcnRpY2xlIiwiQXJ0aWNsZSIsIk9iamVjdCIsImFzc2lnbiIsImNvbm5lY3QiLCJBcnRpY2xlcyJdLCJtYXBwaW5ncyI6Ijs7Ozs7Ozs7QUFBQTtBQUNBO0NBR0E7O0FBQ0E7O0FBRUEsSUFBTUEsZUFBZSxHQUFHLFNBQWxCQSxlQUFrQixDQUFBQyxLQUFLLEVBQUk7QUFBQSx3QkFDUEEsS0FBSyxDQUFDQyxRQURDO0FBQUEsTUFDeEJDLElBRHdCLG1CQUN4QkEsSUFEd0I7QUFBQSxNQUNmQyxJQURlOztBQUcvQixTQUFPO0FBQ0xGLFlBQVEsRUFBRUMsSUFBSSxDQUFDRSxHQUFMLENBQVMsVUFBQ0MsT0FBRDtBQUFBLGFBQWEsSUFBSUMsZ0VBQUosQ0FBWUQsT0FBWixDQUFiO0FBQUEsS0FBVCxDQURMO0FBRUxGLFFBQUksRUFBRUksTUFBTSxDQUFDQyxNQUFQLENBQWMsRUFBZCxFQUFrQkwsSUFBbEI7QUFGRCxHQUFQO0FBSUQsQ0FQRDs7QUFTZU0sMEhBQU8sQ0FBQ1YsZUFBRCxDQUFQLENBQXlCVyw0REFBekIsQ0FBZiIsImZpbGUiOiIuL3Jlc291cmNlcy9qcy9jb21tb24vYXJ0aWNsZXMvbGlzdGluZy9pbmRleC5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbIi8vIGxpYnNcbmltcG9ydCB7IGNvbm5lY3QgfSBmcm9tICdyZWFjdC1yZWR1eCdcbmltcG9ydCBBcnRpY2xlIGZyb20gJy4uLy4uLy4uL21vZHVsZXMvYXJ0aWNsZS9BcnRpY2xlJ1xuXG4vLyBjb21wb25lbnRzXG5pbXBvcnQgQXJ0aWNsZXMgZnJvbSAnLi9jb21wb25lbnRzL0FydGljbGVzJ1xuXG5jb25zdCBtYXBTdGF0ZVRvUHJvcHMgPSBzdGF0ZSA9PiB7XG4gIGNvbnN0IHtkYXRhLCAuLi5tZXRhfSA9IHN0YXRlLmFydGljbGVzXG4gIFxuICByZXR1cm4ge1xuICAgIGFydGljbGVzOiBkYXRhLm1hcCgoYXJ0aWNsZSkgPT4gbmV3IEFydGljbGUoYXJ0aWNsZSkpLFxuICAgIG1ldGE6IE9iamVjdC5hc3NpZ24oe30sIG1ldGEpXG4gIH1cbn1cblxuZXhwb3J0IGRlZmF1bHQgY29ubmVjdChtYXBTdGF0ZVRvUHJvcHMpKEFydGljbGVzKVxuIl0sInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./resources/js/common/articles/listing/index.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react_redux__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n/* harmony import */ var _modules_article_Article__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../../modules/article/Article */ \"./resources/js/modules/article/Article.js\");\n/* harmony import */ var _components_Articles__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./components/Articles */ \"./resources/js/common/articles/listing/components/Articles.js\");\nfunction _objectWithoutProperties(source, excluded) { if (source == null) return {}; var target = _objectWithoutPropertiesLoose(source, excluded); var key, i; if (Object.getOwnPropertySymbols) { var sourceSymbolKeys = Object.getOwnPropertySymbols(source); for (i = 0; i < sourceSymbolKeys.length; i++) { key = sourceSymbolKeys[i]; if (excluded.indexOf(key) >= 0) continue; if (!Object.prototype.propertyIsEnumerable.call(source, key)) continue; target[key] = source[key]; } } return target; }\n\nfunction _objectWithoutPropertiesLoose(source, excluded) { if (source == null) return {}; var target = {}; var sourceKeys = Object.keys(source); var key, i; for (i = 0; i < sourceKeys.length; i++) { key = sourceKeys[i]; if (excluded.indexOf(key) >= 0) continue; target[key] = source[key]; } return target; }\n\n// libs\n\n // components\n\n\n\nconst mapStateToProps = state => {\n  const _state$articles = state.articles,\n        data = _state$articles.data,\n        meta = _objectWithoutProperties(_state$articles, [\"data\"]);\n\n  return {\n    articles: data.map(article => new _modules_article_Article__WEBPACK_IMPORTED_MODULE_1__[\"default\"](article)),\n    meta: Object.assign({}, meta)\n  };\n};\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Object(react_redux__WEBPACK_IMPORTED_MODULE_0__[\"connect\"])(mapStateToProps)(_components_Articles__WEBPACK_IMPORTED_MODULE_2__[\"default\"]));//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvY29tbW9uL2FydGljbGVzL2xpc3RpbmcvaW5kZXguanM/MmJmYSJdLCJuYW1lcyI6WyJtYXBTdGF0ZVRvUHJvcHMiLCJzdGF0ZSIsImFydGljbGVzIiwiZGF0YSIsIm1ldGEiLCJtYXAiLCJhcnRpY2xlIiwiQXJ0aWNsZSIsIk9iamVjdCIsImFzc2lnbiIsImNvbm5lY3QiLCJBcnRpY2xlcyJdLCJtYXBwaW5ncyI6Ijs7Ozs7Ozs7QUFBQTtBQUNBO0NBR0E7O0FBQ0E7O0FBRUEsTUFBTUEsZUFBZSxHQUFHQyxLQUFLLElBQUk7QUFBQSwwQkFDUEEsS0FBSyxDQUFDQyxRQURDO0FBQUEsUUFDeEJDLElBRHdCLG1CQUN4QkEsSUFEd0I7QUFBQSxRQUNmQyxJQURlOztBQUcvQixTQUFPO0FBQ0xGLFlBQVEsRUFBRUMsSUFBSSxDQUFDRSxHQUFMLENBQVVDLE9BQUQsSUFBYSxJQUFJQyxnRUFBSixDQUFZRCxPQUFaLENBQXRCLENBREw7QUFFTEYsUUFBSSxFQUFFSSxNQUFNLENBQUNDLE1BQVAsQ0FBYyxFQUFkLEVBQWtCTCxJQUFsQjtBQUZELEdBQVA7QUFJRCxDQVBEOztBQVNlTSwwSEFBTyxDQUFDVixlQUFELENBQVAsQ0FBeUJXLDREQUF6QixDQUFmIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL2NvbW1vbi9hcnRpY2xlcy9saXN0aW5nL2luZGV4LmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiLy8gbGlic1xuaW1wb3J0IHsgY29ubmVjdCB9IGZyb20gJ3JlYWN0LXJlZHV4J1xuaW1wb3J0IEFydGljbGUgZnJvbSAnLi4vLi4vLi4vbW9kdWxlcy9hcnRpY2xlL0FydGljbGUnXG5cbi8vIGNvbXBvbmVudHNcbmltcG9ydCBBcnRpY2xlcyBmcm9tICcuL2NvbXBvbmVudHMvQXJ0aWNsZXMnXG5cbmNvbnN0IG1hcFN0YXRlVG9Qcm9wcyA9IHN0YXRlID0+IHtcbiAgY29uc3Qge2RhdGEsIC4uLm1ldGF9ID0gc3RhdGUuYXJ0aWNsZXNcbiAgXG4gIHJldHVybiB7XG4gICAgYXJ0aWNsZXM6IGRhdGEubWFwKChhcnRpY2xlKSA9PiBuZXcgQXJ0aWNsZShhcnRpY2xlKSksXG4gICAgbWV0YTogT2JqZWN0LmFzc2lnbih7fSwgbWV0YSlcbiAgfVxufVxuXG5leHBvcnQgZGVmYXVsdCBjb25uZWN0KG1hcFN0YXRlVG9Qcm9wcykoQXJ0aWNsZXMpXG4iXSwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./resources/js/common/articles/listing/index.js\n");
 
 /***/ }),
 
@@ -44,7 +44,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var reac
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var moment__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! moment */ \"./node_modules/moment/moment.js\");\n/* harmony import */ var moment__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(moment__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _utils_Model__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/Model */ \"./resources/js/utils/Model.js\");\n/* harmony import */ var _user_User__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../user/User */ \"./resources/js/modules/user/User.js\");\nfunction _typeof(obj) { if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _get(target, property, receiver) { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get; } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(receiver); } return desc.value; }; } return _get(target, property, receiver || target); }\n\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\n\n\n\n\nvar Article =\n/*#__PURE__*/\nfunction (_Model) {\n  _inherits(Article, _Model);\n\n  function Article(props) {\n    var _this;\n\n    _classCallCheck(this, Article);\n\n    _this = _possibleConstructorReturn(this, _getPrototypeOf(Article).call(this, props));\n\n    _this.initialize(props);\n\n    return _this;\n  }\n\n  _createClass(Article, [{\n    key: \"initialize\",\n    value: function initialize(props) {\n      _get(_getPrototypeOf(Article.prototype), \"initialize\", this).call(this, props);\n\n      this.slug = props.slug || '';\n      this.title = props.title || '';\n      this.description = props.description || '';\n      this.content = props.content || '';\n      this.published = props.published || false;\n      this.publishedAt = props.publishedAt ? moment__WEBPACK_IMPORTED_MODULE_0___default()(props.publishedAt) : null; // relate user model\n\n      this.user = props.user ? new _user_User__WEBPACK_IMPORTED_MODULE_2__[\"default\"](props.user) : null;\n    }\n  }]);\n\n  return Article;\n}(_utils_Model__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Article);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL0FydGljbGUuanM/YjE5MSJdLCJuYW1lcyI6WyJBcnRpY2xlIiwicHJvcHMiLCJpbml0aWFsaXplIiwic2x1ZyIsInRpdGxlIiwiZGVzY3JpcHRpb24iLCJjb250ZW50IiwicHVibGlzaGVkIiwicHVibGlzaGVkQXQiLCJtb21lbnQiLCJ1c2VyIiwiVXNlciIsIk1vZGVsIl0sIm1hcHBpbmdzIjoiOzs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7QUFBQTtBQUNBO0FBQ0E7O0lBRU1BLE87Ozs7O0FBQ0osbUJBQVlDLEtBQVosRUFBbUI7QUFBQTs7QUFBQTs7QUFDakIsaUZBQU1BLEtBQU47O0FBRUEsVUFBS0MsVUFBTCxDQUFnQkQsS0FBaEI7O0FBSGlCO0FBSWxCOzs7OytCQUVVQSxLLEVBQU87QUFDaEIsOEVBQWlCQSxLQUFqQjs7QUFFQSxXQUFLRSxJQUFMLEdBQVlGLEtBQUssQ0FBQ0UsSUFBTixJQUFjLEVBQTFCO0FBQ0EsV0FBS0MsS0FBTCxHQUFhSCxLQUFLLENBQUNHLEtBQU4sSUFBZSxFQUE1QjtBQUNBLFdBQUtDLFdBQUwsR0FBbUJKLEtBQUssQ0FBQ0ksV0FBTixJQUFxQixFQUF4QztBQUNBLFdBQUtDLE9BQUwsR0FBZUwsS0FBSyxDQUFDSyxPQUFOLElBQWlCLEVBQWhDO0FBQ0EsV0FBS0MsU0FBTCxHQUFpQk4sS0FBSyxDQUFDTSxTQUFOLElBQW1CLEtBQXBDO0FBQ0EsV0FBS0MsV0FBTCxHQUFtQlAsS0FBSyxDQUFDTyxXQUFOLEdBQW9CQyw2Q0FBTSxDQUFDUixLQUFLLENBQUNPLFdBQVAsQ0FBMUIsR0FBZ0QsSUFBbkUsQ0FSZ0IsQ0FVaEI7O0FBQ0EsV0FBS0UsSUFBTCxHQUFZVCxLQUFLLENBQUNTLElBQU4sR0FBYSxJQUFJQyxrREFBSixDQUFTVixLQUFLLENBQUNTLElBQWYsQ0FBYixHQUFvQyxJQUFoRDtBQUNEOzs7O0VBbkJtQkUsb0Q7O0FBc0JQWixzRUFBZiIsImZpbGUiOiIuL3Jlc291cmNlcy9qcy9tb2R1bGVzL2FydGljbGUvQXJ0aWNsZS5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBtb21lbnQgZnJvbSAnbW9tZW50J1xuaW1wb3J0IE1vZGVsIGZyb20gJy4uLy4uL3V0aWxzL01vZGVsJ1xuaW1wb3J0IFVzZXIgZnJvbSAnLi4vdXNlci9Vc2VyJ1xuXG5jbGFzcyBBcnRpY2xlIGV4dGVuZHMgTW9kZWwge1xuICBjb25zdHJ1Y3Rvcihwcm9wcykge1xuICAgIHN1cGVyKHByb3BzKVxuXG4gICAgdGhpcy5pbml0aWFsaXplKHByb3BzKVxuICB9XG5cbiAgaW5pdGlhbGl6ZShwcm9wcykge1xuICAgIHN1cGVyLmluaXRpYWxpemUocHJvcHMpXG5cbiAgICB0aGlzLnNsdWcgPSBwcm9wcy5zbHVnIHx8ICcnXG4gICAgdGhpcy50aXRsZSA9IHByb3BzLnRpdGxlIHx8ICcnXG4gICAgdGhpcy5kZXNjcmlwdGlvbiA9IHByb3BzLmRlc2NyaXB0aW9uIHx8ICcnXG4gICAgdGhpcy5jb250ZW50ID0gcHJvcHMuY29udGVudCB8fCAnJ1xuICAgIHRoaXMucHVibGlzaGVkID0gcHJvcHMucHVibGlzaGVkIHx8IGZhbHNlXG4gICAgdGhpcy5wdWJsaXNoZWRBdCA9IHByb3BzLnB1Ymxpc2hlZEF0ID8gbW9tZW50KHByb3BzLnB1Ymxpc2hlZEF0KSA6IG51bGxcblxuICAgIC8vIHJlbGF0ZSB1c2VyIG1vZGVsXG4gICAgdGhpcy51c2VyID0gcHJvcHMudXNlciA/IG5ldyBVc2VyKHByb3BzLnVzZXIpIDogbnVsbFxuICB9XG59XG5cbmV4cG9ydCBkZWZhdWx0IEFydGljbGVcbiJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/article/Article.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var moment__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! moment */ \"./node_modules/moment/moment.js\");\n/* harmony import */ var moment__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(moment__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _utils_Model__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/Model */ \"./resources/js/utils/Model.js\");\n/* harmony import */ var _user_User__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../user/User */ \"./resources/js/modules/user/User.js\");\n\n\n\n\nclass Article extends _utils_Model__WEBPACK_IMPORTED_MODULE_1__[\"default\"] {\n  constructor(props) {\n    super(props);\n    this.initialize(props);\n  }\n\n  initialize(props) {\n    super.initialize(props);\n    this.slug = props.slug || '';\n    this.title = props.title || '';\n    this.description = props.description || '';\n    this.content = props.content || '';\n    this.published = props.published || false;\n    this.publishedAt = props.publishedAt ? moment__WEBPACK_IMPORTED_MODULE_0___default()(props.publishedAt) : null; // relate user model\n\n    this.user = props.user ? new _user_User__WEBPACK_IMPORTED_MODULE_2__[\"default\"](props.user) : null;\n  }\n\n}\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Article);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL0FydGljbGUuanM/YjE5MSJdLCJuYW1lcyI6WyJBcnRpY2xlIiwiTW9kZWwiLCJjb25zdHJ1Y3RvciIsInByb3BzIiwiaW5pdGlhbGl6ZSIsInNsdWciLCJ0aXRsZSIsImRlc2NyaXB0aW9uIiwiY29udGVudCIsInB1Ymxpc2hlZCIsInB1Ymxpc2hlZEF0IiwibW9tZW50IiwidXNlciIsIlVzZXIiXSwibWFwcGluZ3MiOiJBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUNBO0FBQ0E7O0FBRUEsTUFBTUEsT0FBTixTQUFzQkMsb0RBQXRCLENBQTRCO0FBQzFCQyxhQUFXLENBQUNDLEtBQUQsRUFBUTtBQUNqQixVQUFNQSxLQUFOO0FBRUEsU0FBS0MsVUFBTCxDQUFnQkQsS0FBaEI7QUFDRDs7QUFFREMsWUFBVSxDQUFDRCxLQUFELEVBQVE7QUFDaEIsVUFBTUMsVUFBTixDQUFpQkQsS0FBakI7QUFFQSxTQUFLRSxJQUFMLEdBQVlGLEtBQUssQ0FBQ0UsSUFBTixJQUFjLEVBQTFCO0FBQ0EsU0FBS0MsS0FBTCxHQUFhSCxLQUFLLENBQUNHLEtBQU4sSUFBZSxFQUE1QjtBQUNBLFNBQUtDLFdBQUwsR0FBbUJKLEtBQUssQ0FBQ0ksV0FBTixJQUFxQixFQUF4QztBQUNBLFNBQUtDLE9BQUwsR0FBZUwsS0FBSyxDQUFDSyxPQUFOLElBQWlCLEVBQWhDO0FBQ0EsU0FBS0MsU0FBTCxHQUFpQk4sS0FBSyxDQUFDTSxTQUFOLElBQW1CLEtBQXBDO0FBQ0EsU0FBS0MsV0FBTCxHQUFtQlAsS0FBSyxDQUFDTyxXQUFOLEdBQW9CQyw2Q0FBTSxDQUFDUixLQUFLLENBQUNPLFdBQVAsQ0FBMUIsR0FBZ0QsSUFBbkUsQ0FSZ0IsQ0FVaEI7O0FBQ0EsU0FBS0UsSUFBTCxHQUFZVCxLQUFLLENBQUNTLElBQU4sR0FBYSxJQUFJQyxrREFBSixDQUFTVixLQUFLLENBQUNTLElBQWYsQ0FBYixHQUFvQyxJQUFoRDtBQUNEOztBQW5CeUI7O0FBc0JiWixzRUFBZiIsImZpbGUiOiIuL3Jlc291cmNlcy9qcy9tb2R1bGVzL2FydGljbGUvQXJ0aWNsZS5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBtb21lbnQgZnJvbSAnbW9tZW50J1xuaW1wb3J0IE1vZGVsIGZyb20gJy4uLy4uL3V0aWxzL01vZGVsJ1xuaW1wb3J0IFVzZXIgZnJvbSAnLi4vdXNlci9Vc2VyJ1xuXG5jbGFzcyBBcnRpY2xlIGV4dGVuZHMgTW9kZWwge1xuICBjb25zdHJ1Y3Rvcihwcm9wcykge1xuICAgIHN1cGVyKHByb3BzKVxuXG4gICAgdGhpcy5pbml0aWFsaXplKHByb3BzKVxuICB9XG5cbiAgaW5pdGlhbGl6ZShwcm9wcykge1xuICAgIHN1cGVyLmluaXRpYWxpemUocHJvcHMpXG5cbiAgICB0aGlzLnNsdWcgPSBwcm9wcy5zbHVnIHx8ICcnXG4gICAgdGhpcy50aXRsZSA9IHByb3BzLnRpdGxlIHx8ICcnXG4gICAgdGhpcy5kZXNjcmlwdGlvbiA9IHByb3BzLmRlc2NyaXB0aW9uIHx8ICcnXG4gICAgdGhpcy5jb250ZW50ID0gcHJvcHMuY29udGVudCB8fCAnJ1xuICAgIHRoaXMucHVibGlzaGVkID0gcHJvcHMucHVibGlzaGVkIHx8IGZhbHNlXG4gICAgdGhpcy5wdWJsaXNoZWRBdCA9IHByb3BzLnB1Ymxpc2hlZEF0ID8gbW9tZW50KHByb3BzLnB1Ymxpc2hlZEF0KSA6IG51bGxcblxuICAgIC8vIHJlbGF0ZSB1c2VyIG1vZGVsXG4gICAgdGhpcy51c2VyID0gcHJvcHMudXNlciA/IG5ldyBVc2VyKHByb3BzLnVzZXIpIDogbnVsbFxuICB9XG59XG5cbmV4cG9ydCBkZWZhdWx0IEFydGljbGVcbiJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/article/Article.js\n");
 
 /***/ }),
 
@@ -56,7 +56,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var mome
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleAddRequest\", function() { return articleAddRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleUpdateRequest\", function() { return articleUpdateRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleRemoveRequest\", function() { return articleRemoveRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleListRequest\", function() { return articleListRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleEditRequest\", function() { return articleEditRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleFetchRequest\", function() { return articleFetchRequest; });\n/* harmony import */ var _utils_Http__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/Http */ \"./resources/js/utils/Http.js\");\n/* harmony import */ var _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/Transformer */ \"./resources/js/utils/Transformer.js\");\n/* harmony import */ var _store_actions__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./store/actions */ \"./resources/js/modules/article/store/actions.js\");\n\n\n\n\nfunction transformRequest(parms) {\n  return _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].send(parms);\n}\n\nfunction transformResponse(params) {\n  return _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].fetch(params);\n}\n\nfunction articleAddRequest(params) {\n  return function (dispatch) {\n    return new Promise(function (resolve, reject) {\n      _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].post('/articles', transformRequest(params)).then(function (res) {\n        dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n        return resolve();\n      }).catch(function (err) {\n        var statusCode = err.response.status;\n        var data = {\n          error: null,\n          statusCode: statusCode\n        };\n\n        if (statusCode === 422) {\n          var resetErrors = {\n            errors: err.response.data,\n            replace: false,\n            searchStr: '',\n            replaceStr: ''\n          };\n          data.error = _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].resetValidationFields(resetErrors);\n        } else if (statusCode === 401) {\n          data.error = err.response.data.message;\n        }\n\n        return reject(data);\n      });\n    });\n  };\n}\nfunction articleUpdateRequest(params) {\n  return function (dispatch) {\n    return new Promise(function (resolve, reject) {\n      _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].patch(\"articles/\".concat(params.id), transformRequest(params)).then(function (res) {\n        dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n        return resolve();\n      }).catch(function (err) {\n        var statusCode = err.response.status;\n        var data = {\n          error: null,\n          statusCode: statusCode\n        };\n\n        if (statusCode === 422) {\n          var resetErrors = {\n            errors: err.response.data,\n            replace: false,\n            searchStr: '',\n            replaceStr: ''\n          };\n          data.error = _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].resetValidationFields(resetErrors);\n        } else if (statusCode === 401) {\n          data.error = err.response.data.message;\n        }\n\n        return reject(data);\n      });\n    });\n  };\n}\nfunction articleRemoveRequest(id) {\n  return function (dispatch) {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].delete(\"articles/\".concat(id)).then(function () {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"remove\"](id));\n    }).catch(function (err) {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleListRequest(_ref) {\n  var _ref$pageNumber = _ref.pageNumber,\n      pageNumber = _ref$pageNumber === void 0 ? 1 : _ref$pageNumber,\n      _ref$url = _ref.url,\n      url = _ref$url === void 0 ? '/articles' : _ref$url;\n  return function (dispatch) {\n    if (pageNumber > 1) {\n      url = url + \"?page=\".concat(pageNumber);\n    }\n\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(url).then(function (res) {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"list\"](transformResponse(res.data)));\n    }).catch(function (err) {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleEditRequest(id) {\n  return function (dispatch) {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(\"articles/\".concat(id)).then(function (res) {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n    }).catch(function (err) {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleFetchRequest(slug) {\n  return function (dispatch) {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(\"articles/published/\".concat(slug)).then(function (res) {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n    }).catch(function (err) {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL3NlcnZpY2UuanM/Y2YwNiJdLCJuYW1lcyI6WyJ0cmFuc2Zvcm1SZXF1ZXN0IiwicGFybXMiLCJUcmFuc2Zvcm1lciIsInNlbmQiLCJ0cmFuc2Zvcm1SZXNwb25zZSIsInBhcmFtcyIsImZldGNoIiwiYXJ0aWNsZUFkZFJlcXVlc3QiLCJkaXNwYXRjaCIsIlByb21pc2UiLCJyZXNvbHZlIiwicmVqZWN0IiwiSHR0cCIsInBvc3QiLCJ0aGVuIiwicmVzIiwiYXJ0aWNsZUFjdGlvbnMiLCJkYXRhIiwiY2F0Y2giLCJlcnIiLCJzdGF0dXNDb2RlIiwicmVzcG9uc2UiLCJzdGF0dXMiLCJlcnJvciIsInJlc2V0RXJyb3JzIiwiZXJyb3JzIiwicmVwbGFjZSIsInNlYXJjaFN0ciIsInJlcGxhY2VTdHIiLCJyZXNldFZhbGlkYXRpb25GaWVsZHMiLCJtZXNzYWdlIiwiYXJ0aWNsZVVwZGF0ZVJlcXVlc3QiLCJwYXRjaCIsImlkIiwiYXJ0aWNsZVJlbW92ZVJlcXVlc3QiLCJkZWxldGUiLCJjb25zb2xlIiwiYXJ0aWNsZUxpc3RSZXF1ZXN0IiwicGFnZU51bWJlciIsInVybCIsImdldCIsImFydGljbGVFZGl0UmVxdWVzdCIsImFydGljbGVGZXRjaFJlcXVlc3QiLCJzbHVnIl0sIm1hcHBpbmdzIjoiQUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQ0E7QUFDQTs7QUFFQSxTQUFTQSxnQkFBVCxDQUEwQkMsS0FBMUIsRUFBaUM7QUFDL0IsU0FBT0MsMERBQVcsQ0FBQ0MsSUFBWixDQUFpQkYsS0FBakIsQ0FBUDtBQUNEOztBQUVELFNBQVNHLGlCQUFULENBQTJCQyxNQUEzQixFQUFtQztBQUNqQyxTQUFPSCwwREFBVyxDQUFDSSxLQUFaLENBQWtCRCxNQUFsQixDQUFQO0FBQ0Q7O0FBRU0sU0FBU0UsaUJBQVQsQ0FBMkJGLE1BQTNCLEVBQW1DO0FBQ3hDLFNBQU8sVUFBQUcsUUFBUTtBQUFBLFdBQ2IsSUFBSUMsT0FBSixDQUFZLFVBQUNDLE9BQUQsRUFBVUMsTUFBVixFQUFxQjtBQUMvQkMseURBQUksQ0FBQ0MsSUFBTCxDQUFVLFdBQVYsRUFBdUJiLGdCQUFnQixDQUFDSyxNQUFELENBQXZDLEVBQ0dTLElBREgsQ0FDUSxVQUFBQyxHQUFHLEVBQUk7QUFDWFAsZ0JBQVEsQ0FBQ1Esa0RBQUEsQ0FBbUJaLGlCQUFpQixDQUFDVyxHQUFHLENBQUNFLElBQUwsQ0FBcEMsQ0FBRCxDQUFSO0FBQ0EsZUFBT1AsT0FBTyxFQUFkO0FBQ0QsT0FKSCxFQUtHUSxLQUxILENBS1MsVUFBQ0MsR0FBRCxFQUFTO0FBQ2QsWUFBTUMsVUFBVSxHQUFHRCxHQUFHLENBQUNFLFFBQUosQ0FBYUMsTUFBaEM7QUFDQSxZQUFNTCxJQUFJLEdBQUc7QUFDWE0sZUFBSyxFQUFFLElBREk7QUFFWEgsb0JBQVUsRUFBVkE7QUFGVyxTQUFiOztBQUtBLFlBQUlBLFVBQVUsS0FBSyxHQUFuQixFQUF3QjtBQUN0QixjQUFNSSxXQUFXLEdBQUc7QUFDbEJDLGtCQUFNLEVBQUVOLEdBQUcsQ0FBQ0UsUUFBSixDQUFhSixJQURIO0FBRWxCUyxtQkFBTyxFQUFFLEtBRlM7QUFHbEJDLHFCQUFTLEVBQUUsRUFITztBQUlsQkMsc0JBQVUsRUFBRTtBQUpNLFdBQXBCO0FBTUFYLGNBQUksQ0FBQ00sS0FBTCxHQUFhckIsMERBQVcsQ0FBQzJCLHFCQUFaLENBQWtDTCxXQUFsQyxDQUFiO0FBQ0QsU0FSRCxNQVFPLElBQUlKLFVBQVUsS0FBSyxHQUFuQixFQUF3QjtBQUM3QkgsY0FBSSxDQUFDTSxLQUFMLEdBQWFKLEdBQUcsQ0FBQ0UsUUFBSixDQUFhSixJQUFiLENBQWtCYSxPQUEvQjtBQUNEOztBQUNELGVBQU9uQixNQUFNLENBQUNNLElBQUQsQ0FBYjtBQUNELE9BeEJIO0FBeUJELEtBMUJELENBRGE7QUFBQSxHQUFmO0FBNkJEO0FBRU0sU0FBU2Msb0JBQVQsQ0FBOEIxQixNQUE5QixFQUFzQztBQUMzQyxTQUFPLFVBQUFHLFFBQVE7QUFBQSxXQUNiLElBQUlDLE9BQUosQ0FBWSxVQUFDQyxPQUFELEVBQVVDLE1BQVYsRUFBcUI7QUFDL0JDLHlEQUFJLENBQUNvQixLQUFMLG9CQUF1QjNCLE1BQU0sQ0FBQzRCLEVBQTlCLEdBQW9DakMsZ0JBQWdCLENBQUNLLE1BQUQsQ0FBcEQsRUFDR1MsSUFESCxDQUNRLFVBQUFDLEdBQUcsRUFBSTtBQUNYUCxnQkFBUSxDQUFDUSxrREFBQSxDQUFtQlosaUJBQWlCLENBQUNXLEdBQUcsQ0FBQ0UsSUFBTCxDQUFwQyxDQUFELENBQVI7QUFDQSxlQUFPUCxPQUFPLEVBQWQ7QUFDRCxPQUpILEVBS0dRLEtBTEgsQ0FLUyxVQUFDQyxHQUFELEVBQVM7QUFDZCxZQUFNQyxVQUFVLEdBQUdELEdBQUcsQ0FBQ0UsUUFBSixDQUFhQyxNQUFoQztBQUNBLFlBQU1MLElBQUksR0FBRztBQUNYTSxlQUFLLEVBQUUsSUFESTtBQUVYSCxvQkFBVSxFQUFWQTtBQUZXLFNBQWI7O0FBS0EsWUFBSUEsVUFBVSxLQUFLLEdBQW5CLEVBQXdCO0FBQ3RCLGNBQU1JLFdBQVcsR0FBRztBQUNsQkMsa0JBQU0sRUFBRU4sR0FBRyxDQUFDRSxRQUFKLENBQWFKLElBREg7QUFFbEJTLG1CQUFPLEVBQUUsS0FGUztBQUdsQkMscUJBQVMsRUFBRSxFQUhPO0FBSWxCQyxzQkFBVSxFQUFFO0FBSk0sV0FBcEI7QUFNQVgsY0FBSSxDQUFDTSxLQUFMLEdBQWFyQiwwREFBVyxDQUFDMkIscUJBQVosQ0FBa0NMLFdBQWxDLENBQWI7QUFDRCxTQVJELE1BUU8sSUFBSUosVUFBVSxLQUFLLEdBQW5CLEVBQXdCO0FBQzdCSCxjQUFJLENBQUNNLEtBQUwsR0FBYUosR0FBRyxDQUFDRSxRQUFKLENBQWFKLElBQWIsQ0FBa0JhLE9BQS9CO0FBQ0Q7O0FBQ0QsZUFBT25CLE1BQU0sQ0FBQ00sSUFBRCxDQUFiO0FBQ0QsT0F4Qkg7QUF5QkQsS0ExQkQsQ0FEYTtBQUFBLEdBQWY7QUE2QkQ7QUFFTSxTQUFTaUIsb0JBQVQsQ0FBOEJELEVBQTlCLEVBQWtDO0FBQ3ZDLFNBQU8sVUFBQXpCLFFBQVEsRUFBSTtBQUNqQkksdURBQUksQ0FBQ3VCLE1BQUwsb0JBQXdCRixFQUF4QixHQUNHbkIsSUFESCxDQUNRLFlBQU07QUFDVk4sY0FBUSxDQUFDUSxxREFBQSxDQUFzQmlCLEVBQXRCLENBQUQsQ0FBUjtBQUNELEtBSEgsRUFJR2YsS0FKSCxDQUlTLFVBQUNDLEdBQUQsRUFBUztBQUNkO0FBQ0FpQixhQUFPLENBQUNiLEtBQVIsQ0FBY0osR0FBRyxDQUFDRSxRQUFsQjtBQUNELEtBUEg7QUFRRCxHQVREO0FBVUQ7QUFFTSxTQUFTZ0Isa0JBQVQsT0FBaUU7QUFBQSw2QkFBcENDLFVBQW9DO0FBQUEsTUFBcENBLFVBQW9DLGdDQUF2QixDQUF1QjtBQUFBLHNCQUFwQkMsR0FBb0I7QUFBQSxNQUFwQkEsR0FBb0IseUJBQWQsV0FBYztBQUN0RSxTQUFPLFVBQUEvQixRQUFRLEVBQUk7QUFDakIsUUFBSThCLFVBQVUsR0FBRyxDQUFqQixFQUFvQjtBQUNsQkMsU0FBRyxHQUFHQSxHQUFHLG1CQUFZRCxVQUFaLENBQVQ7QUFDRDs7QUFFRDFCLHVEQUFJLENBQUM0QixHQUFMLENBQVNELEdBQVQsRUFDR3pCLElBREgsQ0FDUSxVQUFDQyxHQUFELEVBQVM7QUFDYlAsY0FBUSxDQUFDUSxtREFBQSxDQUFvQlosaUJBQWlCLENBQUNXLEdBQUcsQ0FBQ0UsSUFBTCxDQUFyQyxDQUFELENBQVI7QUFDRCxLQUhILEVBSUdDLEtBSkgsQ0FJUyxVQUFDQyxHQUFELEVBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FiRDtBQWNEO0FBRU0sU0FBU29CLGtCQUFULENBQTRCUixFQUE1QixFQUFnQztBQUNyQyxTQUFPLFVBQUF6QixRQUFRLEVBQUk7QUFDakJJLHVEQUFJLENBQUM0QixHQUFMLG9CQUFxQlAsRUFBckIsR0FDR25CLElBREgsQ0FDUSxVQUFDQyxHQUFELEVBQVM7QUFDYlAsY0FBUSxDQUFDUSxrREFBQSxDQUFtQlosaUJBQWlCLENBQUNXLEdBQUcsQ0FBQ0UsSUFBTCxDQUFwQyxDQUFELENBQVI7QUFDRCxLQUhILEVBSUdDLEtBSkgsQ0FJUyxVQUFDQyxHQUFELEVBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FURDtBQVVEO0FBRU0sU0FBU3FCLG1CQUFULENBQTZCQyxJQUE3QixFQUFtQztBQUN4QyxTQUFPLFVBQUFuQyxRQUFRLEVBQUk7QUFDakJJLHVEQUFJLENBQUM0QixHQUFMLDhCQUErQkcsSUFBL0IsR0FDRzdCLElBREgsQ0FDUSxVQUFDQyxHQUFELEVBQVM7QUFDYlAsY0FBUSxDQUFDUSxrREFBQSxDQUFtQlosaUJBQWlCLENBQUNXLEdBQUcsQ0FBQ0UsSUFBTCxDQUFwQyxDQUFELENBQVI7QUFDRCxLQUhILEVBSUdDLEtBSkgsQ0FJUyxVQUFDQyxHQUFELEVBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FURDtBQVVEIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvYXJ0aWNsZS9zZXJ2aWNlLmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiaW1wb3J0IEh0dHAgZnJvbSAnLi4vLi4vdXRpbHMvSHR0cCdcbmltcG9ydCBUcmFuc2Zvcm1lciBmcm9tICcuLi8uLi91dGlscy9UcmFuc2Zvcm1lcidcbmltcG9ydCAqIGFzIGFydGljbGVBY3Rpb25zIGZyb20gJy4vc3RvcmUvYWN0aW9ucydcblxuZnVuY3Rpb24gdHJhbnNmb3JtUmVxdWVzdChwYXJtcykge1xuICByZXR1cm4gVHJhbnNmb3JtZXIuc2VuZChwYXJtcylcbn1cblxuZnVuY3Rpb24gdHJhbnNmb3JtUmVzcG9uc2UocGFyYW1zKSB7XG4gIHJldHVybiBUcmFuc2Zvcm1lci5mZXRjaChwYXJhbXMpXG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlQWRkUmVxdWVzdChwYXJhbXMpIHtcbiAgcmV0dXJuIGRpc3BhdGNoID0+IChcbiAgICBuZXcgUHJvbWlzZSgocmVzb2x2ZSwgcmVqZWN0KSA9PiB7XG4gICAgICBIdHRwLnBvc3QoJy9hcnRpY2xlcycsIHRyYW5zZm9ybVJlcXVlc3QocGFyYW1zKSlcbiAgICAgICAgLnRoZW4ocmVzID0+IHtcbiAgICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5hZGQodHJhbnNmb3JtUmVzcG9uc2UocmVzLmRhdGEpKSlcbiAgICAgICAgICByZXR1cm4gcmVzb2x2ZSgpXG4gICAgICAgIH0pXG4gICAgICAgIC5jYXRjaCgoZXJyKSA9PiB7XG4gICAgICAgICAgY29uc3Qgc3RhdHVzQ29kZSA9IGVyci5yZXNwb25zZS5zdGF0dXM7XG4gICAgICAgICAgY29uc3QgZGF0YSA9IHtcbiAgICAgICAgICAgIGVycm9yOiBudWxsLFxuICAgICAgICAgICAgc3RhdHVzQ29kZSxcbiAgICAgICAgICB9O1xuXG4gICAgICAgICAgaWYgKHN0YXR1c0NvZGUgPT09IDQyMikge1xuICAgICAgICAgICAgY29uc3QgcmVzZXRFcnJvcnMgPSB7XG4gICAgICAgICAgICAgIGVycm9yczogZXJyLnJlc3BvbnNlLmRhdGEsXG4gICAgICAgICAgICAgIHJlcGxhY2U6IGZhbHNlLFxuICAgICAgICAgICAgICBzZWFyY2hTdHI6ICcnLFxuICAgICAgICAgICAgICByZXBsYWNlU3RyOiAnJyxcbiAgICAgICAgICAgIH07XG4gICAgICAgICAgICBkYXRhLmVycm9yID0gVHJhbnNmb3JtZXIucmVzZXRWYWxpZGF0aW9uRmllbGRzKHJlc2V0RXJyb3JzKTtcbiAgICAgICAgICB9IGVsc2UgaWYgKHN0YXR1c0NvZGUgPT09IDQwMSkge1xuICAgICAgICAgICAgZGF0YS5lcnJvciA9IGVyci5yZXNwb25zZS5kYXRhLm1lc3NhZ2U7XG4gICAgICAgICAgfVxuICAgICAgICAgIHJldHVybiByZWplY3QoZGF0YSk7XG4gICAgICAgIH0pXG4gICAgfSlcbiAgKVxufVxuXG5leHBvcnQgZnVuY3Rpb24gYXJ0aWNsZVVwZGF0ZVJlcXVlc3QocGFyYW1zKSB7XG4gIHJldHVybiBkaXNwYXRjaCA9PiAoXG4gICAgbmV3IFByb21pc2UoKHJlc29sdmUsIHJlamVjdCkgPT4ge1xuICAgICAgSHR0cC5wYXRjaChgYXJ0aWNsZXMvJHtwYXJhbXMuaWR9YCwgdHJhbnNmb3JtUmVxdWVzdChwYXJhbXMpKVxuICAgICAgICAudGhlbihyZXMgPT4ge1xuICAgICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLmFkZCh0cmFuc2Zvcm1SZXNwb25zZShyZXMuZGF0YSkpKVxuICAgICAgICAgIHJldHVybiByZXNvbHZlKClcbiAgICAgICAgfSlcbiAgICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgICBjb25zdCBzdGF0dXNDb2RlID0gZXJyLnJlc3BvbnNlLnN0YXR1cztcbiAgICAgICAgICBjb25zdCBkYXRhID0ge1xuICAgICAgICAgICAgZXJyb3I6IG51bGwsXG4gICAgICAgICAgICBzdGF0dXNDb2RlLFxuICAgICAgICAgIH07XG5cbiAgICAgICAgICBpZiAoc3RhdHVzQ29kZSA9PT0gNDIyKSB7XG4gICAgICAgICAgICBjb25zdCByZXNldEVycm9ycyA9IHtcbiAgICAgICAgICAgICAgZXJyb3JzOiBlcnIucmVzcG9uc2UuZGF0YSxcbiAgICAgICAgICAgICAgcmVwbGFjZTogZmFsc2UsXG4gICAgICAgICAgICAgIHNlYXJjaFN0cjogJycsXG4gICAgICAgICAgICAgIHJlcGxhY2VTdHI6ICcnLFxuICAgICAgICAgICAgfTtcbiAgICAgICAgICAgIGRhdGEuZXJyb3IgPSBUcmFuc2Zvcm1lci5yZXNldFZhbGlkYXRpb25GaWVsZHMocmVzZXRFcnJvcnMpO1xuICAgICAgICAgIH0gZWxzZSBpZiAoc3RhdHVzQ29kZSA9PT0gNDAxKSB7XG4gICAgICAgICAgICBkYXRhLmVycm9yID0gZXJyLnJlc3BvbnNlLmRhdGEubWVzc2FnZTtcbiAgICAgICAgICB9XG4gICAgICAgICAgcmV0dXJuIHJlamVjdChkYXRhKTtcbiAgICAgICAgfSlcbiAgICB9KVxuICApXG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlUmVtb3ZlUmVxdWVzdChpZCkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIEh0dHAuZGVsZXRlKGBhcnRpY2xlcy8ke2lkfWApXG4gICAgICAudGhlbigoKSA9PiB7XG4gICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLnJlbW92ZShpZCkpXG4gICAgICB9KVxuICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgLy8gVE9ETzogaGFuZGxlIGVyclxuICAgICAgICBjb25zb2xlLmVycm9yKGVyci5yZXNwb25zZSlcbiAgICAgIH0pXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIGFydGljbGVMaXN0UmVxdWVzdCh7cGFnZU51bWJlciA9IDEsIHVybCA9ICcvYXJ0aWNsZXMnfSkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIGlmIChwYWdlTnVtYmVyID4gMSkge1xuICAgICAgdXJsID0gdXJsICsgYD9wYWdlPSR7cGFnZU51bWJlcn1gXG4gICAgfVxuXG4gICAgSHR0cC5nZXQodXJsKVxuICAgICAgLnRoZW4oKHJlcykgPT4ge1xuICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5saXN0KHRyYW5zZm9ybVJlc3BvbnNlKHJlcy5kYXRhKSkpXG4gICAgICB9KVxuICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgLy8gVE9ETzogaGFuZGxlIGVyclxuICAgICAgICBjb25zb2xlLmVycm9yKGVyci5yZXNwb25zZSlcbiAgICAgIH0pXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIGFydGljbGVFZGl0UmVxdWVzdChpZCkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIEh0dHAuZ2V0KGBhcnRpY2xlcy8ke2lkfWApXG4gICAgICAudGhlbigocmVzKSA9PiB7XG4gICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLmFkZCh0cmFuc2Zvcm1SZXNwb25zZShyZXMuZGF0YSkpKVxuICAgICAgfSlcbiAgICAgIC5jYXRjaCgoZXJyKSA9PiB7XG4gICAgICAgIC8vIFRPRE86IGhhbmRsZSBlcnJcbiAgICAgICAgY29uc29sZS5lcnJvcihlcnIucmVzcG9uc2UpXG4gICAgICB9KVxuICB9XG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlRmV0Y2hSZXF1ZXN0KHNsdWcpIHtcbiAgcmV0dXJuIGRpc3BhdGNoID0+IHtcbiAgICBIdHRwLmdldChgYXJ0aWNsZXMvcHVibGlzaGVkLyR7c2x1Z31gKVxuICAgICAgLnRoZW4oKHJlcykgPT4ge1xuICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5hZGQodHJhbnNmb3JtUmVzcG9uc2UocmVzLmRhdGEpKSlcbiAgICAgIH0pXG4gICAgICAuY2F0Y2goKGVycikgPT4ge1xuICAgICAgICAvLyBUT0RPOiBoYW5kbGUgZXJyXG4gICAgICAgIGNvbnNvbGUuZXJyb3IoZXJyLnJlc3BvbnNlKVxuICAgICAgfSlcbiAgfVxufSJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/article/service.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleAddRequest\", function() { return articleAddRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleUpdateRequest\", function() { return articleUpdateRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleRemoveRequest\", function() { return articleRemoveRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleListRequest\", function() { return articleListRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleEditRequest\", function() { return articleEditRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleFetchRequest\", function() { return articleFetchRequest; });\n/* harmony import */ var _utils_Http__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/Http */ \"./resources/js/utils/Http.js\");\n/* harmony import */ var _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/Transformer */ \"./resources/js/utils/Transformer.js\");\n/* harmony import */ var _store_actions__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./store/actions */ \"./resources/js/modules/article/store/actions.js\");\n\n\n\n\nfunction transformRequest(parms) {\n  return _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].send(parms);\n}\n\nfunction transformResponse(params) {\n  return _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].fetch(params);\n}\n\nfunction articleAddRequest(params) {\n  return dispatch => new Promise((resolve, reject) => {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].post('/articles', transformRequest(params)).then(res => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n      return resolve();\n    }).catch(err => {\n      const statusCode = err.response.status;\n      const data = {\n        error: null,\n        statusCode\n      };\n\n      if (statusCode === 422) {\n        const resetErrors = {\n          errors: err.response.data,\n          replace: false,\n          searchStr: '',\n          replaceStr: ''\n        };\n        data.error = _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].resetValidationFields(resetErrors);\n      } else if (statusCode === 401) {\n        data.error = err.response.data.message;\n      }\n\n      return reject(data);\n    });\n  });\n}\nfunction articleUpdateRequest(params) {\n  return dispatch => new Promise((resolve, reject) => {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].patch(\"articles/\".concat(params.id), transformRequest(params)).then(res => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n      return resolve();\n    }).catch(err => {\n      const statusCode = err.response.status;\n      const data = {\n        error: null,\n        statusCode\n      };\n\n      if (statusCode === 422) {\n        const resetErrors = {\n          errors: err.response.data,\n          replace: false,\n          searchStr: '',\n          replaceStr: ''\n        };\n        data.error = _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].resetValidationFields(resetErrors);\n      } else if (statusCode === 401) {\n        data.error = err.response.data.message;\n      }\n\n      return reject(data);\n    });\n  });\n}\nfunction articleRemoveRequest(id) {\n  return dispatch => {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].delete(\"articles/\".concat(id)).then(() => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"remove\"](id));\n    }).catch(err => {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleListRequest({\n  pageNumber = 1,\n  url = '/articles'\n}) {\n  return dispatch => {\n    if (pageNumber > 1) {\n      url = url + \"?page=\".concat(pageNumber);\n    }\n\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(url).then(res => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"list\"](transformResponse(res.data)));\n    }).catch(err => {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleEditRequest(id) {\n  return dispatch => {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(\"articles/\".concat(id)).then(res => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n    }).catch(err => {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleFetchRequest(slug) {\n  return dispatch => {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(\"articles/published/\".concat(slug)).then(res => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n    }).catch(err => {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL3NlcnZpY2UuanM/Y2YwNiJdLCJuYW1lcyI6WyJ0cmFuc2Zvcm1SZXF1ZXN0IiwicGFybXMiLCJUcmFuc2Zvcm1lciIsInNlbmQiLCJ0cmFuc2Zvcm1SZXNwb25zZSIsInBhcmFtcyIsImZldGNoIiwiYXJ0aWNsZUFkZFJlcXVlc3QiLCJkaXNwYXRjaCIsIlByb21pc2UiLCJyZXNvbHZlIiwicmVqZWN0IiwiSHR0cCIsInBvc3QiLCJ0aGVuIiwicmVzIiwiYXJ0aWNsZUFjdGlvbnMiLCJkYXRhIiwiY2F0Y2giLCJlcnIiLCJzdGF0dXNDb2RlIiwicmVzcG9uc2UiLCJzdGF0dXMiLCJlcnJvciIsInJlc2V0RXJyb3JzIiwiZXJyb3JzIiwicmVwbGFjZSIsInNlYXJjaFN0ciIsInJlcGxhY2VTdHIiLCJyZXNldFZhbGlkYXRpb25GaWVsZHMiLCJtZXNzYWdlIiwiYXJ0aWNsZVVwZGF0ZVJlcXVlc3QiLCJwYXRjaCIsImlkIiwiYXJ0aWNsZVJlbW92ZVJlcXVlc3QiLCJkZWxldGUiLCJjb25zb2xlIiwiYXJ0aWNsZUxpc3RSZXF1ZXN0IiwicGFnZU51bWJlciIsInVybCIsImdldCIsImFydGljbGVFZGl0UmVxdWVzdCIsImFydGljbGVGZXRjaFJlcXVlc3QiLCJzbHVnIl0sIm1hcHBpbmdzIjoiQUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQ0E7QUFDQTs7QUFFQSxTQUFTQSxnQkFBVCxDQUEwQkMsS0FBMUIsRUFBaUM7QUFDL0IsU0FBT0MsMERBQVcsQ0FBQ0MsSUFBWixDQUFpQkYsS0FBakIsQ0FBUDtBQUNEOztBQUVELFNBQVNHLGlCQUFULENBQTJCQyxNQUEzQixFQUFtQztBQUNqQyxTQUFPSCwwREFBVyxDQUFDSSxLQUFaLENBQWtCRCxNQUFsQixDQUFQO0FBQ0Q7O0FBRU0sU0FBU0UsaUJBQVQsQ0FBMkJGLE1BQTNCLEVBQW1DO0FBQ3hDLFNBQU9HLFFBQVEsSUFDYixJQUFJQyxPQUFKLENBQVksQ0FBQ0MsT0FBRCxFQUFVQyxNQUFWLEtBQXFCO0FBQy9CQyx1REFBSSxDQUFDQyxJQUFMLENBQVUsV0FBVixFQUF1QmIsZ0JBQWdCLENBQUNLLE1BQUQsQ0FBdkMsRUFDR1MsSUFESCxDQUNRQyxHQUFHLElBQUk7QUFDWFAsY0FBUSxDQUFDUSxrREFBQSxDQUFtQlosaUJBQWlCLENBQUNXLEdBQUcsQ0FBQ0UsSUFBTCxDQUFwQyxDQUFELENBQVI7QUFDQSxhQUFPUCxPQUFPLEVBQWQ7QUFDRCxLQUpILEVBS0dRLEtBTEgsQ0FLVUMsR0FBRCxJQUFTO0FBQ2QsWUFBTUMsVUFBVSxHQUFHRCxHQUFHLENBQUNFLFFBQUosQ0FBYUMsTUFBaEM7QUFDQSxZQUFNTCxJQUFJLEdBQUc7QUFDWE0sYUFBSyxFQUFFLElBREk7QUFFWEg7QUFGVyxPQUFiOztBQUtBLFVBQUlBLFVBQVUsS0FBSyxHQUFuQixFQUF3QjtBQUN0QixjQUFNSSxXQUFXLEdBQUc7QUFDbEJDLGdCQUFNLEVBQUVOLEdBQUcsQ0FBQ0UsUUFBSixDQUFhSixJQURIO0FBRWxCUyxpQkFBTyxFQUFFLEtBRlM7QUFHbEJDLG1CQUFTLEVBQUUsRUFITztBQUlsQkMsb0JBQVUsRUFBRTtBQUpNLFNBQXBCO0FBTUFYLFlBQUksQ0FBQ00sS0FBTCxHQUFhckIsMERBQVcsQ0FBQzJCLHFCQUFaLENBQWtDTCxXQUFsQyxDQUFiO0FBQ0QsT0FSRCxNQVFPLElBQUlKLFVBQVUsS0FBSyxHQUFuQixFQUF3QjtBQUM3QkgsWUFBSSxDQUFDTSxLQUFMLEdBQWFKLEdBQUcsQ0FBQ0UsUUFBSixDQUFhSixJQUFiLENBQWtCYSxPQUEvQjtBQUNEOztBQUNELGFBQU9uQixNQUFNLENBQUNNLElBQUQsQ0FBYjtBQUNELEtBeEJIO0FBeUJELEdBMUJELENBREY7QUE2QkQ7QUFFTSxTQUFTYyxvQkFBVCxDQUE4QjFCLE1BQTlCLEVBQXNDO0FBQzNDLFNBQU9HLFFBQVEsSUFDYixJQUFJQyxPQUFKLENBQVksQ0FBQ0MsT0FBRCxFQUFVQyxNQUFWLEtBQXFCO0FBQy9CQyx1REFBSSxDQUFDb0IsS0FBTCxvQkFBdUIzQixNQUFNLENBQUM0QixFQUE5QixHQUFvQ2pDLGdCQUFnQixDQUFDSyxNQUFELENBQXBELEVBQ0dTLElBREgsQ0FDUUMsR0FBRyxJQUFJO0FBQ1hQLGNBQVEsQ0FBQ1Esa0RBQUEsQ0FBbUJaLGlCQUFpQixDQUFDVyxHQUFHLENBQUNFLElBQUwsQ0FBcEMsQ0FBRCxDQUFSO0FBQ0EsYUFBT1AsT0FBTyxFQUFkO0FBQ0QsS0FKSCxFQUtHUSxLQUxILENBS1VDLEdBQUQsSUFBUztBQUNkLFlBQU1DLFVBQVUsR0FBR0QsR0FBRyxDQUFDRSxRQUFKLENBQWFDLE1BQWhDO0FBQ0EsWUFBTUwsSUFBSSxHQUFHO0FBQ1hNLGFBQUssRUFBRSxJQURJO0FBRVhIO0FBRlcsT0FBYjs7QUFLQSxVQUFJQSxVQUFVLEtBQUssR0FBbkIsRUFBd0I7QUFDdEIsY0FBTUksV0FBVyxHQUFHO0FBQ2xCQyxnQkFBTSxFQUFFTixHQUFHLENBQUNFLFFBQUosQ0FBYUosSUFESDtBQUVsQlMsaUJBQU8sRUFBRSxLQUZTO0FBR2xCQyxtQkFBUyxFQUFFLEVBSE87QUFJbEJDLG9CQUFVLEVBQUU7QUFKTSxTQUFwQjtBQU1BWCxZQUFJLENBQUNNLEtBQUwsR0FBYXJCLDBEQUFXLENBQUMyQixxQkFBWixDQUFrQ0wsV0FBbEMsQ0FBYjtBQUNELE9BUkQsTUFRTyxJQUFJSixVQUFVLEtBQUssR0FBbkIsRUFBd0I7QUFDN0JILFlBQUksQ0FBQ00sS0FBTCxHQUFhSixHQUFHLENBQUNFLFFBQUosQ0FBYUosSUFBYixDQUFrQmEsT0FBL0I7QUFDRDs7QUFDRCxhQUFPbkIsTUFBTSxDQUFDTSxJQUFELENBQWI7QUFDRCxLQXhCSDtBQXlCRCxHQTFCRCxDQURGO0FBNkJEO0FBRU0sU0FBU2lCLG9CQUFULENBQThCRCxFQUE5QixFQUFrQztBQUN2QyxTQUFPekIsUUFBUSxJQUFJO0FBQ2pCSSx1REFBSSxDQUFDdUIsTUFBTCxvQkFBd0JGLEVBQXhCLEdBQ0duQixJQURILENBQ1EsTUFBTTtBQUNWTixjQUFRLENBQUNRLHFEQUFBLENBQXNCaUIsRUFBdEIsQ0FBRCxDQUFSO0FBQ0QsS0FISCxFQUlHZixLQUpILENBSVVDLEdBQUQsSUFBUztBQUNkO0FBQ0FpQixhQUFPLENBQUNiLEtBQVIsQ0FBY0osR0FBRyxDQUFDRSxRQUFsQjtBQUNELEtBUEg7QUFRRCxHQVREO0FBVUQ7QUFFTSxTQUFTZ0Isa0JBQVQsQ0FBNEI7QUFBQ0MsWUFBVSxHQUFHLENBQWQ7QUFBaUJDLEtBQUcsR0FBRztBQUF2QixDQUE1QixFQUFpRTtBQUN0RSxTQUFPL0IsUUFBUSxJQUFJO0FBQ2pCLFFBQUk4QixVQUFVLEdBQUcsQ0FBakIsRUFBb0I7QUFDbEJDLFNBQUcsR0FBR0EsR0FBRyxtQkFBWUQsVUFBWixDQUFUO0FBQ0Q7O0FBRUQxQix1REFBSSxDQUFDNEIsR0FBTCxDQUFTRCxHQUFULEVBQ0d6QixJQURILENBQ1NDLEdBQUQsSUFBUztBQUNiUCxjQUFRLENBQUNRLG1EQUFBLENBQW9CWixpQkFBaUIsQ0FBQ1csR0FBRyxDQUFDRSxJQUFMLENBQXJDLENBQUQsQ0FBUjtBQUNELEtBSEgsRUFJR0MsS0FKSCxDQUlVQyxHQUFELElBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FiRDtBQWNEO0FBRU0sU0FBU29CLGtCQUFULENBQTRCUixFQUE1QixFQUFnQztBQUNyQyxTQUFPekIsUUFBUSxJQUFJO0FBQ2pCSSx1REFBSSxDQUFDNEIsR0FBTCxvQkFBcUJQLEVBQXJCLEdBQ0duQixJQURILENBQ1NDLEdBQUQsSUFBUztBQUNiUCxjQUFRLENBQUNRLGtEQUFBLENBQW1CWixpQkFBaUIsQ0FBQ1csR0FBRyxDQUFDRSxJQUFMLENBQXBDLENBQUQsQ0FBUjtBQUNELEtBSEgsRUFJR0MsS0FKSCxDQUlVQyxHQUFELElBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FURDtBQVVEO0FBRU0sU0FBU3FCLG1CQUFULENBQTZCQyxJQUE3QixFQUFtQztBQUN4QyxTQUFPbkMsUUFBUSxJQUFJO0FBQ2pCSSx1REFBSSxDQUFDNEIsR0FBTCw4QkFBK0JHLElBQS9CLEdBQ0c3QixJQURILENBQ1NDLEdBQUQsSUFBUztBQUNiUCxjQUFRLENBQUNRLGtEQUFBLENBQW1CWixpQkFBaUIsQ0FBQ1csR0FBRyxDQUFDRSxJQUFMLENBQXBDLENBQUQsQ0FBUjtBQUNELEtBSEgsRUFJR0MsS0FKSCxDQUlVQyxHQUFELElBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FURDtBQVVEIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvYXJ0aWNsZS9zZXJ2aWNlLmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiaW1wb3J0IEh0dHAgZnJvbSAnLi4vLi4vdXRpbHMvSHR0cCdcbmltcG9ydCBUcmFuc2Zvcm1lciBmcm9tICcuLi8uLi91dGlscy9UcmFuc2Zvcm1lcidcbmltcG9ydCAqIGFzIGFydGljbGVBY3Rpb25zIGZyb20gJy4vc3RvcmUvYWN0aW9ucydcblxuZnVuY3Rpb24gdHJhbnNmb3JtUmVxdWVzdChwYXJtcykge1xuICByZXR1cm4gVHJhbnNmb3JtZXIuc2VuZChwYXJtcylcbn1cblxuZnVuY3Rpb24gdHJhbnNmb3JtUmVzcG9uc2UocGFyYW1zKSB7XG4gIHJldHVybiBUcmFuc2Zvcm1lci5mZXRjaChwYXJhbXMpXG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlQWRkUmVxdWVzdChwYXJhbXMpIHtcbiAgcmV0dXJuIGRpc3BhdGNoID0+IChcbiAgICBuZXcgUHJvbWlzZSgocmVzb2x2ZSwgcmVqZWN0KSA9PiB7XG4gICAgICBIdHRwLnBvc3QoJy9hcnRpY2xlcycsIHRyYW5zZm9ybVJlcXVlc3QocGFyYW1zKSlcbiAgICAgICAgLnRoZW4ocmVzID0+IHtcbiAgICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5hZGQodHJhbnNmb3JtUmVzcG9uc2UocmVzLmRhdGEpKSlcbiAgICAgICAgICByZXR1cm4gcmVzb2x2ZSgpXG4gICAgICAgIH0pXG4gICAgICAgIC5jYXRjaCgoZXJyKSA9PiB7XG4gICAgICAgICAgY29uc3Qgc3RhdHVzQ29kZSA9IGVyci5yZXNwb25zZS5zdGF0dXM7XG4gICAgICAgICAgY29uc3QgZGF0YSA9IHtcbiAgICAgICAgICAgIGVycm9yOiBudWxsLFxuICAgICAgICAgICAgc3RhdHVzQ29kZSxcbiAgICAgICAgICB9O1xuXG4gICAgICAgICAgaWYgKHN0YXR1c0NvZGUgPT09IDQyMikge1xuICAgICAgICAgICAgY29uc3QgcmVzZXRFcnJvcnMgPSB7XG4gICAgICAgICAgICAgIGVycm9yczogZXJyLnJlc3BvbnNlLmRhdGEsXG4gICAgICAgICAgICAgIHJlcGxhY2U6IGZhbHNlLFxuICAgICAgICAgICAgICBzZWFyY2hTdHI6ICcnLFxuICAgICAgICAgICAgICByZXBsYWNlU3RyOiAnJyxcbiAgICAgICAgICAgIH07XG4gICAgICAgICAgICBkYXRhLmVycm9yID0gVHJhbnNmb3JtZXIucmVzZXRWYWxpZGF0aW9uRmllbGRzKHJlc2V0RXJyb3JzKTtcbiAgICAgICAgICB9IGVsc2UgaWYgKHN0YXR1c0NvZGUgPT09IDQwMSkge1xuICAgICAgICAgICAgZGF0YS5lcnJvciA9IGVyci5yZXNwb25zZS5kYXRhLm1lc3NhZ2U7XG4gICAgICAgICAgfVxuICAgICAgICAgIHJldHVybiByZWplY3QoZGF0YSk7XG4gICAgICAgIH0pXG4gICAgfSlcbiAgKVxufVxuXG5leHBvcnQgZnVuY3Rpb24gYXJ0aWNsZVVwZGF0ZVJlcXVlc3QocGFyYW1zKSB7XG4gIHJldHVybiBkaXNwYXRjaCA9PiAoXG4gICAgbmV3IFByb21pc2UoKHJlc29sdmUsIHJlamVjdCkgPT4ge1xuICAgICAgSHR0cC5wYXRjaChgYXJ0aWNsZXMvJHtwYXJhbXMuaWR9YCwgdHJhbnNmb3JtUmVxdWVzdChwYXJhbXMpKVxuICAgICAgICAudGhlbihyZXMgPT4ge1xuICAgICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLmFkZCh0cmFuc2Zvcm1SZXNwb25zZShyZXMuZGF0YSkpKVxuICAgICAgICAgIHJldHVybiByZXNvbHZlKClcbiAgICAgICAgfSlcbiAgICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgICBjb25zdCBzdGF0dXNDb2RlID0gZXJyLnJlc3BvbnNlLnN0YXR1cztcbiAgICAgICAgICBjb25zdCBkYXRhID0ge1xuICAgICAgICAgICAgZXJyb3I6IG51bGwsXG4gICAgICAgICAgICBzdGF0dXNDb2RlLFxuICAgICAgICAgIH07XG5cbiAgICAgICAgICBpZiAoc3RhdHVzQ29kZSA9PT0gNDIyKSB7XG4gICAgICAgICAgICBjb25zdCByZXNldEVycm9ycyA9IHtcbiAgICAgICAgICAgICAgZXJyb3JzOiBlcnIucmVzcG9uc2UuZGF0YSxcbiAgICAgICAgICAgICAgcmVwbGFjZTogZmFsc2UsXG4gICAgICAgICAgICAgIHNlYXJjaFN0cjogJycsXG4gICAgICAgICAgICAgIHJlcGxhY2VTdHI6ICcnLFxuICAgICAgICAgICAgfTtcbiAgICAgICAgICAgIGRhdGEuZXJyb3IgPSBUcmFuc2Zvcm1lci5yZXNldFZhbGlkYXRpb25GaWVsZHMocmVzZXRFcnJvcnMpO1xuICAgICAgICAgIH0gZWxzZSBpZiAoc3RhdHVzQ29kZSA9PT0gNDAxKSB7XG4gICAgICAgICAgICBkYXRhLmVycm9yID0gZXJyLnJlc3BvbnNlLmRhdGEubWVzc2FnZTtcbiAgICAgICAgICB9XG4gICAgICAgICAgcmV0dXJuIHJlamVjdChkYXRhKTtcbiAgICAgICAgfSlcbiAgICB9KVxuICApXG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlUmVtb3ZlUmVxdWVzdChpZCkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIEh0dHAuZGVsZXRlKGBhcnRpY2xlcy8ke2lkfWApXG4gICAgICAudGhlbigoKSA9PiB7XG4gICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLnJlbW92ZShpZCkpXG4gICAgICB9KVxuICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgLy8gVE9ETzogaGFuZGxlIGVyclxuICAgICAgICBjb25zb2xlLmVycm9yKGVyci5yZXNwb25zZSlcbiAgICAgIH0pXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIGFydGljbGVMaXN0UmVxdWVzdCh7cGFnZU51bWJlciA9IDEsIHVybCA9ICcvYXJ0aWNsZXMnfSkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIGlmIChwYWdlTnVtYmVyID4gMSkge1xuICAgICAgdXJsID0gdXJsICsgYD9wYWdlPSR7cGFnZU51bWJlcn1gXG4gICAgfVxuXG4gICAgSHR0cC5nZXQodXJsKVxuICAgICAgLnRoZW4oKHJlcykgPT4ge1xuICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5saXN0KHRyYW5zZm9ybVJlc3BvbnNlKHJlcy5kYXRhKSkpXG4gICAgICB9KVxuICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgLy8gVE9ETzogaGFuZGxlIGVyclxuICAgICAgICBjb25zb2xlLmVycm9yKGVyci5yZXNwb25zZSlcbiAgICAgIH0pXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIGFydGljbGVFZGl0UmVxdWVzdChpZCkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIEh0dHAuZ2V0KGBhcnRpY2xlcy8ke2lkfWApXG4gICAgICAudGhlbigocmVzKSA9PiB7XG4gICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLmFkZCh0cmFuc2Zvcm1SZXNwb25zZShyZXMuZGF0YSkpKVxuICAgICAgfSlcbiAgICAgIC5jYXRjaCgoZXJyKSA9PiB7XG4gICAgICAgIC8vIFRPRE86IGhhbmRsZSBlcnJcbiAgICAgICAgY29uc29sZS5lcnJvcihlcnIucmVzcG9uc2UpXG4gICAgICB9KVxuICB9XG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlRmV0Y2hSZXF1ZXN0KHNsdWcpIHtcbiAgcmV0dXJuIGRpc3BhdGNoID0+IHtcbiAgICBIdHRwLmdldChgYXJ0aWNsZXMvcHVibGlzaGVkLyR7c2x1Z31gKVxuICAgICAgLnRoZW4oKHJlcykgPT4ge1xuICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5hZGQodHJhbnNmb3JtUmVzcG9uc2UocmVzLmRhdGEpKSlcbiAgICAgIH0pXG4gICAgICAuY2F0Y2goKGVycikgPT4ge1xuICAgICAgICAvLyBUT0RPOiBoYW5kbGUgZXJyXG4gICAgICAgIGNvbnNvbGUuZXJyb3IoZXJyLnJlc3BvbnNlKVxuICAgICAgfSlcbiAgfVxufSJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/article/service.js\n");
 
 /***/ }),
 
@@ -68,7 +68,43 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"add\", function() { return add; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"update\", function() { return update; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"remove\", function() { return remove; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"list\", function() { return list; });\n/* harmony import */ var _action_types__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./action-types */ \"./resources/js/modules/article/store/action-types.js\");\n/* ============\n * Actions for the article module\n * ============\n *\n * The actions that are available on the\n * article module.\n */\n\nfunction add(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_ADD\"],\n    payload: payload\n  };\n}\nfunction update(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_UPDATE\"],\n    payload: payload\n  };\n}\nfunction remove(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_REMOVE\"],\n    payload: payload\n  };\n}\nfunction list(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_LIST\"],\n    payload: payload\n  };\n}//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL3N0b3JlL2FjdGlvbnMuanM/ZjY0ZCJdLCJuYW1lcyI6WyJhZGQiLCJwYXlsb2FkIiwidHlwZSIsIkFSVElDTEVfQUREIiwidXBkYXRlIiwiQVJUSUNMRV9VUERBVEUiLCJyZW1vdmUiLCJBUlRJQ0xFX1JFTU9WRSIsImxpc3QiLCJBUlRJQ0xFX0xJU1QiXSwibWFwcGluZ3MiOiJBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBOzs7Ozs7O0FBUUE7QUFPTyxTQUFTQSxHQUFULENBQWFDLE9BQWIsRUFBc0I7QUFDM0IsU0FBTztBQUNMQyxRQUFJLEVBQUVDLHlEQUREO0FBRUxGLFdBQU8sRUFBUEE7QUFGSyxHQUFQO0FBSUQ7QUFFTSxTQUFTRyxNQUFULENBQWdCSCxPQUFoQixFQUF5QjtBQUM5QixTQUFPO0FBQ0xDLFFBQUksRUFBRUcsNERBREQ7QUFFTEosV0FBTyxFQUFQQTtBQUZLLEdBQVA7QUFJRDtBQUVNLFNBQVNLLE1BQVQsQ0FBZ0JMLE9BQWhCLEVBQXlCO0FBQzlCLFNBQU87QUFDTEMsUUFBSSxFQUFFSyw0REFERDtBQUVMTixXQUFPLEVBQVBBO0FBRkssR0FBUDtBQUlEO0FBRU0sU0FBU08sSUFBVCxDQUFjUCxPQUFkLEVBQXVCO0FBQzVCLFNBQU87QUFDTEMsUUFBSSxFQUFFTywwREFERDtBQUVMUixXQUFPLEVBQVBBO0FBRkssR0FBUDtBQUlEIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvYXJ0aWNsZS9zdG9yZS9hY3Rpb25zLmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiLyogPT09PT09PT09PT09XG4gKiBBY3Rpb25zIGZvciB0aGUgYXJ0aWNsZSBtb2R1bGVcbiAqID09PT09PT09PT09PVxuICpcbiAqIFRoZSBhY3Rpb25zIHRoYXQgYXJlIGF2YWlsYWJsZSBvbiB0aGVcbiAqIGFydGljbGUgbW9kdWxlLlxuICovXG5cbmltcG9ydCB7XG4gIEFSVElDTEVfQURELFxuICBBUlRJQ0xFX1VQREFURSxcbiAgQVJUSUNMRV9SRU1PVkUsXG4gIEFSVElDTEVfTElTVCxcbn0gZnJvbSAnLi9hY3Rpb24tdHlwZXMnO1xuXG5leHBvcnQgZnVuY3Rpb24gYWRkKHBheWxvYWQpIHtcbiAgcmV0dXJuIHtcbiAgICB0eXBlOiBBUlRJQ0xFX0FERCxcbiAgICBwYXlsb2FkXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIHVwZGF0ZShwYXlsb2FkKSB7XG4gIHJldHVybiB7XG4gICAgdHlwZTogQVJUSUNMRV9VUERBVEUsXG4gICAgcGF5bG9hZFxuICB9XG59XG5cbmV4cG9ydCBmdW5jdGlvbiByZW1vdmUocGF5bG9hZCkge1xuICByZXR1cm4ge1xuICAgIHR5cGU6IEFSVElDTEVfUkVNT1ZFLFxuICAgIHBheWxvYWRcbiAgfVxufVxuXG5leHBvcnQgZnVuY3Rpb24gbGlzdChwYXlsb2FkKSB7XG4gIHJldHVybiB7XG4gICAgdHlwZTogQVJUSUNMRV9MSVNULFxuICAgIHBheWxvYWRcbiAgfVxufSJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/article/store/actions.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"add\", function() { return add; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"update\", function() { return update; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"remove\", function() { return remove; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"list\", function() { return list; });\n/* harmony import */ var _action_types__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./action-types */ \"./resources/js/modules/article/store/action-types.js\");\n/* ============\n * Actions for the article module\n * ============\n *\n * The actions that are available on the\n * article module.\n */\n\nfunction add(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_ADD\"],\n    payload\n  };\n}\nfunction update(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_UPDATE\"],\n    payload\n  };\n}\nfunction remove(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_REMOVE\"],\n    payload\n  };\n}\nfunction list(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_LIST\"],\n    payload\n  };\n}//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL3N0b3JlL2FjdGlvbnMuanM/ZjY0ZCJdLCJuYW1lcyI6WyJhZGQiLCJwYXlsb2FkIiwidHlwZSIsIkFSVElDTEVfQUREIiwidXBkYXRlIiwiQVJUSUNMRV9VUERBVEUiLCJyZW1vdmUiLCJBUlRJQ0xFX1JFTU9WRSIsImxpc3QiLCJBUlRJQ0xFX0xJU1QiXSwibWFwcGluZ3MiOiJBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBOzs7Ozs7O0FBUUE7QUFPTyxTQUFTQSxHQUFULENBQWFDLE9BQWIsRUFBc0I7QUFDM0IsU0FBTztBQUNMQyxRQUFJLEVBQUVDLHlEQUREO0FBRUxGO0FBRkssR0FBUDtBQUlEO0FBRU0sU0FBU0csTUFBVCxDQUFnQkgsT0FBaEIsRUFBeUI7QUFDOUIsU0FBTztBQUNMQyxRQUFJLEVBQUVHLDREQUREO0FBRUxKO0FBRkssR0FBUDtBQUlEO0FBRU0sU0FBU0ssTUFBVCxDQUFnQkwsT0FBaEIsRUFBeUI7QUFDOUIsU0FBTztBQUNMQyxRQUFJLEVBQUVLLDREQUREO0FBRUxOO0FBRkssR0FBUDtBQUlEO0FBRU0sU0FBU08sSUFBVCxDQUFjUCxPQUFkLEVBQXVCO0FBQzVCLFNBQU87QUFDTEMsUUFBSSxFQUFFTywwREFERDtBQUVMUjtBQUZLLEdBQVA7QUFJRCIsImZpbGUiOiIuL3Jlc291cmNlcy9qcy9tb2R1bGVzL2FydGljbGUvc3RvcmUvYWN0aW9ucy5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbIi8qID09PT09PT09PT09PVxuICogQWN0aW9ucyBmb3IgdGhlIGFydGljbGUgbW9kdWxlXG4gKiA9PT09PT09PT09PT1cbiAqXG4gKiBUaGUgYWN0aW9ucyB0aGF0IGFyZSBhdmFpbGFibGUgb24gdGhlXG4gKiBhcnRpY2xlIG1vZHVsZS5cbiAqL1xuXG5pbXBvcnQge1xuICBBUlRJQ0xFX0FERCxcbiAgQVJUSUNMRV9VUERBVEUsXG4gIEFSVElDTEVfUkVNT1ZFLFxuICBBUlRJQ0xFX0xJU1QsXG59IGZyb20gJy4vYWN0aW9uLXR5cGVzJztcblxuZXhwb3J0IGZ1bmN0aW9uIGFkZChwYXlsb2FkKSB7XG4gIHJldHVybiB7XG4gICAgdHlwZTogQVJUSUNMRV9BREQsXG4gICAgcGF5bG9hZFxuICB9XG59XG5cbmV4cG9ydCBmdW5jdGlvbiB1cGRhdGUocGF5bG9hZCkge1xuICByZXR1cm4ge1xuICAgIHR5cGU6IEFSVElDTEVfVVBEQVRFLFxuICAgIHBheWxvYWRcbiAgfVxufVxuXG5leHBvcnQgZnVuY3Rpb24gcmVtb3ZlKHBheWxvYWQpIHtcbiAgcmV0dXJuIHtcbiAgICB0eXBlOiBBUlRJQ0xFX1JFTU9WRSxcbiAgICBwYXlsb2FkXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIGxpc3QocGF5bG9hZCkge1xuICByZXR1cm4ge1xuICAgIHR5cGU6IEFSVElDTEVfTElTVCxcbiAgICBwYXlsb2FkXG4gIH1cbn0iXSwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./resources/js/modules/article/store/actions.js\n");
+
+/***/ }),
+
+/***/ "./resources/js/modules/web/pages/home/Page.js":
+/*!*****************************************************!*\
+  !*** ./resources/js/modules/web/pages/home/Page.js ***!
+  \*****************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\");\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var _components_Header__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./components/Header */ \"./resources/js/modules/web/pages/home/components/Header.js\");\n/* harmony import */ var _common_articles_listing_index__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../../../common/articles/listing/index */ \"./resources/js/common/articles/listing/index.js\");\n/* harmony import */ var _article_service__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../../article/service */ \"./resources/js/modules/article/service.js\");\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\n\n // import components\n\n\n // import services\n\n\n\nclass Page extends react__WEBPACK_IMPORTED_MODULE_0__[\"Component\"] {\n  componentDidMount() {\n    this.props.dispatch(Object(_article_service__WEBPACK_IMPORTED_MODULE_4__[\"articleListRequest\"])({\n      url: '/articles/published'\n    }));\n  }\n\n  render() {\n    return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", null, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_components_Header__WEBPACK_IMPORTED_MODULE_2__[\"default\"], null), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(_common_articles_listing_index__WEBPACK_IMPORTED_MODULE_3__[\"default\"], null));\n  }\n\n}\n\n_defineProperty(Page, \"displayName\", \"HomePage\");\n\n_defineProperty(Page, \"propTypes\", {\n  dispatch: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.func.isRequired\n});\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Page);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy93ZWIvcGFnZXMvaG9tZS9QYWdlLmpzPzMwMGIiXSwibmFtZXMiOlsiUGFnZSIsIkNvbXBvbmVudCIsImNvbXBvbmVudERpZE1vdW50IiwicHJvcHMiLCJkaXNwYXRjaCIsImFydGljbGVMaXN0UmVxdWVzdCIsInVybCIsInJlbmRlciIsIlByb3BUeXBlcyIsImZ1bmMiLCJpc1JlcXVpcmVkIl0sIm1hcHBpbmdzIjoiOzs7Ozs7Ozs7O0FBQUE7Q0FHQTs7QUFDQTtDQUdBOztBQUNBOztBQUVBLE1BQU1BLElBQU4sU0FBbUJDLCtDQUFuQixDQUE2QjtBQU0zQkMsbUJBQWlCLEdBQUc7QUFDbEIsU0FBS0MsS0FBTCxDQUFXQyxRQUFYLENBQW9CQywyRUFBa0IsQ0FBQztBQUFFQyxTQUFHLEVBQUU7QUFBUCxLQUFELENBQXRDO0FBQ0Q7O0FBRURDLFFBQU0sR0FBRztBQUNQLFdBQU8sd0VBQ0wsMkRBQUMsMERBQUQsT0FESyxFQUVMLDJEQUFDLHNFQUFELE9BRkssQ0FBUDtBQUlEOztBQWYwQjs7Z0JBQXZCUCxJLGlCQUNpQixVOztnQkFEakJBLEksZUFFZTtBQUNqQkksVUFBUSxFQUFFSSxpREFBUyxDQUFDQyxJQUFWLENBQWVDO0FBRFIsQzs7QUFnQk5WLG1FQUFmIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvd2ViL3BhZ2VzL2hvbWUvUGFnZS5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBSZWFjdCwgeyBDb21wb25lbnQgfSBmcm9tIFwicmVhY3RcIlxuaW1wb3J0IFByb3BUeXBlcyBmcm9tIFwicHJvcC10eXBlc1wiXG5cbi8vIGltcG9ydCBjb21wb25lbnRzXG5pbXBvcnQgSGVhZGVyIGZyb20gXCIuL2NvbXBvbmVudHMvSGVhZGVyXCJcbmltcG9ydCBBcnRpY2xlcyBmcm9tIFwiLi4vLi4vLi4vLi4vY29tbW9uL2FydGljbGVzL2xpc3RpbmcvaW5kZXhcIlxuXG4vLyBpbXBvcnQgc2VydmljZXNcbmltcG9ydCB7IGFydGljbGVMaXN0UmVxdWVzdCB9IGZyb20gXCIuLi8uLi8uLi9hcnRpY2xlL3NlcnZpY2VcIlxuXG5jbGFzcyBQYWdlIGV4dGVuZHMgQ29tcG9uZW50IHtcbiAgc3RhdGljIGRpc3BsYXlOYW1lID0gXCJIb21lUGFnZVwiXG4gIHN0YXRpYyBwcm9wVHlwZXMgPSB7XG4gICAgZGlzcGF0Y2g6IFByb3BUeXBlcy5mdW5jLmlzUmVxdWlyZWQsXG4gIH1cblxuICBjb21wb25lbnREaWRNb3VudCgpIHtcbiAgICB0aGlzLnByb3BzLmRpc3BhdGNoKGFydGljbGVMaXN0UmVxdWVzdCh7IHVybDogJy9hcnRpY2xlcy9wdWJsaXNoZWQnIH0pKVxuICB9XG5cbiAgcmVuZGVyKCkge1xuICAgIHJldHVybiA8ZGl2PlxuICAgICAgPEhlYWRlci8+XG4gICAgICA8QXJ0aWNsZXMvPlxuICAgIDwvZGl2PlxuICB9XG59XG5cbmV4cG9ydCBkZWZhdWx0IFBhZ2VcbiJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/web/pages/home/Page.js\n");
+
+/***/ }),
+
+/***/ "./resources/js/modules/web/pages/home/components/Header.js":
+/*!******************************************************************!*\
+  !*** ./resources/js/modules/web/pages/home/components/Header.js ***!
+  \******************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n\nconst displayName = \"HomePageHeader\";\n\nfunction Header() {\n  return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"header\", {\n    className: \"bg-primary text-white\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n    className: \"container text-center\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"img\", {\n    width: \"125\",\n    height: \"125\",\n    src: \"https://graph.facebook.com/1243067599/picture?type=square\",\n    alt: \"...\",\n    className: \"rounded-circle\"\n  }), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"h1\", null, \"Moeen Farooq Basra\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"p\", {\n    className: \"lead\"\n  }, \"Master in Information Technology\"), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"p\", {\n    className: \"lead\"\n  }, \"Fullstack Developer at\\xA0\", react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"a\", {\n    className: \"text-white\",\n    href: \"http://awok.com\",\n    target: \"_blank\",\n    rel: \"noreferrer noopener\"\n  }, \"Awok.com\")), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"p\", {\n    className: \"lead\"\n  }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"i\", {\n    className: \"fa fa-heart text-danger\"\n  }), \"{ PHP, JavaScript, Python, MySQL, MongoDB }\")));\n}\n\nHeader.displayName = displayName;\n/* harmony default export */ __webpack_exports__[\"default\"] = (Header);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy93ZWIvcGFnZXMvaG9tZS9jb21wb25lbnRzL0hlYWRlci5qcz81ZGNkIl0sIm5hbWVzIjpbImRpc3BsYXlOYW1lIiwiSGVhZGVyIl0sIm1hcHBpbmdzIjoiQUFBQTtBQUFBO0FBQUE7QUFBQTtBQUVBLE1BQU1BLFdBQVcsR0FBRyxnQkFBcEI7O0FBRUEsU0FBU0MsTUFBVCxHQUFrQjtBQUNoQixTQUFPO0FBQVEsYUFBUyxFQUFDO0FBQWxCLEtBQ0w7QUFBSyxhQUFTLEVBQUM7QUFBZixLQUNFO0FBQUssU0FBSyxFQUFDLEtBQVg7QUFBaUIsVUFBTSxFQUFDLEtBQXhCO0FBQThCLE9BQUcsRUFBQywyREFBbEM7QUFBOEYsT0FBRyxFQUFDLEtBQWxHO0FBQXdHLGFBQVMsRUFBQztBQUFsSCxJQURGLEVBRUUsNEZBRkYsRUFHRTtBQUFHLGFBQVMsRUFBQztBQUFiLHdDQUhGLEVBSUU7QUFBRyxhQUFTLEVBQUM7QUFBYixtQ0FDRTtBQUFHLGFBQVMsRUFBQyxZQUFiO0FBQ0csUUFBSSxFQUFDLGlCQURSO0FBRUcsVUFBTSxFQUFDLFFBRlY7QUFHRyxPQUFHLEVBQUM7QUFIUCxnQkFERixDQUpGLEVBVUU7QUFBRyxhQUFTLEVBQUM7QUFBYixLQUFvQjtBQUFHLGFBQVMsRUFBQztBQUFiLElBQXBCLGdEQVZGLENBREssQ0FBUDtBQWNEOztBQUNEQSxNQUFNLENBQUNELFdBQVAsR0FBcUJBLFdBQXJCO0FBRWVDLHFFQUFmIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvd2ViL3BhZ2VzL2hvbWUvY29tcG9uZW50cy9IZWFkZXIuanMuanMiLCJzb3VyY2VzQ29udGVudCI6WyJpbXBvcnQgUmVhY3QgZnJvbSBcInJlYWN0XCJcblxuY29uc3QgZGlzcGxheU5hbWUgPSBcIkhvbWVQYWdlSGVhZGVyXCJcblxuZnVuY3Rpb24gSGVhZGVyKCkge1xuICByZXR1cm4gPGhlYWRlciBjbGFzc05hbWU9XCJiZy1wcmltYXJ5IHRleHQtd2hpdGVcIj5cbiAgICA8ZGl2IGNsYXNzTmFtZT1cImNvbnRhaW5lciB0ZXh0LWNlbnRlclwiPlxuICAgICAgPGltZyB3aWR0aD1cIjEyNVwiIGhlaWdodD1cIjEyNVwiIHNyYz1cImh0dHBzOi8vZ3JhcGguZmFjZWJvb2suY29tLzEyNDMwNjc1OTkvcGljdHVyZT90eXBlPXNxdWFyZVwiIGFsdD1cIi4uLlwiIGNsYXNzTmFtZT1cInJvdW5kZWQtY2lyY2xlXCIgLz5cbiAgICAgIDxoMT5Nb2VlbiBGYXJvb3EgQmFzcmE8L2gxPlxuICAgICAgPHAgY2xhc3NOYW1lPVwibGVhZFwiPk1hc3RlciBpbiBJbmZvcm1hdGlvbiBUZWNobm9sb2d5PC9wPlxuICAgICAgPHAgY2xhc3NOYW1lPVwibGVhZFwiPkZ1bGxzdGFjayBEZXZlbG9wZXIgYXQmbmJzcDtcbiAgICAgICAgPGEgY2xhc3NOYW1lPVwidGV4dC13aGl0ZVwiXG4gICAgICAgICAgIGhyZWY9XCJodHRwOi8vYXdvay5jb21cIlxuICAgICAgICAgICB0YXJnZXQ9XCJfYmxhbmtcIlxuICAgICAgICAgICByZWw9XCJub3JlZmVycmVyIG5vb3BlbmVyXCI+QXdvay5jb208L2E+XG4gICAgICA8L3A+XG4gICAgICA8cCBjbGFzc05hbWU9XCJsZWFkXCI+PGkgY2xhc3NOYW1lPVwiZmEgZmEtaGVhcnQgdGV4dC1kYW5nZXJcIiAvPntgeyBQSFAsIEphdmFTY3JpcHQsIFB5dGhvbiwgTXlTUUwsIE1vbmdvREIgfWB9PC9wPlxuICAgIDwvZGl2PlxuICA8L2hlYWRlcj5cbn1cbkhlYWRlci5kaXNwbGF5TmFtZSA9IGRpc3BsYXlOYW1lXG5cbmV4cG9ydCBkZWZhdWx0IEhlYWRlclxuIl0sInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./resources/js/modules/web/pages/home/components/Header.js\n");
+
+/***/ }),
+
+/***/ "./resources/js/modules/web/pages/home/index.js":
+/*!******************************************************!*\
+  !*** ./resources/js/modules/web/pages/home/index.js ***!
+  \******************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react_redux__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n/* harmony import */ var _Page__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./Page */ \"./resources/js/modules/web/pages/home/Page.js\");\n// import libs\n // import components\n\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Object(react_redux__WEBPACK_IMPORTED_MODULE_0__[\"connect\"])()(_Page__WEBPACK_IMPORTED_MODULE_1__[\"default\"]));//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy93ZWIvcGFnZXMvaG9tZS9pbmRleC5qcz8xMGRmIl0sIm5hbWVzIjpbImNvbm5lY3QiLCJQYWdlIl0sIm1hcHBpbmdzIjoiQUFBQTtBQUFBO0FBQUE7QUFBQTtDQUdBOztBQUNBO0FBRWVBLDBIQUFPLEdBQUdDLDZDQUFILENBQXRCIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvd2ViL3BhZ2VzL2hvbWUvaW5kZXguanMuanMiLCJzb3VyY2VzQ29udGVudCI6WyIvLyBpbXBvcnQgbGlic1xuaW1wb3J0IHsgY29ubmVjdCB9IGZyb20gXCJyZWFjdC1yZWR1eFwiXG5cbi8vIGltcG9ydCBjb21wb25lbnRzXG5pbXBvcnQgUGFnZSBmcm9tIFwiLi9QYWdlXCJcblxuZXhwb3J0IGRlZmF1bHQgY29ubmVjdCgpKFBhZ2UpXG4iXSwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./resources/js/modules/web/pages/home/index.js\n");
 
 /***/ })
 
