@@ -1,14 +1,37 @@
 (window["webpackJsonp"] = window["webpackJsonp"] || []).push([[3],{
 
-/***/ "./resources/js/common/wysiwyg-editor/index.js":
-/*!*****************************************************!*\
-  !*** ./resources/js/common/wysiwyg-editor/index.js ***!
-  \*****************************************************/
-/*! exports provided: default */
-/***/ (function(module, __webpack_exports__, __webpack_require__) {
+/***/ "./node_modules/react-document-title/index.js":
+/*!****************************************************!*\
+  !*** ./node_modules/react-document-title/index.js ***!
+  \****************************************************/
+/*! no static exports found */
+/***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\");\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var draft_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! draft-js */ \"./node_modules/draft-js/lib/Draft.js\");\n/* harmony import */ var draft_js__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(draft_js__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var react_draft_wysiwyg__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! react-draft-wysiwyg */ \"./node_modules/react-draft-wysiwyg/dist/react-draft-wysiwyg.js\");\n/* harmony import */ var react_draft_wysiwyg__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(react_draft_wysiwyg__WEBPACK_IMPORTED_MODULE_3__);\n/* harmony import */ var draftjs_to_html__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! draftjs-to-html */ \"./node_modules/draftjs-to-html/lib/draftjs-to-html.js\");\n/* harmony import */ var draftjs_to_html__WEBPACK_IMPORTED_MODULE_4___default = /*#__PURE__*/__webpack_require__.n(draftjs_to_html__WEBPACK_IMPORTED_MODULE_4__);\n/* harmony import */ var react_draft_wysiwyg_dist_react_draft_wysiwyg_css__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! react-draft-wysiwyg/dist/react-draft-wysiwyg.css */ \"./node_modules/react-draft-wysiwyg/dist/react-draft-wysiwyg.css\");\n/* harmony import */ var react_draft_wysiwyg_dist_react_draft_wysiwyg_css__WEBPACK_IMPORTED_MODULE_5___default = /*#__PURE__*/__webpack_require__.n(react_draft_wysiwyg_dist_react_draft_wysiwyg_css__WEBPACK_IMPORTED_MODULE_5__);\nfunction _typeof(obj) { if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\n\n\n\n\n\n\n\nvar WYSIWYG =\n/*#__PURE__*/\nfunction (_Component) {\n  _inherits(WYSIWYG, _Component);\n\n  function WYSIWYG(props) {\n    var _this;\n\n    _classCallCheck(this, WYSIWYG);\n\n    _this = _possibleConstructorReturn(this, _getPrototypeOf(WYSIWYG).call(this, props));\n    _this.state = {\n      editorState: _this.convertHtmlToEditorState(_this.props.value)\n    };\n    _this.onEditorStateChange = _this.onEditorStateChange.bind(_assertThisInitialized(_this));\n    return _this;\n  }\n\n  _createClass(WYSIWYG, [{\n    key: \"convertHtmlToEditorState\",\n    value: function convertHtmlToEditorState(value) {\n      if (value) {\n        var blocksFromHTML = Object(draft_js__WEBPACK_IMPORTED_MODULE_2__[\"convertFromHTML\"])(value);\n        var state = draft_js__WEBPACK_IMPORTED_MODULE_2__[\"ContentState\"].createFromBlockArray(blocksFromHTML.contentBlocks, blocksFromHTML.entityMap);\n        return draft_js__WEBPACK_IMPORTED_MODULE_2__[\"EditorState\"].createWithContent(state);\n      }\n    }\n  }, {\n    key: \"onEditorStateChange\",\n    value: function onEditorStateChange(editorState) {\n      this.setState({\n        editorState: editorState\n      });\n      this.props.onChange(draftjs_to_html__WEBPACK_IMPORTED_MODULE_4___default()(Object(draft_js__WEBPACK_IMPORTED_MODULE_2__[\"convertToRaw\"])(editorState.getCurrentContent())));\n    }\n  }, {\n    key: \"render\",\n    value: function render() {\n      var editorState = this.state.editorState;\n      return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", null, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(react_draft_wysiwyg__WEBPACK_IMPORTED_MODULE_3__[\"Editor\"], {\n        editorState: editorState,\n        wrapperClassName: \"demo-wrapper\",\n        editorClassName: \"form-control\",\n        onEditorStateChange: this.onEditorStateChange\n      }));\n    }\n  }]);\n\n  return WYSIWYG;\n}(react__WEBPACK_IMPORTED_MODULE_0__[\"Component\"]);\n\n_defineProperty(WYSIWYG, \"propTypes\", {\n  value: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.string.isRequired,\n  onChange: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.func.isRequired\n});\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (WYSIWYG);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvY29tbW9uL3d5c2l3eWctZWRpdG9yL2luZGV4LmpzPzNjM2UiXSwibmFtZXMiOlsiV1lTSVdZRyIsInByb3BzIiwic3RhdGUiLCJlZGl0b3JTdGF0ZSIsImNvbnZlcnRIdG1sVG9FZGl0b3JTdGF0ZSIsInZhbHVlIiwib25FZGl0b3JTdGF0ZUNoYW5nZSIsImJpbmQiLCJibG9ja3NGcm9tSFRNTCIsImNvbnZlcnRGcm9tSFRNTCIsIkNvbnRlbnRTdGF0ZSIsImNyZWF0ZUZyb21CbG9ja0FycmF5IiwiY29udGVudEJsb2NrcyIsImVudGl0eU1hcCIsIkVkaXRvclN0YXRlIiwiY3JlYXRlV2l0aENvbnRlbnQiLCJzZXRTdGF0ZSIsIm9uQ2hhbmdlIiwiZHJhZnRUb0h0bWwiLCJjb252ZXJ0VG9SYXciLCJnZXRDdXJyZW50Q29udGVudCIsIkNvbXBvbmVudCIsIlByb3BUeXBlcyIsInN0cmluZyIsImlzUmVxdWlyZWQiLCJmdW5jIl0sIm1hcHBpbmdzIjoiOzs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7QUFBQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBRUE7O0lBRU1BLE87Ozs7O0FBTUosbUJBQVlDLEtBQVosRUFBbUI7QUFBQTs7QUFBQTs7QUFDakIsaUZBQU1BLEtBQU47QUFFQSxVQUFLQyxLQUFMLEdBQWE7QUFDWEMsaUJBQVcsRUFBRSxNQUFLQyx3QkFBTCxDQUE4QixNQUFLSCxLQUFMLENBQVdJLEtBQXpDO0FBREYsS0FBYjtBQUlBLFVBQUtDLG1CQUFMLEdBQTJCLE1BQUtBLG1CQUFMLENBQXlCQyxJQUF6QiwrQkFBM0I7QUFQaUI7QUFRbEI7Ozs7NkNBRXdCRixLLEVBQU87QUFDOUIsVUFBSUEsS0FBSixFQUFVO0FBQ1IsWUFBTUcsY0FBYyxHQUFHQyxnRUFBZSxDQUFDSixLQUFELENBQXRDO0FBQ0EsWUFBTUgsS0FBSyxHQUFHUSxxREFBWSxDQUFDQyxvQkFBYixDQUNaSCxjQUFjLENBQUNJLGFBREgsRUFFWkosY0FBYyxDQUFDSyxTQUZILENBQWQ7QUFLQSxlQUFPQyxvREFBVyxDQUFDQyxpQkFBWixDQUE4QmIsS0FBOUIsQ0FBUDtBQUNEO0FBQ0Y7Ozt3Q0FFbUJDLFcsRUFBYTtBQUMvQixXQUFLYSxRQUFMLENBQWM7QUFDWmIsbUJBQVcsRUFBWEE7QUFEWSxPQUFkO0FBSUEsV0FBS0YsS0FBTCxDQUFXZ0IsUUFBWCxDQUFvQkMsc0RBQVcsQ0FBQ0MsNkRBQVksQ0FBQ2hCLFdBQVcsQ0FBQ2lCLGlCQUFaLEVBQUQsQ0FBYixDQUEvQjtBQUNEOzs7NkJBRVE7QUFBQSxVQUNDakIsV0FERCxHQUNpQixLQUFLRCxLQUR0QixDQUNDQyxXQUREO0FBRVAsYUFDRSx3RUFDRSwyREFBQywwREFBRDtBQUNFLG1CQUFXLEVBQUVBLFdBRGY7QUFFRSx3QkFBZ0IsRUFBQyxjQUZuQjtBQUdFLHVCQUFlLEVBQUMsY0FIbEI7QUFJRSwyQkFBbUIsRUFBRSxLQUFLRztBQUo1QixRQURGLENBREY7QUFVRDs7OztFQWhEbUJlLCtDOztnQkFBaEJyQixPLGVBQ2U7QUFDakJLLE9BQUssRUFBRWlCLGlEQUFTLENBQUNDLE1BQVYsQ0FBaUJDLFVBRFA7QUFFakJQLFVBQVEsRUFBRUssaURBQVMsQ0FBQ0csSUFBVixDQUFlRDtBQUZSLEM7O0FBa0ROeEIsc0VBQWYiLCJmaWxlIjoiLi9yZXNvdXJjZXMvanMvY29tbW9uL3d5c2l3eWctZWRpdG9yL2luZGV4LmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiaW1wb3J0IFJlYWN0LCB7IENvbXBvbmVudCB9IGZyb20gJ3JlYWN0J1xuaW1wb3J0IFByb3BUeXBlcyBmcm9tICdwcm9wLXR5cGVzJ1xuaW1wb3J0IHsgRWRpdG9yU3RhdGUsIGNvbnZlcnRUb1JhdywgQ29udGVudFN0YXRlLCBjb252ZXJ0RnJvbUhUTUwgfSBmcm9tICdkcmFmdC1qcydcbmltcG9ydCB7IEVkaXRvciB9IGZyb20gJ3JlYWN0LWRyYWZ0LXd5c2l3eWcnXG5pbXBvcnQgZHJhZnRUb0h0bWwgZnJvbSAnZHJhZnRqcy10by1odG1sJ1xuXG5pbXBvcnQgJ3JlYWN0LWRyYWZ0LXd5c2l3eWcvZGlzdC9yZWFjdC1kcmFmdC13eXNpd3lnLmNzcydcblxuY2xhc3MgV1lTSVdZRyBleHRlbmRzIENvbXBvbmVudCB7XG4gIHN0YXRpYyBwcm9wVHlwZXMgPSB7XG4gICAgdmFsdWU6IFByb3BUeXBlcy5zdHJpbmcuaXNSZXF1aXJlZCxcbiAgICBvbkNoYW5nZTogUHJvcFR5cGVzLmZ1bmMuaXNSZXF1aXJlZCxcbiAgfVxuXG4gIGNvbnN0cnVjdG9yKHByb3BzKSB7XG4gICAgc3VwZXIocHJvcHMpXG5cbiAgICB0aGlzLnN0YXRlID0ge1xuICAgICAgZWRpdG9yU3RhdGU6IHRoaXMuY29udmVydEh0bWxUb0VkaXRvclN0YXRlKHRoaXMucHJvcHMudmFsdWUpLFxuICAgIH1cblxuICAgIHRoaXMub25FZGl0b3JTdGF0ZUNoYW5nZSA9IHRoaXMub25FZGl0b3JTdGF0ZUNoYW5nZS5iaW5kKHRoaXMpXG4gIH1cblxuICBjb252ZXJ0SHRtbFRvRWRpdG9yU3RhdGUodmFsdWUpIHtcbiAgICBpZiAodmFsdWUpe1xuICAgICAgY29uc3QgYmxvY2tzRnJvbUhUTUwgPSBjb252ZXJ0RnJvbUhUTUwodmFsdWUpXG4gICAgICBjb25zdCBzdGF0ZSA9IENvbnRlbnRTdGF0ZS5jcmVhdGVGcm9tQmxvY2tBcnJheShcbiAgICAgICAgYmxvY2tzRnJvbUhUTUwuY29udGVudEJsb2NrcyxcbiAgICAgICAgYmxvY2tzRnJvbUhUTUwuZW50aXR5TWFwXG4gICAgICApXG5cbiAgICAgIHJldHVybiBFZGl0b3JTdGF0ZS5jcmVhdGVXaXRoQ29udGVudChzdGF0ZSlcbiAgICB9XG4gIH1cblxuICBvbkVkaXRvclN0YXRlQ2hhbmdlKGVkaXRvclN0YXRlKSB7XG4gICAgdGhpcy5zZXRTdGF0ZSh7XG4gICAgICBlZGl0b3JTdGF0ZSxcbiAgICB9KVxuXG4gICAgdGhpcy5wcm9wcy5vbkNoYW5nZShkcmFmdFRvSHRtbChjb252ZXJ0VG9SYXcoZWRpdG9yU3RhdGUuZ2V0Q3VycmVudENvbnRlbnQoKSkpKVxuICB9XG5cbiAgcmVuZGVyKCkge1xuICAgIGNvbnN0IHsgZWRpdG9yU3RhdGUgfSA9IHRoaXMuc3RhdGVcbiAgICByZXR1cm4gKFxuICAgICAgPGRpdj5cbiAgICAgICAgPEVkaXRvclxuICAgICAgICAgIGVkaXRvclN0YXRlPXtlZGl0b3JTdGF0ZX1cbiAgICAgICAgICB3cmFwcGVyQ2xhc3NOYW1lPVwiZGVtby13cmFwcGVyXCJcbiAgICAgICAgICBlZGl0b3JDbGFzc05hbWU9XCJmb3JtLWNvbnRyb2xcIlxuICAgICAgICAgIG9uRWRpdG9yU3RhdGVDaGFuZ2U9e3RoaXMub25FZGl0b3JTdGF0ZUNoYW5nZX1cbiAgICAgICAgLz5cbiAgICAgIDwvZGl2PlxuICAgIClcbiAgfVxufVxuXG5leHBvcnQgZGVmYXVsdCBXWVNJV1lHXG4iXSwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./resources/js/common/wysiwyg-editor/index.js\n");
+eval("\n\nvar React = __webpack_require__(/*! react */ \"./node_modules/react/index.js\"),\n    PropTypes = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\"),\n    withSideEffect = __webpack_require__(/*! react-side-effect */ \"./node_modules/react-side-effect/lib/index.js\");\n\nfunction reducePropsToState(propsList) {\n  var innermostProps = propsList[propsList.length - 1];\n  if (innermostProps) {\n    return innermostProps.title;\n  }\n}\n\nfunction handleStateChangeOnClient(title) {\n  var nextTitle = title || '';\n  if (nextTitle !== document.title) {\n    document.title = nextTitle;\n  }\n}\n\nfunction DocumentTitle() {}\nDocumentTitle.prototype = Object.create(React.Component.prototype);\n\nDocumentTitle.displayName = 'DocumentTitle';\nDocumentTitle.propTypes = {\n  title: PropTypes.string.isRequired\n};\n\nDocumentTitle.prototype.render = function() {\n  if (this.props.children) {\n    return React.Children.only(this.props.children);\n  } else {\n    return null;\n  }\n};\n\nmodule.exports = withSideEffect(\n  reducePropsToState,\n  handleStateChangeOnClient\n)(DocumentTitle);\n//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9ub2RlX21vZHVsZXMvcmVhY3QtZG9jdW1lbnQtdGl0bGUvaW5kZXguanM/NjQ1YyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiQUFBYTs7QUFFYixZQUFZLG1CQUFPLENBQUMsNENBQU87QUFDM0IsZ0JBQWdCLG1CQUFPLENBQUMsc0RBQVk7QUFDcEMscUJBQXFCLG1CQUFPLENBQUMsd0VBQW1COztBQUVoRDtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7O0FBRUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBOztBQUVBO0FBQ0E7O0FBRUE7QUFDQTtBQUNBO0FBQ0E7O0FBRUE7QUFDQTtBQUNBO0FBQ0EsR0FBRztBQUNIO0FBQ0E7QUFDQTs7QUFFQTtBQUNBO0FBQ0E7QUFDQSIsImZpbGUiOiIuL25vZGVfbW9kdWxlcy9yZWFjdC1kb2N1bWVudC10aXRsZS9pbmRleC5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbIid1c2Ugc3RyaWN0JztcblxudmFyIFJlYWN0ID0gcmVxdWlyZSgncmVhY3QnKSxcbiAgICBQcm9wVHlwZXMgPSByZXF1aXJlKCdwcm9wLXR5cGVzJyksXG4gICAgd2l0aFNpZGVFZmZlY3QgPSByZXF1aXJlKCdyZWFjdC1zaWRlLWVmZmVjdCcpO1xuXG5mdW5jdGlvbiByZWR1Y2VQcm9wc1RvU3RhdGUocHJvcHNMaXN0KSB7XG4gIHZhciBpbm5lcm1vc3RQcm9wcyA9IHByb3BzTGlzdFtwcm9wc0xpc3QubGVuZ3RoIC0gMV07XG4gIGlmIChpbm5lcm1vc3RQcm9wcykge1xuICAgIHJldHVybiBpbm5lcm1vc3RQcm9wcy50aXRsZTtcbiAgfVxufVxuXG5mdW5jdGlvbiBoYW5kbGVTdGF0ZUNoYW5nZU9uQ2xpZW50KHRpdGxlKSB7XG4gIHZhciBuZXh0VGl0bGUgPSB0aXRsZSB8fCAnJztcbiAgaWYgKG5leHRUaXRsZSAhPT0gZG9jdW1lbnQudGl0bGUpIHtcbiAgICBkb2N1bWVudC50aXRsZSA9IG5leHRUaXRsZTtcbiAgfVxufVxuXG5mdW5jdGlvbiBEb2N1bWVudFRpdGxlKCkge31cbkRvY3VtZW50VGl0bGUucHJvdG90eXBlID0gT2JqZWN0LmNyZWF0ZShSZWFjdC5Db21wb25lbnQucHJvdG90eXBlKTtcblxuRG9jdW1lbnRUaXRsZS5kaXNwbGF5TmFtZSA9ICdEb2N1bWVudFRpdGxlJztcbkRvY3VtZW50VGl0bGUucHJvcFR5cGVzID0ge1xuICB0aXRsZTogUHJvcFR5cGVzLnN0cmluZy5pc1JlcXVpcmVkXG59O1xuXG5Eb2N1bWVudFRpdGxlLnByb3RvdHlwZS5yZW5kZXIgPSBmdW5jdGlvbigpIHtcbiAgaWYgKHRoaXMucHJvcHMuY2hpbGRyZW4pIHtcbiAgICByZXR1cm4gUmVhY3QuQ2hpbGRyZW4ub25seSh0aGlzLnByb3BzLmNoaWxkcmVuKTtcbiAgfSBlbHNlIHtcbiAgICByZXR1cm4gbnVsbDtcbiAgfVxufTtcblxubW9kdWxlLmV4cG9ydHMgPSB3aXRoU2lkZUVmZmVjdChcbiAgcmVkdWNlUHJvcHNUb1N0YXRlLFxuICBoYW5kbGVTdGF0ZUNoYW5nZU9uQ2xpZW50XG4pKERvY3VtZW50VGl0bGUpO1xuIl0sInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./node_modules/react-document-title/index.js\n");
+
+/***/ }),
+
+/***/ "./node_modules/react-side-effect/lib/index.js":
+/*!*****************************************************!*\
+  !*** ./node_modules/react-side-effect/lib/index.js ***!
+  \*****************************************************/
+/*! no static exports found */
+/***/ (function(module, exports, __webpack_require__) {
+
+"use strict";
+eval("\n\nfunction _interopDefault (ex) { return (ex && (typeof ex === 'object') && 'default' in ex) ? ex['default'] : ex; }\n\nvar React = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\nvar React__default = _interopDefault(React);\nvar shallowEqual = _interopDefault(__webpack_require__(/*! shallowequal */ \"./node_modules/shallowequal/index.js\"));\n\nfunction _defineProperty(obj, key, value) {\n  if (key in obj) {\n    Object.defineProperty(obj, key, {\n      value: value,\n      enumerable: true,\n      configurable: true,\n      writable: true\n    });\n  } else {\n    obj[key] = value;\n  }\n\n  return obj;\n}\n\nfunction _inheritsLoose(subClass, superClass) {\n  subClass.prototype = Object.create(superClass.prototype);\n  subClass.prototype.constructor = subClass;\n  subClass.__proto__ = superClass;\n}\n\nvar canUseDOM = !!(typeof window !== 'undefined' && window.document && window.document.createElement);\nfunction withSideEffect(reducePropsToState, handleStateChangeOnClient, mapStateOnServer) {\n  if (typeof reducePropsToState !== 'function') {\n    throw new Error('Expected reducePropsToState to be a function.');\n  }\n\n  if (typeof handleStateChangeOnClient !== 'function') {\n    throw new Error('Expected handleStateChangeOnClient to be a function.');\n  }\n\n  if (typeof mapStateOnServer !== 'undefined' && typeof mapStateOnServer !== 'function') {\n    throw new Error('Expected mapStateOnServer to either be undefined or a function.');\n  }\n\n  function getDisplayName(WrappedComponent) {\n    return WrappedComponent.displayName || WrappedComponent.name || 'Component';\n  }\n\n  return function wrap(WrappedComponent) {\n    if (typeof WrappedComponent !== 'function') {\n      throw new Error('Expected WrappedComponent to be a React component.');\n    }\n\n    var mountedInstances = [];\n    var state;\n\n    function emitChange() {\n      state = reducePropsToState(mountedInstances.map(function (instance) {\n        return instance.props;\n      }));\n\n      if (SideEffect.canUseDOM) {\n        handleStateChangeOnClient(state);\n      } else if (mapStateOnServer) {\n        state = mapStateOnServer(state);\n      }\n    }\n\n    var SideEffect =\n    /*#__PURE__*/\n    function (_Component) {\n      _inheritsLoose(SideEffect, _Component);\n\n      function SideEffect() {\n        return _Component.apply(this, arguments) || this;\n      }\n\n      // Try to use displayName of wrapped component\n      // Expose canUseDOM so tests can monkeypatch it\n      SideEffect.peek = function peek() {\n        return state;\n      };\n\n      SideEffect.rewind = function rewind() {\n        if (SideEffect.canUseDOM) {\n          throw new Error('You may only call rewind() on the server. Call peek() to read the current state.');\n        }\n\n        var recordedState = state;\n        state = undefined;\n        mountedInstances = [];\n        return recordedState;\n      };\n\n      var _proto = SideEffect.prototype;\n\n      _proto.shouldComponentUpdate = function shouldComponentUpdate(nextProps) {\n        return !shallowEqual(nextProps, this.props);\n      };\n\n      _proto.componentWillMount = function componentWillMount() {\n        mountedInstances.push(this);\n        emitChange();\n      };\n\n      _proto.componentDidUpdate = function componentDidUpdate() {\n        emitChange();\n      };\n\n      _proto.componentWillUnmount = function componentWillUnmount() {\n        var index = mountedInstances.indexOf(this);\n        mountedInstances.splice(index, 1);\n        emitChange();\n      };\n\n      _proto.render = function render() {\n        return React__default.createElement(WrappedComponent, this.props);\n      };\n\n      return SideEffect;\n    }(React.Component);\n\n    _defineProperty(SideEffect, \"displayName\", \"SideEffect(\" + getDisplayName(WrappedComponent) + \")\");\n\n    _defineProperty(SideEffect, \"canUseDOM\", canUseDOM);\n\n    return SideEffect;\n  };\n}\n\nmodule.exports = withSideEffect;\n//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9ub2RlX21vZHVsZXMvcmVhY3Qtc2lkZS1lZmZlY3QvbGliL2luZGV4LmpzP2YzZWIiXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IkFBQWE7O0FBRWIsK0JBQStCLGlGQUFpRjs7QUFFaEgsWUFBWSxtQkFBTyxDQUFDLDRDQUFPO0FBQzNCO0FBQ0EsbUNBQW1DLG1CQUFPLENBQUMsMERBQWM7O0FBRXpEO0FBQ0E7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBO0FBQ0EsS0FBSztBQUNMLEdBQUc7QUFDSDtBQUNBOztBQUVBO0FBQ0E7O0FBRUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTs7QUFFQTtBQUNBO0FBQ0E7QUFDQTtBQUNBOztBQUVBO0FBQ0E7QUFDQTs7QUFFQTtBQUNBO0FBQ0E7O0FBRUE7QUFDQTtBQUNBOztBQUVBO0FBQ0E7QUFDQTtBQUNBOztBQUVBO0FBQ0E7O0FBRUE7QUFDQTtBQUNBO0FBQ0EsT0FBTzs7QUFFUDtBQUNBO0FBQ0EsT0FBTztBQUNQO0FBQ0E7QUFDQTs7QUFFQTtBQUNBO0FBQ0E7QUFDQTs7QUFFQTtBQUNBO0FBQ0E7O0FBRUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTs7QUFFQTtBQUNBO0FBQ0E7QUFDQTs7QUFFQTtBQUNBO0FBQ0E7QUFDQTtBQUNBOztBQUVBOztBQUVBO0FBQ0E7QUFDQTs7QUFFQTtBQUNBO0FBQ0E7QUFDQTs7QUFFQTtBQUNBO0FBQ0E7O0FBRUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTs7QUFFQTtBQUNBO0FBQ0E7O0FBRUE7QUFDQSxLQUFLOztBQUVMOztBQUVBOztBQUVBO0FBQ0E7QUFDQTs7QUFFQSIsImZpbGUiOiIuL25vZGVfbW9kdWxlcy9yZWFjdC1zaWRlLWVmZmVjdC9saWIvaW5kZXguanMuanMiLCJzb3VyY2VzQ29udGVudCI6WyIndXNlIHN0cmljdCc7XG5cbmZ1bmN0aW9uIF9pbnRlcm9wRGVmYXVsdCAoZXgpIHsgcmV0dXJuIChleCAmJiAodHlwZW9mIGV4ID09PSAnb2JqZWN0JykgJiYgJ2RlZmF1bHQnIGluIGV4KSA/IGV4WydkZWZhdWx0J10gOiBleDsgfVxuXG52YXIgUmVhY3QgPSByZXF1aXJlKCdyZWFjdCcpO1xudmFyIFJlYWN0X19kZWZhdWx0ID0gX2ludGVyb3BEZWZhdWx0KFJlYWN0KTtcbnZhciBzaGFsbG93RXF1YWwgPSBfaW50ZXJvcERlZmF1bHQocmVxdWlyZSgnc2hhbGxvd2VxdWFsJykpO1xuXG5mdW5jdGlvbiBfZGVmaW5lUHJvcGVydHkob2JqLCBrZXksIHZhbHVlKSB7XG4gIGlmIChrZXkgaW4gb2JqKSB7XG4gICAgT2JqZWN0LmRlZmluZVByb3BlcnR5KG9iaiwga2V5LCB7XG4gICAgICB2YWx1ZTogdmFsdWUsXG4gICAgICBlbnVtZXJhYmxlOiB0cnVlLFxuICAgICAgY29uZmlndXJhYmxlOiB0cnVlLFxuICAgICAgd3JpdGFibGU6IHRydWVcbiAgICB9KTtcbiAgfSBlbHNlIHtcbiAgICBvYmpba2V5XSA9IHZhbHVlO1xuICB9XG5cbiAgcmV0dXJuIG9iajtcbn1cblxuZnVuY3Rpb24gX2luaGVyaXRzTG9vc2Uoc3ViQ2xhc3MsIHN1cGVyQ2xhc3MpIHtcbiAgc3ViQ2xhc3MucHJvdG90eXBlID0gT2JqZWN0LmNyZWF0ZShzdXBlckNsYXNzLnByb3RvdHlwZSk7XG4gIHN1YkNsYXNzLnByb3RvdHlwZS5jb25zdHJ1Y3RvciA9IHN1YkNsYXNzO1xuICBzdWJDbGFzcy5fX3Byb3RvX18gPSBzdXBlckNsYXNzO1xufVxuXG52YXIgY2FuVXNlRE9NID0gISEodHlwZW9mIHdpbmRvdyAhPT0gJ3VuZGVmaW5lZCcgJiYgd2luZG93LmRvY3VtZW50ICYmIHdpbmRvdy5kb2N1bWVudC5jcmVhdGVFbGVtZW50KTtcbmZ1bmN0aW9uIHdpdGhTaWRlRWZmZWN0KHJlZHVjZVByb3BzVG9TdGF0ZSwgaGFuZGxlU3RhdGVDaGFuZ2VPbkNsaWVudCwgbWFwU3RhdGVPblNlcnZlcikge1xuICBpZiAodHlwZW9mIHJlZHVjZVByb3BzVG9TdGF0ZSAhPT0gJ2Z1bmN0aW9uJykge1xuICAgIHRocm93IG5ldyBFcnJvcignRXhwZWN0ZWQgcmVkdWNlUHJvcHNUb1N0YXRlIHRvIGJlIGEgZnVuY3Rpb24uJyk7XG4gIH1cblxuICBpZiAodHlwZW9mIGhhbmRsZVN0YXRlQ2hhbmdlT25DbGllbnQgIT09ICdmdW5jdGlvbicpIHtcbiAgICB0aHJvdyBuZXcgRXJyb3IoJ0V4cGVjdGVkIGhhbmRsZVN0YXRlQ2hhbmdlT25DbGllbnQgdG8gYmUgYSBmdW5jdGlvbi4nKTtcbiAgfVxuXG4gIGlmICh0eXBlb2YgbWFwU3RhdGVPblNlcnZlciAhPT0gJ3VuZGVmaW5lZCcgJiYgdHlwZW9mIG1hcFN0YXRlT25TZXJ2ZXIgIT09ICdmdW5jdGlvbicpIHtcbiAgICB0aHJvdyBuZXcgRXJyb3IoJ0V4cGVjdGVkIG1hcFN0YXRlT25TZXJ2ZXIgdG8gZWl0aGVyIGJlIHVuZGVmaW5lZCBvciBhIGZ1bmN0aW9uLicpO1xuICB9XG5cbiAgZnVuY3Rpb24gZ2V0RGlzcGxheU5hbWUoV3JhcHBlZENvbXBvbmVudCkge1xuICAgIHJldHVybiBXcmFwcGVkQ29tcG9uZW50LmRpc3BsYXlOYW1lIHx8IFdyYXBwZWRDb21wb25lbnQubmFtZSB8fCAnQ29tcG9uZW50JztcbiAgfVxuXG4gIHJldHVybiBmdW5jdGlvbiB3cmFwKFdyYXBwZWRDb21wb25lbnQpIHtcbiAgICBpZiAodHlwZW9mIFdyYXBwZWRDb21wb25lbnQgIT09ICdmdW5jdGlvbicpIHtcbiAgICAgIHRocm93IG5ldyBFcnJvcignRXhwZWN0ZWQgV3JhcHBlZENvbXBvbmVudCB0byBiZSBhIFJlYWN0IGNvbXBvbmVudC4nKTtcbiAgICB9XG5cbiAgICB2YXIgbW91bnRlZEluc3RhbmNlcyA9IFtdO1xuICAgIHZhciBzdGF0ZTtcblxuICAgIGZ1bmN0aW9uIGVtaXRDaGFuZ2UoKSB7XG4gICAgICBzdGF0ZSA9IHJlZHVjZVByb3BzVG9TdGF0ZShtb3VudGVkSW5zdGFuY2VzLm1hcChmdW5jdGlvbiAoaW5zdGFuY2UpIHtcbiAgICAgICAgcmV0dXJuIGluc3RhbmNlLnByb3BzO1xuICAgICAgfSkpO1xuXG4gICAgICBpZiAoU2lkZUVmZmVjdC5jYW5Vc2VET00pIHtcbiAgICAgICAgaGFuZGxlU3RhdGVDaGFuZ2VPbkNsaWVudChzdGF0ZSk7XG4gICAgICB9IGVsc2UgaWYgKG1hcFN0YXRlT25TZXJ2ZXIpIHtcbiAgICAgICAgc3RhdGUgPSBtYXBTdGF0ZU9uU2VydmVyKHN0YXRlKTtcbiAgICAgIH1cbiAgICB9XG5cbiAgICB2YXIgU2lkZUVmZmVjdCA9XG4gICAgLyojX19QVVJFX18qL1xuICAgIGZ1bmN0aW9uIChfQ29tcG9uZW50KSB7XG4gICAgICBfaW5oZXJpdHNMb29zZShTaWRlRWZmZWN0LCBfQ29tcG9uZW50KTtcblxuICAgICAgZnVuY3Rpb24gU2lkZUVmZmVjdCgpIHtcbiAgICAgICAgcmV0dXJuIF9Db21wb25lbnQuYXBwbHkodGhpcywgYXJndW1lbnRzKSB8fCB0aGlzO1xuICAgICAgfVxuXG4gICAgICAvLyBUcnkgdG8gdXNlIGRpc3BsYXlOYW1lIG9mIHdyYXBwZWQgY29tcG9uZW50XG4gICAgICAvLyBFeHBvc2UgY2FuVXNlRE9NIHNvIHRlc3RzIGNhbiBtb25rZXlwYXRjaCBpdFxuICAgICAgU2lkZUVmZmVjdC5wZWVrID0gZnVuY3Rpb24gcGVlaygpIHtcbiAgICAgICAgcmV0dXJuIHN0YXRlO1xuICAgICAgfTtcblxuICAgICAgU2lkZUVmZmVjdC5yZXdpbmQgPSBmdW5jdGlvbiByZXdpbmQoKSB7XG4gICAgICAgIGlmIChTaWRlRWZmZWN0LmNhblVzZURPTSkge1xuICAgICAgICAgIHRocm93IG5ldyBFcnJvcignWW91IG1heSBvbmx5IGNhbGwgcmV3aW5kKCkgb24gdGhlIHNlcnZlci4gQ2FsbCBwZWVrKCkgdG8gcmVhZCB0aGUgY3VycmVudCBzdGF0ZS4nKTtcbiAgICAgICAgfVxuXG4gICAgICAgIHZhciByZWNvcmRlZFN0YXRlID0gc3RhdGU7XG4gICAgICAgIHN0YXRlID0gdW5kZWZpbmVkO1xuICAgICAgICBtb3VudGVkSW5zdGFuY2VzID0gW107XG4gICAgICAgIHJldHVybiByZWNvcmRlZFN0YXRlO1xuICAgICAgfTtcblxuICAgICAgdmFyIF9wcm90byA9IFNpZGVFZmZlY3QucHJvdG90eXBlO1xuXG4gICAgICBfcHJvdG8uc2hvdWxkQ29tcG9uZW50VXBkYXRlID0gZnVuY3Rpb24gc2hvdWxkQ29tcG9uZW50VXBkYXRlKG5leHRQcm9wcykge1xuICAgICAgICByZXR1cm4gIXNoYWxsb3dFcXVhbChuZXh0UHJvcHMsIHRoaXMucHJvcHMpO1xuICAgICAgfTtcblxuICAgICAgX3Byb3RvLmNvbXBvbmVudFdpbGxNb3VudCA9IGZ1bmN0aW9uIGNvbXBvbmVudFdpbGxNb3VudCgpIHtcbiAgICAgICAgbW91bnRlZEluc3RhbmNlcy5wdXNoKHRoaXMpO1xuICAgICAgICBlbWl0Q2hhbmdlKCk7XG4gICAgICB9O1xuXG4gICAgICBfcHJvdG8uY29tcG9uZW50RGlkVXBkYXRlID0gZnVuY3Rpb24gY29tcG9uZW50RGlkVXBkYXRlKCkge1xuICAgICAgICBlbWl0Q2hhbmdlKCk7XG4gICAgICB9O1xuXG4gICAgICBfcHJvdG8uY29tcG9uZW50V2lsbFVubW91bnQgPSBmdW5jdGlvbiBjb21wb25lbnRXaWxsVW5tb3VudCgpIHtcbiAgICAgICAgdmFyIGluZGV4ID0gbW91bnRlZEluc3RhbmNlcy5pbmRleE9mKHRoaXMpO1xuICAgICAgICBtb3VudGVkSW5zdGFuY2VzLnNwbGljZShpbmRleCwgMSk7XG4gICAgICAgIGVtaXRDaGFuZ2UoKTtcbiAgICAgIH07XG5cbiAgICAgIF9wcm90by5yZW5kZXIgPSBmdW5jdGlvbiByZW5kZXIoKSB7XG4gICAgICAgIHJldHVybiBSZWFjdF9fZGVmYXVsdC5jcmVhdGVFbGVtZW50KFdyYXBwZWRDb21wb25lbnQsIHRoaXMucHJvcHMpO1xuICAgICAgfTtcblxuICAgICAgcmV0dXJuIFNpZGVFZmZlY3Q7XG4gICAgfShSZWFjdC5Db21wb25lbnQpO1xuXG4gICAgX2RlZmluZVByb3BlcnR5KFNpZGVFZmZlY3QsIFwiZGlzcGxheU5hbWVcIiwgXCJTaWRlRWZmZWN0KFwiICsgZ2V0RGlzcGxheU5hbWUoV3JhcHBlZENvbXBvbmVudCkgKyBcIilcIik7XG5cbiAgICBfZGVmaW5lUHJvcGVydHkoU2lkZUVmZmVjdCwgXCJjYW5Vc2VET01cIiwgY2FuVXNlRE9NKTtcblxuICAgIHJldHVybiBTaWRlRWZmZWN0O1xuICB9O1xufVxuXG5tb2R1bGUuZXhwb3J0cyA9IHdpdGhTaWRlRWZmZWN0O1xuIl0sInNvdXJjZVJvb3QiOiIifQ==\n//# sourceURL=webpack-internal:///./node_modules/react-side-effect/lib/index.js\n");
+
+/***/ }),
+
+/***/ "./node_modules/shallowequal/index.js":
+/*!********************************************!*\
+  !*** ./node_modules/shallowequal/index.js ***!
+  \********************************************/
+/*! no static exports found */
+/***/ (function(module, exports) {
+
+eval("//\n\nmodule.exports = function shallowEqual(objA, objB, compare, compareContext) {\n  var ret = compare ? compare.call(compareContext, objA, objB) : void 0;\n\n  if (ret !== void 0) {\n    return !!ret;\n  }\n\n  if (objA === objB) {\n    return true;\n  }\n\n  if (typeof objA !== \"object\" || !objA || typeof objB !== \"object\" || !objB) {\n    return false;\n  }\n\n  var keysA = Object.keys(objA);\n  var keysB = Object.keys(objB);\n\n  if (keysA.length !== keysB.length) {\n    return false;\n  }\n\n  var bHasOwnProperty = Object.prototype.hasOwnProperty.bind(objB);\n\n  // Test for A's keys different from B.\n  for (var idx = 0; idx < keysA.length; idx++) {\n    var key = keysA[idx];\n\n    if (!bHasOwnProperty(key)) {\n      return false;\n    }\n\n    var valueA = objA[key];\n    var valueB = objB[key];\n\n    ret = compare ? compare.call(compareContext, valueA, valueB, key) : void 0;\n\n    if (ret === false || (ret === void 0 && valueA !== valueB)) {\n      return false;\n    }\n  }\n\n  return true;\n};\n//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9ub2RlX21vZHVsZXMvc2hhbGxvd2VxdWFsL2luZGV4LmpzPzFiMmIiXSwibmFtZXMiOltdLCJtYXBwaW5ncyI6IkFBQUE7O0FBRUE7QUFDQTs7QUFFQTtBQUNBO0FBQ0E7O0FBRUE7QUFDQTtBQUNBOztBQUVBO0FBQ0E7QUFDQTs7QUFFQTtBQUNBOztBQUVBO0FBQ0E7QUFDQTs7QUFFQTs7QUFFQTtBQUNBLG1CQUFtQixvQkFBb0I7QUFDdkM7O0FBRUE7QUFDQTtBQUNBOztBQUVBO0FBQ0E7O0FBRUE7O0FBRUE7QUFDQTtBQUNBO0FBQ0E7O0FBRUE7QUFDQSIsImZpbGUiOiIuL25vZGVfbW9kdWxlcy9zaGFsbG93ZXF1YWwvaW5kZXguanMuanMiLCJzb3VyY2VzQ29udGVudCI6WyIvL1xuXG5tb2R1bGUuZXhwb3J0cyA9IGZ1bmN0aW9uIHNoYWxsb3dFcXVhbChvYmpBLCBvYmpCLCBjb21wYXJlLCBjb21wYXJlQ29udGV4dCkge1xuICB2YXIgcmV0ID0gY29tcGFyZSA/IGNvbXBhcmUuY2FsbChjb21wYXJlQ29udGV4dCwgb2JqQSwgb2JqQikgOiB2b2lkIDA7XG5cbiAgaWYgKHJldCAhPT0gdm9pZCAwKSB7XG4gICAgcmV0dXJuICEhcmV0O1xuICB9XG5cbiAgaWYgKG9iakEgPT09IG9iakIpIHtcbiAgICByZXR1cm4gdHJ1ZTtcbiAgfVxuXG4gIGlmICh0eXBlb2Ygb2JqQSAhPT0gXCJvYmplY3RcIiB8fCAhb2JqQSB8fCB0eXBlb2Ygb2JqQiAhPT0gXCJvYmplY3RcIiB8fCAhb2JqQikge1xuICAgIHJldHVybiBmYWxzZTtcbiAgfVxuXG4gIHZhciBrZXlzQSA9IE9iamVjdC5rZXlzKG9iakEpO1xuICB2YXIga2V5c0IgPSBPYmplY3Qua2V5cyhvYmpCKTtcblxuICBpZiAoa2V5c0EubGVuZ3RoICE9PSBrZXlzQi5sZW5ndGgpIHtcbiAgICByZXR1cm4gZmFsc2U7XG4gIH1cblxuICB2YXIgYkhhc093blByb3BlcnR5ID0gT2JqZWN0LnByb3RvdHlwZS5oYXNPd25Qcm9wZXJ0eS5iaW5kKG9iakIpO1xuXG4gIC8vIFRlc3QgZm9yIEEncyBrZXlzIGRpZmZlcmVudCBmcm9tIEIuXG4gIGZvciAodmFyIGlkeCA9IDA7IGlkeCA8IGtleXNBLmxlbmd0aDsgaWR4KyspIHtcbiAgICB2YXIga2V5ID0ga2V5c0FbaWR4XTtcblxuICAgIGlmICghYkhhc093blByb3BlcnR5KGtleSkpIHtcbiAgICAgIHJldHVybiBmYWxzZTtcbiAgICB9XG5cbiAgICB2YXIgdmFsdWVBID0gb2JqQVtrZXldO1xuICAgIHZhciB2YWx1ZUIgPSBvYmpCW2tleV07XG5cbiAgICByZXQgPSBjb21wYXJlID8gY29tcGFyZS5jYWxsKGNvbXBhcmVDb250ZXh0LCB2YWx1ZUEsIHZhbHVlQiwga2V5KSA6IHZvaWQgMDtcblxuICAgIGlmIChyZXQgPT09IGZhbHNlIHx8IChyZXQgPT09IHZvaWQgMCAmJiB2YWx1ZUEgIT09IHZhbHVlQikpIHtcbiAgICAgIHJldHVybiBmYWxzZTtcbiAgICB9XG4gIH1cblxuICByZXR1cm4gdHJ1ZTtcbn07XG4iXSwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./node_modules/shallowequal/index.js\n");
 
 /***/ }),
 
@@ -20,7 +43,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var reac
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var moment__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! moment */ \"./node_modules/moment/moment.js\");\n/* harmony import */ var moment__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(moment__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _utils_Model__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/Model */ \"./resources/js/utils/Model.js\");\n/* harmony import */ var _user_User__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../user/User */ \"./resources/js/modules/user/User.js\");\nfunction _typeof(obj) { if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _get(target, property, receiver) { if (typeof Reflect !== \"undefined\" && Reflect.get) { _get = Reflect.get; } else { _get = function _get(target, property, receiver) { var base = _superPropBase(target, property); if (!base) return; var desc = Object.getOwnPropertyDescriptor(base, property); if (desc.get) { return desc.get.call(receiver); } return desc.value; }; } return _get(target, property, receiver || target); }\n\nfunction _superPropBase(object, property) { while (!Object.prototype.hasOwnProperty.call(object, property)) { object = _getPrototypeOf(object); if (object === null) break; } return object; }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\n\n\n\n\nvar Article =\n/*#__PURE__*/\nfunction (_Model) {\n  _inherits(Article, _Model);\n\n  function Article(props) {\n    var _this;\n\n    _classCallCheck(this, Article);\n\n    _this = _possibleConstructorReturn(this, _getPrototypeOf(Article).call(this, props));\n\n    _this.initialize(props);\n\n    return _this;\n  }\n\n  _createClass(Article, [{\n    key: \"initialize\",\n    value: function initialize(props) {\n      _get(_getPrototypeOf(Article.prototype), \"initialize\", this).call(this, props);\n\n      this.slug = props.slug || '';\n      this.title = props.title || '';\n      this.description = props.description || '';\n      this.content = props.content || '';\n      this.published = props.published || false;\n      this.publishedAt = props.publishedAt ? moment__WEBPACK_IMPORTED_MODULE_0___default()(props.publishedAt) : null; // relate user model\n\n      this.user = props.user ? new _user_User__WEBPACK_IMPORTED_MODULE_2__[\"default\"](props.user) : null;\n    }\n  }]);\n\n  return Article;\n}(_utils_Model__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Article);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL0FydGljbGUuanM/YjE5MSJdLCJuYW1lcyI6WyJBcnRpY2xlIiwicHJvcHMiLCJpbml0aWFsaXplIiwic2x1ZyIsInRpdGxlIiwiZGVzY3JpcHRpb24iLCJjb250ZW50IiwicHVibGlzaGVkIiwicHVibGlzaGVkQXQiLCJtb21lbnQiLCJ1c2VyIiwiVXNlciIsIk1vZGVsIl0sIm1hcHBpbmdzIjoiOzs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7Ozs7QUFBQTtBQUNBO0FBQ0E7O0lBRU1BLE87Ozs7O0FBQ0osbUJBQVlDLEtBQVosRUFBbUI7QUFBQTs7QUFBQTs7QUFDakIsaUZBQU1BLEtBQU47O0FBRUEsVUFBS0MsVUFBTCxDQUFnQkQsS0FBaEI7O0FBSGlCO0FBSWxCOzs7OytCQUVVQSxLLEVBQU87QUFDaEIsOEVBQWlCQSxLQUFqQjs7QUFFQSxXQUFLRSxJQUFMLEdBQVlGLEtBQUssQ0FBQ0UsSUFBTixJQUFjLEVBQTFCO0FBQ0EsV0FBS0MsS0FBTCxHQUFhSCxLQUFLLENBQUNHLEtBQU4sSUFBZSxFQUE1QjtBQUNBLFdBQUtDLFdBQUwsR0FBbUJKLEtBQUssQ0FBQ0ksV0FBTixJQUFxQixFQUF4QztBQUNBLFdBQUtDLE9BQUwsR0FBZUwsS0FBSyxDQUFDSyxPQUFOLElBQWlCLEVBQWhDO0FBQ0EsV0FBS0MsU0FBTCxHQUFpQk4sS0FBSyxDQUFDTSxTQUFOLElBQW1CLEtBQXBDO0FBQ0EsV0FBS0MsV0FBTCxHQUFtQlAsS0FBSyxDQUFDTyxXQUFOLEdBQW9CQyw2Q0FBTSxDQUFDUixLQUFLLENBQUNPLFdBQVAsQ0FBMUIsR0FBZ0QsSUFBbkUsQ0FSZ0IsQ0FVaEI7O0FBQ0EsV0FBS0UsSUFBTCxHQUFZVCxLQUFLLENBQUNTLElBQU4sR0FBYSxJQUFJQyxrREFBSixDQUFTVixLQUFLLENBQUNTLElBQWYsQ0FBYixHQUFvQyxJQUFoRDtBQUNEOzs7O0VBbkJtQkUsb0Q7O0FBc0JQWixzRUFBZiIsImZpbGUiOiIuL3Jlc291cmNlcy9qcy9tb2R1bGVzL2FydGljbGUvQXJ0aWNsZS5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBtb21lbnQgZnJvbSAnbW9tZW50J1xuaW1wb3J0IE1vZGVsIGZyb20gJy4uLy4uL3V0aWxzL01vZGVsJ1xuaW1wb3J0IFVzZXIgZnJvbSAnLi4vdXNlci9Vc2VyJ1xuXG5jbGFzcyBBcnRpY2xlIGV4dGVuZHMgTW9kZWwge1xuICBjb25zdHJ1Y3Rvcihwcm9wcykge1xuICAgIHN1cGVyKHByb3BzKVxuXG4gICAgdGhpcy5pbml0aWFsaXplKHByb3BzKVxuICB9XG5cbiAgaW5pdGlhbGl6ZShwcm9wcykge1xuICAgIHN1cGVyLmluaXRpYWxpemUocHJvcHMpXG5cbiAgICB0aGlzLnNsdWcgPSBwcm9wcy5zbHVnIHx8ICcnXG4gICAgdGhpcy50aXRsZSA9IHByb3BzLnRpdGxlIHx8ICcnXG4gICAgdGhpcy5kZXNjcmlwdGlvbiA9IHByb3BzLmRlc2NyaXB0aW9uIHx8ICcnXG4gICAgdGhpcy5jb250ZW50ID0gcHJvcHMuY29udGVudCB8fCAnJ1xuICAgIHRoaXMucHVibGlzaGVkID0gcHJvcHMucHVibGlzaGVkIHx8IGZhbHNlXG4gICAgdGhpcy5wdWJsaXNoZWRBdCA9IHByb3BzLnB1Ymxpc2hlZEF0ID8gbW9tZW50KHByb3BzLnB1Ymxpc2hlZEF0KSA6IG51bGxcblxuICAgIC8vIHJlbGF0ZSB1c2VyIG1vZGVsXG4gICAgdGhpcy51c2VyID0gcHJvcHMudXNlciA/IG5ldyBVc2VyKHByb3BzLnVzZXIpIDogbnVsbFxuICB9XG59XG5cbmV4cG9ydCBkZWZhdWx0IEFydGljbGVcbiJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/article/Article.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var moment__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! moment */ \"./node_modules/moment/moment.js\");\n/* harmony import */ var moment__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(moment__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var _utils_Model__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/Model */ \"./resources/js/utils/Model.js\");\n/* harmony import */ var _user_User__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../user/User */ \"./resources/js/modules/user/User.js\");\n\n\n\n\nclass Article extends _utils_Model__WEBPACK_IMPORTED_MODULE_1__[\"default\"] {\n  constructor(props) {\n    super(props);\n    this.initialize(props);\n  }\n\n  initialize(props) {\n    super.initialize(props);\n    this.slug = props.slug || '';\n    this.title = props.title || '';\n    this.description = props.description || '';\n    this.content = props.content || '';\n    this.published = props.published || false;\n    this.publishedAt = props.publishedAt ? moment__WEBPACK_IMPORTED_MODULE_0___default()(props.publishedAt) : null; // relate user model\n\n    this.user = props.user ? new _user_User__WEBPACK_IMPORTED_MODULE_2__[\"default\"](props.user) : null;\n  }\n\n}\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Article);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL0FydGljbGUuanM/YjE5MSJdLCJuYW1lcyI6WyJBcnRpY2xlIiwiTW9kZWwiLCJjb25zdHJ1Y3RvciIsInByb3BzIiwiaW5pdGlhbGl6ZSIsInNsdWciLCJ0aXRsZSIsImRlc2NyaXB0aW9uIiwiY29udGVudCIsInB1Ymxpc2hlZCIsInB1Ymxpc2hlZEF0IiwibW9tZW50IiwidXNlciIsIlVzZXIiXSwibWFwcGluZ3MiOiJBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUNBO0FBQ0E7O0FBRUEsTUFBTUEsT0FBTixTQUFzQkMsb0RBQXRCLENBQTRCO0FBQzFCQyxhQUFXLENBQUNDLEtBQUQsRUFBUTtBQUNqQixVQUFNQSxLQUFOO0FBRUEsU0FBS0MsVUFBTCxDQUFnQkQsS0FBaEI7QUFDRDs7QUFFREMsWUFBVSxDQUFDRCxLQUFELEVBQVE7QUFDaEIsVUFBTUMsVUFBTixDQUFpQkQsS0FBakI7QUFFQSxTQUFLRSxJQUFMLEdBQVlGLEtBQUssQ0FBQ0UsSUFBTixJQUFjLEVBQTFCO0FBQ0EsU0FBS0MsS0FBTCxHQUFhSCxLQUFLLENBQUNHLEtBQU4sSUFBZSxFQUE1QjtBQUNBLFNBQUtDLFdBQUwsR0FBbUJKLEtBQUssQ0FBQ0ksV0FBTixJQUFxQixFQUF4QztBQUNBLFNBQUtDLE9BQUwsR0FBZUwsS0FBSyxDQUFDSyxPQUFOLElBQWlCLEVBQWhDO0FBQ0EsU0FBS0MsU0FBTCxHQUFpQk4sS0FBSyxDQUFDTSxTQUFOLElBQW1CLEtBQXBDO0FBQ0EsU0FBS0MsV0FBTCxHQUFtQlAsS0FBSyxDQUFDTyxXQUFOLEdBQW9CQyw2Q0FBTSxDQUFDUixLQUFLLENBQUNPLFdBQVAsQ0FBMUIsR0FBZ0QsSUFBbkUsQ0FSZ0IsQ0FVaEI7O0FBQ0EsU0FBS0UsSUFBTCxHQUFZVCxLQUFLLENBQUNTLElBQU4sR0FBYSxJQUFJQyxrREFBSixDQUFTVixLQUFLLENBQUNTLElBQWYsQ0FBYixHQUFvQyxJQUFoRDtBQUNEOztBQW5CeUI7O0FBc0JiWixzRUFBZiIsImZpbGUiOiIuL3Jlc291cmNlcy9qcy9tb2R1bGVzL2FydGljbGUvQXJ0aWNsZS5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbImltcG9ydCBtb21lbnQgZnJvbSAnbW9tZW50J1xuaW1wb3J0IE1vZGVsIGZyb20gJy4uLy4uL3V0aWxzL01vZGVsJ1xuaW1wb3J0IFVzZXIgZnJvbSAnLi4vdXNlci9Vc2VyJ1xuXG5jbGFzcyBBcnRpY2xlIGV4dGVuZHMgTW9kZWwge1xuICBjb25zdHJ1Y3Rvcihwcm9wcykge1xuICAgIHN1cGVyKHByb3BzKVxuXG4gICAgdGhpcy5pbml0aWFsaXplKHByb3BzKVxuICB9XG5cbiAgaW5pdGlhbGl6ZShwcm9wcykge1xuICAgIHN1cGVyLmluaXRpYWxpemUocHJvcHMpXG5cbiAgICB0aGlzLnNsdWcgPSBwcm9wcy5zbHVnIHx8ICcnXG4gICAgdGhpcy50aXRsZSA9IHByb3BzLnRpdGxlIHx8ICcnXG4gICAgdGhpcy5kZXNjcmlwdGlvbiA9IHByb3BzLmRlc2NyaXB0aW9uIHx8ICcnXG4gICAgdGhpcy5jb250ZW50ID0gcHJvcHMuY29udGVudCB8fCAnJ1xuICAgIHRoaXMucHVibGlzaGVkID0gcHJvcHMucHVibGlzaGVkIHx8IGZhbHNlXG4gICAgdGhpcy5wdWJsaXNoZWRBdCA9IHByb3BzLnB1Ymxpc2hlZEF0ID8gbW9tZW50KHByb3BzLnB1Ymxpc2hlZEF0KSA6IG51bGxcblxuICAgIC8vIHJlbGF0ZSB1c2VyIG1vZGVsXG4gICAgdGhpcy51c2VyID0gcHJvcHMudXNlciA/IG5ldyBVc2VyKHByb3BzLnVzZXIpIDogbnVsbFxuICB9XG59XG5cbmV4cG9ydCBkZWZhdWx0IEFydGljbGVcbiJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/article/Article.js\n");
 
 /***/ }),
 
@@ -32,7 +55,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var mome
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleAddRequest\", function() { return articleAddRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleUpdateRequest\", function() { return articleUpdateRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleRemoveRequest\", function() { return articleRemoveRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleListRequest\", function() { return articleListRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleEditRequest\", function() { return articleEditRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleFetchRequest\", function() { return articleFetchRequest; });\n/* harmony import */ var _utils_Http__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/Http */ \"./resources/js/utils/Http.js\");\n/* harmony import */ var _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/Transformer */ \"./resources/js/utils/Transformer.js\");\n/* harmony import */ var _store_actions__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./store/actions */ \"./resources/js/modules/article/store/actions.js\");\n\n\n\n\nfunction transformRequest(parms) {\n  return _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].send(parms);\n}\n\nfunction transformResponse(params) {\n  return _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].fetch(params);\n}\n\nfunction articleAddRequest(params) {\n  return function (dispatch) {\n    return new Promise(function (resolve, reject) {\n      _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].post('/articles', transformRequest(params)).then(function (res) {\n        dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n        return resolve();\n      }).catch(function (err) {\n        var statusCode = err.response.status;\n        var data = {\n          error: null,\n          statusCode: statusCode\n        };\n\n        if (statusCode === 422) {\n          var resetErrors = {\n            errors: err.response.data,\n            replace: false,\n            searchStr: '',\n            replaceStr: ''\n          };\n          data.error = _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].resetValidationFields(resetErrors);\n        } else if (statusCode === 401) {\n          data.error = err.response.data.message;\n        }\n\n        return reject(data);\n      });\n    });\n  };\n}\nfunction articleUpdateRequest(params) {\n  return function (dispatch) {\n    return new Promise(function (resolve, reject) {\n      _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].patch(\"articles/\".concat(params.id), transformRequest(params)).then(function (res) {\n        dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n        return resolve();\n      }).catch(function (err) {\n        var statusCode = err.response.status;\n        var data = {\n          error: null,\n          statusCode: statusCode\n        };\n\n        if (statusCode === 422) {\n          var resetErrors = {\n            errors: err.response.data,\n            replace: false,\n            searchStr: '',\n            replaceStr: ''\n          };\n          data.error = _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].resetValidationFields(resetErrors);\n        } else if (statusCode === 401) {\n          data.error = err.response.data.message;\n        }\n\n        return reject(data);\n      });\n    });\n  };\n}\nfunction articleRemoveRequest(id) {\n  return function (dispatch) {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].delete(\"articles/\".concat(id)).then(function () {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"remove\"](id));\n    }).catch(function (err) {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleListRequest(_ref) {\n  var _ref$pageNumber = _ref.pageNumber,\n      pageNumber = _ref$pageNumber === void 0 ? 1 : _ref$pageNumber,\n      _ref$url = _ref.url,\n      url = _ref$url === void 0 ? '/articles' : _ref$url;\n  return function (dispatch) {\n    if (pageNumber > 1) {\n      url = url + \"?page=\".concat(pageNumber);\n    }\n\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(url).then(function (res) {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"list\"](transformResponse(res.data)));\n    }).catch(function (err) {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleEditRequest(id) {\n  return function (dispatch) {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(\"articles/\".concat(id)).then(function (res) {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n    }).catch(function (err) {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleFetchRequest(slug) {\n  return function (dispatch) {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(\"articles/published/\".concat(slug)).then(function (res) {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n    }).catch(function (err) {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL3NlcnZpY2UuanM/Y2YwNiJdLCJuYW1lcyI6WyJ0cmFuc2Zvcm1SZXF1ZXN0IiwicGFybXMiLCJUcmFuc2Zvcm1lciIsInNlbmQiLCJ0cmFuc2Zvcm1SZXNwb25zZSIsInBhcmFtcyIsImZldGNoIiwiYXJ0aWNsZUFkZFJlcXVlc3QiLCJkaXNwYXRjaCIsIlByb21pc2UiLCJyZXNvbHZlIiwicmVqZWN0IiwiSHR0cCIsInBvc3QiLCJ0aGVuIiwicmVzIiwiYXJ0aWNsZUFjdGlvbnMiLCJkYXRhIiwiY2F0Y2giLCJlcnIiLCJzdGF0dXNDb2RlIiwicmVzcG9uc2UiLCJzdGF0dXMiLCJlcnJvciIsInJlc2V0RXJyb3JzIiwiZXJyb3JzIiwicmVwbGFjZSIsInNlYXJjaFN0ciIsInJlcGxhY2VTdHIiLCJyZXNldFZhbGlkYXRpb25GaWVsZHMiLCJtZXNzYWdlIiwiYXJ0aWNsZVVwZGF0ZVJlcXVlc3QiLCJwYXRjaCIsImlkIiwiYXJ0aWNsZVJlbW92ZVJlcXVlc3QiLCJkZWxldGUiLCJjb25zb2xlIiwiYXJ0aWNsZUxpc3RSZXF1ZXN0IiwicGFnZU51bWJlciIsInVybCIsImdldCIsImFydGljbGVFZGl0UmVxdWVzdCIsImFydGljbGVGZXRjaFJlcXVlc3QiLCJzbHVnIl0sIm1hcHBpbmdzIjoiQUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQ0E7QUFDQTs7QUFFQSxTQUFTQSxnQkFBVCxDQUEwQkMsS0FBMUIsRUFBaUM7QUFDL0IsU0FBT0MsMERBQVcsQ0FBQ0MsSUFBWixDQUFpQkYsS0FBakIsQ0FBUDtBQUNEOztBQUVELFNBQVNHLGlCQUFULENBQTJCQyxNQUEzQixFQUFtQztBQUNqQyxTQUFPSCwwREFBVyxDQUFDSSxLQUFaLENBQWtCRCxNQUFsQixDQUFQO0FBQ0Q7O0FBRU0sU0FBU0UsaUJBQVQsQ0FBMkJGLE1BQTNCLEVBQW1DO0FBQ3hDLFNBQU8sVUFBQUcsUUFBUTtBQUFBLFdBQ2IsSUFBSUMsT0FBSixDQUFZLFVBQUNDLE9BQUQsRUFBVUMsTUFBVixFQUFxQjtBQUMvQkMseURBQUksQ0FBQ0MsSUFBTCxDQUFVLFdBQVYsRUFBdUJiLGdCQUFnQixDQUFDSyxNQUFELENBQXZDLEVBQ0dTLElBREgsQ0FDUSxVQUFBQyxHQUFHLEVBQUk7QUFDWFAsZ0JBQVEsQ0FBQ1Esa0RBQUEsQ0FBbUJaLGlCQUFpQixDQUFDVyxHQUFHLENBQUNFLElBQUwsQ0FBcEMsQ0FBRCxDQUFSO0FBQ0EsZUFBT1AsT0FBTyxFQUFkO0FBQ0QsT0FKSCxFQUtHUSxLQUxILENBS1MsVUFBQ0MsR0FBRCxFQUFTO0FBQ2QsWUFBTUMsVUFBVSxHQUFHRCxHQUFHLENBQUNFLFFBQUosQ0FBYUMsTUFBaEM7QUFDQSxZQUFNTCxJQUFJLEdBQUc7QUFDWE0sZUFBSyxFQUFFLElBREk7QUFFWEgsb0JBQVUsRUFBVkE7QUFGVyxTQUFiOztBQUtBLFlBQUlBLFVBQVUsS0FBSyxHQUFuQixFQUF3QjtBQUN0QixjQUFNSSxXQUFXLEdBQUc7QUFDbEJDLGtCQUFNLEVBQUVOLEdBQUcsQ0FBQ0UsUUFBSixDQUFhSixJQURIO0FBRWxCUyxtQkFBTyxFQUFFLEtBRlM7QUFHbEJDLHFCQUFTLEVBQUUsRUFITztBQUlsQkMsc0JBQVUsRUFBRTtBQUpNLFdBQXBCO0FBTUFYLGNBQUksQ0FBQ00sS0FBTCxHQUFhckIsMERBQVcsQ0FBQzJCLHFCQUFaLENBQWtDTCxXQUFsQyxDQUFiO0FBQ0QsU0FSRCxNQVFPLElBQUlKLFVBQVUsS0FBSyxHQUFuQixFQUF3QjtBQUM3QkgsY0FBSSxDQUFDTSxLQUFMLEdBQWFKLEdBQUcsQ0FBQ0UsUUFBSixDQUFhSixJQUFiLENBQWtCYSxPQUEvQjtBQUNEOztBQUNELGVBQU9uQixNQUFNLENBQUNNLElBQUQsQ0FBYjtBQUNELE9BeEJIO0FBeUJELEtBMUJELENBRGE7QUFBQSxHQUFmO0FBNkJEO0FBRU0sU0FBU2Msb0JBQVQsQ0FBOEIxQixNQUE5QixFQUFzQztBQUMzQyxTQUFPLFVBQUFHLFFBQVE7QUFBQSxXQUNiLElBQUlDLE9BQUosQ0FBWSxVQUFDQyxPQUFELEVBQVVDLE1BQVYsRUFBcUI7QUFDL0JDLHlEQUFJLENBQUNvQixLQUFMLG9CQUF1QjNCLE1BQU0sQ0FBQzRCLEVBQTlCLEdBQW9DakMsZ0JBQWdCLENBQUNLLE1BQUQsQ0FBcEQsRUFDR1MsSUFESCxDQUNRLFVBQUFDLEdBQUcsRUFBSTtBQUNYUCxnQkFBUSxDQUFDUSxrREFBQSxDQUFtQlosaUJBQWlCLENBQUNXLEdBQUcsQ0FBQ0UsSUFBTCxDQUFwQyxDQUFELENBQVI7QUFDQSxlQUFPUCxPQUFPLEVBQWQ7QUFDRCxPQUpILEVBS0dRLEtBTEgsQ0FLUyxVQUFDQyxHQUFELEVBQVM7QUFDZCxZQUFNQyxVQUFVLEdBQUdELEdBQUcsQ0FBQ0UsUUFBSixDQUFhQyxNQUFoQztBQUNBLFlBQU1MLElBQUksR0FBRztBQUNYTSxlQUFLLEVBQUUsSUFESTtBQUVYSCxvQkFBVSxFQUFWQTtBQUZXLFNBQWI7O0FBS0EsWUFBSUEsVUFBVSxLQUFLLEdBQW5CLEVBQXdCO0FBQ3RCLGNBQU1JLFdBQVcsR0FBRztBQUNsQkMsa0JBQU0sRUFBRU4sR0FBRyxDQUFDRSxRQUFKLENBQWFKLElBREg7QUFFbEJTLG1CQUFPLEVBQUUsS0FGUztBQUdsQkMscUJBQVMsRUFBRSxFQUhPO0FBSWxCQyxzQkFBVSxFQUFFO0FBSk0sV0FBcEI7QUFNQVgsY0FBSSxDQUFDTSxLQUFMLEdBQWFyQiwwREFBVyxDQUFDMkIscUJBQVosQ0FBa0NMLFdBQWxDLENBQWI7QUFDRCxTQVJELE1BUU8sSUFBSUosVUFBVSxLQUFLLEdBQW5CLEVBQXdCO0FBQzdCSCxjQUFJLENBQUNNLEtBQUwsR0FBYUosR0FBRyxDQUFDRSxRQUFKLENBQWFKLElBQWIsQ0FBa0JhLE9BQS9CO0FBQ0Q7O0FBQ0QsZUFBT25CLE1BQU0sQ0FBQ00sSUFBRCxDQUFiO0FBQ0QsT0F4Qkg7QUF5QkQsS0ExQkQsQ0FEYTtBQUFBLEdBQWY7QUE2QkQ7QUFFTSxTQUFTaUIsb0JBQVQsQ0FBOEJELEVBQTlCLEVBQWtDO0FBQ3ZDLFNBQU8sVUFBQXpCLFFBQVEsRUFBSTtBQUNqQkksdURBQUksQ0FBQ3VCLE1BQUwsb0JBQXdCRixFQUF4QixHQUNHbkIsSUFESCxDQUNRLFlBQU07QUFDVk4sY0FBUSxDQUFDUSxxREFBQSxDQUFzQmlCLEVBQXRCLENBQUQsQ0FBUjtBQUNELEtBSEgsRUFJR2YsS0FKSCxDQUlTLFVBQUNDLEdBQUQsRUFBUztBQUNkO0FBQ0FpQixhQUFPLENBQUNiLEtBQVIsQ0FBY0osR0FBRyxDQUFDRSxRQUFsQjtBQUNELEtBUEg7QUFRRCxHQVREO0FBVUQ7QUFFTSxTQUFTZ0Isa0JBQVQsT0FBaUU7QUFBQSw2QkFBcENDLFVBQW9DO0FBQUEsTUFBcENBLFVBQW9DLGdDQUF2QixDQUF1QjtBQUFBLHNCQUFwQkMsR0FBb0I7QUFBQSxNQUFwQkEsR0FBb0IseUJBQWQsV0FBYztBQUN0RSxTQUFPLFVBQUEvQixRQUFRLEVBQUk7QUFDakIsUUFBSThCLFVBQVUsR0FBRyxDQUFqQixFQUFvQjtBQUNsQkMsU0FBRyxHQUFHQSxHQUFHLG1CQUFZRCxVQUFaLENBQVQ7QUFDRDs7QUFFRDFCLHVEQUFJLENBQUM0QixHQUFMLENBQVNELEdBQVQsRUFDR3pCLElBREgsQ0FDUSxVQUFDQyxHQUFELEVBQVM7QUFDYlAsY0FBUSxDQUFDUSxtREFBQSxDQUFvQlosaUJBQWlCLENBQUNXLEdBQUcsQ0FBQ0UsSUFBTCxDQUFyQyxDQUFELENBQVI7QUFDRCxLQUhILEVBSUdDLEtBSkgsQ0FJUyxVQUFDQyxHQUFELEVBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FiRDtBQWNEO0FBRU0sU0FBU29CLGtCQUFULENBQTRCUixFQUE1QixFQUFnQztBQUNyQyxTQUFPLFVBQUF6QixRQUFRLEVBQUk7QUFDakJJLHVEQUFJLENBQUM0QixHQUFMLG9CQUFxQlAsRUFBckIsR0FDR25CLElBREgsQ0FDUSxVQUFDQyxHQUFELEVBQVM7QUFDYlAsY0FBUSxDQUFDUSxrREFBQSxDQUFtQlosaUJBQWlCLENBQUNXLEdBQUcsQ0FBQ0UsSUFBTCxDQUFwQyxDQUFELENBQVI7QUFDRCxLQUhILEVBSUdDLEtBSkgsQ0FJUyxVQUFDQyxHQUFELEVBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FURDtBQVVEO0FBRU0sU0FBU3FCLG1CQUFULENBQTZCQyxJQUE3QixFQUFtQztBQUN4QyxTQUFPLFVBQUFuQyxRQUFRLEVBQUk7QUFDakJJLHVEQUFJLENBQUM0QixHQUFMLDhCQUErQkcsSUFBL0IsR0FDRzdCLElBREgsQ0FDUSxVQUFDQyxHQUFELEVBQVM7QUFDYlAsY0FBUSxDQUFDUSxrREFBQSxDQUFtQlosaUJBQWlCLENBQUNXLEdBQUcsQ0FBQ0UsSUFBTCxDQUFwQyxDQUFELENBQVI7QUFDRCxLQUhILEVBSUdDLEtBSkgsQ0FJUyxVQUFDQyxHQUFELEVBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FURDtBQVVEIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvYXJ0aWNsZS9zZXJ2aWNlLmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiaW1wb3J0IEh0dHAgZnJvbSAnLi4vLi4vdXRpbHMvSHR0cCdcbmltcG9ydCBUcmFuc2Zvcm1lciBmcm9tICcuLi8uLi91dGlscy9UcmFuc2Zvcm1lcidcbmltcG9ydCAqIGFzIGFydGljbGVBY3Rpb25zIGZyb20gJy4vc3RvcmUvYWN0aW9ucydcblxuZnVuY3Rpb24gdHJhbnNmb3JtUmVxdWVzdChwYXJtcykge1xuICByZXR1cm4gVHJhbnNmb3JtZXIuc2VuZChwYXJtcylcbn1cblxuZnVuY3Rpb24gdHJhbnNmb3JtUmVzcG9uc2UocGFyYW1zKSB7XG4gIHJldHVybiBUcmFuc2Zvcm1lci5mZXRjaChwYXJhbXMpXG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlQWRkUmVxdWVzdChwYXJhbXMpIHtcbiAgcmV0dXJuIGRpc3BhdGNoID0+IChcbiAgICBuZXcgUHJvbWlzZSgocmVzb2x2ZSwgcmVqZWN0KSA9PiB7XG4gICAgICBIdHRwLnBvc3QoJy9hcnRpY2xlcycsIHRyYW5zZm9ybVJlcXVlc3QocGFyYW1zKSlcbiAgICAgICAgLnRoZW4ocmVzID0+IHtcbiAgICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5hZGQodHJhbnNmb3JtUmVzcG9uc2UocmVzLmRhdGEpKSlcbiAgICAgICAgICByZXR1cm4gcmVzb2x2ZSgpXG4gICAgICAgIH0pXG4gICAgICAgIC5jYXRjaCgoZXJyKSA9PiB7XG4gICAgICAgICAgY29uc3Qgc3RhdHVzQ29kZSA9IGVyci5yZXNwb25zZS5zdGF0dXM7XG4gICAgICAgICAgY29uc3QgZGF0YSA9IHtcbiAgICAgICAgICAgIGVycm9yOiBudWxsLFxuICAgICAgICAgICAgc3RhdHVzQ29kZSxcbiAgICAgICAgICB9O1xuXG4gICAgICAgICAgaWYgKHN0YXR1c0NvZGUgPT09IDQyMikge1xuICAgICAgICAgICAgY29uc3QgcmVzZXRFcnJvcnMgPSB7XG4gICAgICAgICAgICAgIGVycm9yczogZXJyLnJlc3BvbnNlLmRhdGEsXG4gICAgICAgICAgICAgIHJlcGxhY2U6IGZhbHNlLFxuICAgICAgICAgICAgICBzZWFyY2hTdHI6ICcnLFxuICAgICAgICAgICAgICByZXBsYWNlU3RyOiAnJyxcbiAgICAgICAgICAgIH07XG4gICAgICAgICAgICBkYXRhLmVycm9yID0gVHJhbnNmb3JtZXIucmVzZXRWYWxpZGF0aW9uRmllbGRzKHJlc2V0RXJyb3JzKTtcbiAgICAgICAgICB9IGVsc2UgaWYgKHN0YXR1c0NvZGUgPT09IDQwMSkge1xuICAgICAgICAgICAgZGF0YS5lcnJvciA9IGVyci5yZXNwb25zZS5kYXRhLm1lc3NhZ2U7XG4gICAgICAgICAgfVxuICAgICAgICAgIHJldHVybiByZWplY3QoZGF0YSk7XG4gICAgICAgIH0pXG4gICAgfSlcbiAgKVxufVxuXG5leHBvcnQgZnVuY3Rpb24gYXJ0aWNsZVVwZGF0ZVJlcXVlc3QocGFyYW1zKSB7XG4gIHJldHVybiBkaXNwYXRjaCA9PiAoXG4gICAgbmV3IFByb21pc2UoKHJlc29sdmUsIHJlamVjdCkgPT4ge1xuICAgICAgSHR0cC5wYXRjaChgYXJ0aWNsZXMvJHtwYXJhbXMuaWR9YCwgdHJhbnNmb3JtUmVxdWVzdChwYXJhbXMpKVxuICAgICAgICAudGhlbihyZXMgPT4ge1xuICAgICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLmFkZCh0cmFuc2Zvcm1SZXNwb25zZShyZXMuZGF0YSkpKVxuICAgICAgICAgIHJldHVybiByZXNvbHZlKClcbiAgICAgICAgfSlcbiAgICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgICBjb25zdCBzdGF0dXNDb2RlID0gZXJyLnJlc3BvbnNlLnN0YXR1cztcbiAgICAgICAgICBjb25zdCBkYXRhID0ge1xuICAgICAgICAgICAgZXJyb3I6IG51bGwsXG4gICAgICAgICAgICBzdGF0dXNDb2RlLFxuICAgICAgICAgIH07XG5cbiAgICAgICAgICBpZiAoc3RhdHVzQ29kZSA9PT0gNDIyKSB7XG4gICAgICAgICAgICBjb25zdCByZXNldEVycm9ycyA9IHtcbiAgICAgICAgICAgICAgZXJyb3JzOiBlcnIucmVzcG9uc2UuZGF0YSxcbiAgICAgICAgICAgICAgcmVwbGFjZTogZmFsc2UsXG4gICAgICAgICAgICAgIHNlYXJjaFN0cjogJycsXG4gICAgICAgICAgICAgIHJlcGxhY2VTdHI6ICcnLFxuICAgICAgICAgICAgfTtcbiAgICAgICAgICAgIGRhdGEuZXJyb3IgPSBUcmFuc2Zvcm1lci5yZXNldFZhbGlkYXRpb25GaWVsZHMocmVzZXRFcnJvcnMpO1xuICAgICAgICAgIH0gZWxzZSBpZiAoc3RhdHVzQ29kZSA9PT0gNDAxKSB7XG4gICAgICAgICAgICBkYXRhLmVycm9yID0gZXJyLnJlc3BvbnNlLmRhdGEubWVzc2FnZTtcbiAgICAgICAgICB9XG4gICAgICAgICAgcmV0dXJuIHJlamVjdChkYXRhKTtcbiAgICAgICAgfSlcbiAgICB9KVxuICApXG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlUmVtb3ZlUmVxdWVzdChpZCkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIEh0dHAuZGVsZXRlKGBhcnRpY2xlcy8ke2lkfWApXG4gICAgICAudGhlbigoKSA9PiB7XG4gICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLnJlbW92ZShpZCkpXG4gICAgICB9KVxuICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgLy8gVE9ETzogaGFuZGxlIGVyclxuICAgICAgICBjb25zb2xlLmVycm9yKGVyci5yZXNwb25zZSlcbiAgICAgIH0pXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIGFydGljbGVMaXN0UmVxdWVzdCh7cGFnZU51bWJlciA9IDEsIHVybCA9ICcvYXJ0aWNsZXMnfSkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIGlmIChwYWdlTnVtYmVyID4gMSkge1xuICAgICAgdXJsID0gdXJsICsgYD9wYWdlPSR7cGFnZU51bWJlcn1gXG4gICAgfVxuXG4gICAgSHR0cC5nZXQodXJsKVxuICAgICAgLnRoZW4oKHJlcykgPT4ge1xuICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5saXN0KHRyYW5zZm9ybVJlc3BvbnNlKHJlcy5kYXRhKSkpXG4gICAgICB9KVxuICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgLy8gVE9ETzogaGFuZGxlIGVyclxuICAgICAgICBjb25zb2xlLmVycm9yKGVyci5yZXNwb25zZSlcbiAgICAgIH0pXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIGFydGljbGVFZGl0UmVxdWVzdChpZCkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIEh0dHAuZ2V0KGBhcnRpY2xlcy8ke2lkfWApXG4gICAgICAudGhlbigocmVzKSA9PiB7XG4gICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLmFkZCh0cmFuc2Zvcm1SZXNwb25zZShyZXMuZGF0YSkpKVxuICAgICAgfSlcbiAgICAgIC5jYXRjaCgoZXJyKSA9PiB7XG4gICAgICAgIC8vIFRPRE86IGhhbmRsZSBlcnJcbiAgICAgICAgY29uc29sZS5lcnJvcihlcnIucmVzcG9uc2UpXG4gICAgICB9KVxuICB9XG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlRmV0Y2hSZXF1ZXN0KHNsdWcpIHtcbiAgcmV0dXJuIGRpc3BhdGNoID0+IHtcbiAgICBIdHRwLmdldChgYXJ0aWNsZXMvcHVibGlzaGVkLyR7c2x1Z31gKVxuICAgICAgLnRoZW4oKHJlcykgPT4ge1xuICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5hZGQodHJhbnNmb3JtUmVzcG9uc2UocmVzLmRhdGEpKSlcbiAgICAgIH0pXG4gICAgICAuY2F0Y2goKGVycikgPT4ge1xuICAgICAgICAvLyBUT0RPOiBoYW5kbGUgZXJyXG4gICAgICAgIGNvbnNvbGUuZXJyb3IoZXJyLnJlc3BvbnNlKVxuICAgICAgfSlcbiAgfVxufSJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/article/service.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleAddRequest\", function() { return articleAddRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleUpdateRequest\", function() { return articleUpdateRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleRemoveRequest\", function() { return articleRemoveRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleListRequest\", function() { return articleListRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleEditRequest\", function() { return articleEditRequest; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"articleFetchRequest\", function() { return articleFetchRequest; });\n/* harmony import */ var _utils_Http__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../../utils/Http */ \"./resources/js/utils/Http.js\");\n/* harmony import */ var _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../utils/Transformer */ \"./resources/js/utils/Transformer.js\");\n/* harmony import */ var _store_actions__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./store/actions */ \"./resources/js/modules/article/store/actions.js\");\n\n\n\n\nfunction transformRequest(parms) {\n  return _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].send(parms);\n}\n\nfunction transformResponse(params) {\n  return _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].fetch(params);\n}\n\nfunction articleAddRequest(params) {\n  return dispatch => new Promise((resolve, reject) => {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].post('/articles', transformRequest(params)).then(res => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n      return resolve();\n    }).catch(err => {\n      const statusCode = err.response.status;\n      const data = {\n        error: null,\n        statusCode\n      };\n\n      if (statusCode === 422) {\n        const resetErrors = {\n          errors: err.response.data,\n          replace: false,\n          searchStr: '',\n          replaceStr: ''\n        };\n        data.error = _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].resetValidationFields(resetErrors);\n      } else if (statusCode === 401) {\n        data.error = err.response.data.message;\n      }\n\n      return reject(data);\n    });\n  });\n}\nfunction articleUpdateRequest(params) {\n  return dispatch => new Promise((resolve, reject) => {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].patch(\"articles/\".concat(params.id), transformRequest(params)).then(res => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n      return resolve();\n    }).catch(err => {\n      const statusCode = err.response.status;\n      const data = {\n        error: null,\n        statusCode\n      };\n\n      if (statusCode === 422) {\n        const resetErrors = {\n          errors: err.response.data,\n          replace: false,\n          searchStr: '',\n          replaceStr: ''\n        };\n        data.error = _utils_Transformer__WEBPACK_IMPORTED_MODULE_1__[\"default\"].resetValidationFields(resetErrors);\n      } else if (statusCode === 401) {\n        data.error = err.response.data.message;\n      }\n\n      return reject(data);\n    });\n  });\n}\nfunction articleRemoveRequest(id) {\n  return dispatch => {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].delete(\"articles/\".concat(id)).then(() => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"remove\"](id));\n    }).catch(err => {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleListRequest({\n  pageNumber = 1,\n  url = '/articles'\n}) {\n  return dispatch => {\n    if (pageNumber > 1) {\n      url = url + \"?page=\".concat(pageNumber);\n    }\n\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(url).then(res => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"list\"](transformResponse(res.data)));\n    }).catch(err => {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleEditRequest(id) {\n  return dispatch => {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(\"articles/\".concat(id)).then(res => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n    }).catch(err => {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}\nfunction articleFetchRequest(slug) {\n  return dispatch => {\n    _utils_Http__WEBPACK_IMPORTED_MODULE_0__[\"default\"].get(\"articles/published/\".concat(slug)).then(res => {\n      dispatch(_store_actions__WEBPACK_IMPORTED_MODULE_2__[\"add\"](transformResponse(res.data)));\n    }).catch(err => {\n      // TODO: handle err\n      console.error(err.response);\n    });\n  };\n}//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL3NlcnZpY2UuanM/Y2YwNiJdLCJuYW1lcyI6WyJ0cmFuc2Zvcm1SZXF1ZXN0IiwicGFybXMiLCJUcmFuc2Zvcm1lciIsInNlbmQiLCJ0cmFuc2Zvcm1SZXNwb25zZSIsInBhcmFtcyIsImZldGNoIiwiYXJ0aWNsZUFkZFJlcXVlc3QiLCJkaXNwYXRjaCIsIlByb21pc2UiLCJyZXNvbHZlIiwicmVqZWN0IiwiSHR0cCIsInBvc3QiLCJ0aGVuIiwicmVzIiwiYXJ0aWNsZUFjdGlvbnMiLCJkYXRhIiwiY2F0Y2giLCJlcnIiLCJzdGF0dXNDb2RlIiwicmVzcG9uc2UiLCJzdGF0dXMiLCJlcnJvciIsInJlc2V0RXJyb3JzIiwiZXJyb3JzIiwicmVwbGFjZSIsInNlYXJjaFN0ciIsInJlcGxhY2VTdHIiLCJyZXNldFZhbGlkYXRpb25GaWVsZHMiLCJtZXNzYWdlIiwiYXJ0aWNsZVVwZGF0ZVJlcXVlc3QiLCJwYXRjaCIsImlkIiwiYXJ0aWNsZVJlbW92ZVJlcXVlc3QiLCJkZWxldGUiLCJjb25zb2xlIiwiYXJ0aWNsZUxpc3RSZXF1ZXN0IiwicGFnZU51bWJlciIsInVybCIsImdldCIsImFydGljbGVFZGl0UmVxdWVzdCIsImFydGljbGVGZXRjaFJlcXVlc3QiLCJzbHVnIl0sIm1hcHBpbmdzIjoiQUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQ0E7QUFDQTs7QUFFQSxTQUFTQSxnQkFBVCxDQUEwQkMsS0FBMUIsRUFBaUM7QUFDL0IsU0FBT0MsMERBQVcsQ0FBQ0MsSUFBWixDQUFpQkYsS0FBakIsQ0FBUDtBQUNEOztBQUVELFNBQVNHLGlCQUFULENBQTJCQyxNQUEzQixFQUFtQztBQUNqQyxTQUFPSCwwREFBVyxDQUFDSSxLQUFaLENBQWtCRCxNQUFsQixDQUFQO0FBQ0Q7O0FBRU0sU0FBU0UsaUJBQVQsQ0FBMkJGLE1BQTNCLEVBQW1DO0FBQ3hDLFNBQU9HLFFBQVEsSUFDYixJQUFJQyxPQUFKLENBQVksQ0FBQ0MsT0FBRCxFQUFVQyxNQUFWLEtBQXFCO0FBQy9CQyx1REFBSSxDQUFDQyxJQUFMLENBQVUsV0FBVixFQUF1QmIsZ0JBQWdCLENBQUNLLE1BQUQsQ0FBdkMsRUFDR1MsSUFESCxDQUNRQyxHQUFHLElBQUk7QUFDWFAsY0FBUSxDQUFDUSxrREFBQSxDQUFtQlosaUJBQWlCLENBQUNXLEdBQUcsQ0FBQ0UsSUFBTCxDQUFwQyxDQUFELENBQVI7QUFDQSxhQUFPUCxPQUFPLEVBQWQ7QUFDRCxLQUpILEVBS0dRLEtBTEgsQ0FLVUMsR0FBRCxJQUFTO0FBQ2QsWUFBTUMsVUFBVSxHQUFHRCxHQUFHLENBQUNFLFFBQUosQ0FBYUMsTUFBaEM7QUFDQSxZQUFNTCxJQUFJLEdBQUc7QUFDWE0sYUFBSyxFQUFFLElBREk7QUFFWEg7QUFGVyxPQUFiOztBQUtBLFVBQUlBLFVBQVUsS0FBSyxHQUFuQixFQUF3QjtBQUN0QixjQUFNSSxXQUFXLEdBQUc7QUFDbEJDLGdCQUFNLEVBQUVOLEdBQUcsQ0FBQ0UsUUFBSixDQUFhSixJQURIO0FBRWxCUyxpQkFBTyxFQUFFLEtBRlM7QUFHbEJDLG1CQUFTLEVBQUUsRUFITztBQUlsQkMsb0JBQVUsRUFBRTtBQUpNLFNBQXBCO0FBTUFYLFlBQUksQ0FBQ00sS0FBTCxHQUFhckIsMERBQVcsQ0FBQzJCLHFCQUFaLENBQWtDTCxXQUFsQyxDQUFiO0FBQ0QsT0FSRCxNQVFPLElBQUlKLFVBQVUsS0FBSyxHQUFuQixFQUF3QjtBQUM3QkgsWUFBSSxDQUFDTSxLQUFMLEdBQWFKLEdBQUcsQ0FBQ0UsUUFBSixDQUFhSixJQUFiLENBQWtCYSxPQUEvQjtBQUNEOztBQUNELGFBQU9uQixNQUFNLENBQUNNLElBQUQsQ0FBYjtBQUNELEtBeEJIO0FBeUJELEdBMUJELENBREY7QUE2QkQ7QUFFTSxTQUFTYyxvQkFBVCxDQUE4QjFCLE1BQTlCLEVBQXNDO0FBQzNDLFNBQU9HLFFBQVEsSUFDYixJQUFJQyxPQUFKLENBQVksQ0FBQ0MsT0FBRCxFQUFVQyxNQUFWLEtBQXFCO0FBQy9CQyx1REFBSSxDQUFDb0IsS0FBTCxvQkFBdUIzQixNQUFNLENBQUM0QixFQUE5QixHQUFvQ2pDLGdCQUFnQixDQUFDSyxNQUFELENBQXBELEVBQ0dTLElBREgsQ0FDUUMsR0FBRyxJQUFJO0FBQ1hQLGNBQVEsQ0FBQ1Esa0RBQUEsQ0FBbUJaLGlCQUFpQixDQUFDVyxHQUFHLENBQUNFLElBQUwsQ0FBcEMsQ0FBRCxDQUFSO0FBQ0EsYUFBT1AsT0FBTyxFQUFkO0FBQ0QsS0FKSCxFQUtHUSxLQUxILENBS1VDLEdBQUQsSUFBUztBQUNkLFlBQU1DLFVBQVUsR0FBR0QsR0FBRyxDQUFDRSxRQUFKLENBQWFDLE1BQWhDO0FBQ0EsWUFBTUwsSUFBSSxHQUFHO0FBQ1hNLGFBQUssRUFBRSxJQURJO0FBRVhIO0FBRlcsT0FBYjs7QUFLQSxVQUFJQSxVQUFVLEtBQUssR0FBbkIsRUFBd0I7QUFDdEIsY0FBTUksV0FBVyxHQUFHO0FBQ2xCQyxnQkFBTSxFQUFFTixHQUFHLENBQUNFLFFBQUosQ0FBYUosSUFESDtBQUVsQlMsaUJBQU8sRUFBRSxLQUZTO0FBR2xCQyxtQkFBUyxFQUFFLEVBSE87QUFJbEJDLG9CQUFVLEVBQUU7QUFKTSxTQUFwQjtBQU1BWCxZQUFJLENBQUNNLEtBQUwsR0FBYXJCLDBEQUFXLENBQUMyQixxQkFBWixDQUFrQ0wsV0FBbEMsQ0FBYjtBQUNELE9BUkQsTUFRTyxJQUFJSixVQUFVLEtBQUssR0FBbkIsRUFBd0I7QUFDN0JILFlBQUksQ0FBQ00sS0FBTCxHQUFhSixHQUFHLENBQUNFLFFBQUosQ0FBYUosSUFBYixDQUFrQmEsT0FBL0I7QUFDRDs7QUFDRCxhQUFPbkIsTUFBTSxDQUFDTSxJQUFELENBQWI7QUFDRCxLQXhCSDtBQXlCRCxHQTFCRCxDQURGO0FBNkJEO0FBRU0sU0FBU2lCLG9CQUFULENBQThCRCxFQUE5QixFQUFrQztBQUN2QyxTQUFPekIsUUFBUSxJQUFJO0FBQ2pCSSx1REFBSSxDQUFDdUIsTUFBTCxvQkFBd0JGLEVBQXhCLEdBQ0duQixJQURILENBQ1EsTUFBTTtBQUNWTixjQUFRLENBQUNRLHFEQUFBLENBQXNCaUIsRUFBdEIsQ0FBRCxDQUFSO0FBQ0QsS0FISCxFQUlHZixLQUpILENBSVVDLEdBQUQsSUFBUztBQUNkO0FBQ0FpQixhQUFPLENBQUNiLEtBQVIsQ0FBY0osR0FBRyxDQUFDRSxRQUFsQjtBQUNELEtBUEg7QUFRRCxHQVREO0FBVUQ7QUFFTSxTQUFTZ0Isa0JBQVQsQ0FBNEI7QUFBQ0MsWUFBVSxHQUFHLENBQWQ7QUFBaUJDLEtBQUcsR0FBRztBQUF2QixDQUE1QixFQUFpRTtBQUN0RSxTQUFPL0IsUUFBUSxJQUFJO0FBQ2pCLFFBQUk4QixVQUFVLEdBQUcsQ0FBakIsRUFBb0I7QUFDbEJDLFNBQUcsR0FBR0EsR0FBRyxtQkFBWUQsVUFBWixDQUFUO0FBQ0Q7O0FBRUQxQix1REFBSSxDQUFDNEIsR0FBTCxDQUFTRCxHQUFULEVBQ0d6QixJQURILENBQ1NDLEdBQUQsSUFBUztBQUNiUCxjQUFRLENBQUNRLG1EQUFBLENBQW9CWixpQkFBaUIsQ0FBQ1csR0FBRyxDQUFDRSxJQUFMLENBQXJDLENBQUQsQ0FBUjtBQUNELEtBSEgsRUFJR0MsS0FKSCxDQUlVQyxHQUFELElBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FiRDtBQWNEO0FBRU0sU0FBU29CLGtCQUFULENBQTRCUixFQUE1QixFQUFnQztBQUNyQyxTQUFPekIsUUFBUSxJQUFJO0FBQ2pCSSx1REFBSSxDQUFDNEIsR0FBTCxvQkFBcUJQLEVBQXJCLEdBQ0duQixJQURILENBQ1NDLEdBQUQsSUFBUztBQUNiUCxjQUFRLENBQUNRLGtEQUFBLENBQW1CWixpQkFBaUIsQ0FBQ1csR0FBRyxDQUFDRSxJQUFMLENBQXBDLENBQUQsQ0FBUjtBQUNELEtBSEgsRUFJR0MsS0FKSCxDQUlVQyxHQUFELElBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FURDtBQVVEO0FBRU0sU0FBU3FCLG1CQUFULENBQTZCQyxJQUE3QixFQUFtQztBQUN4QyxTQUFPbkMsUUFBUSxJQUFJO0FBQ2pCSSx1REFBSSxDQUFDNEIsR0FBTCw4QkFBK0JHLElBQS9CLEdBQ0c3QixJQURILENBQ1NDLEdBQUQsSUFBUztBQUNiUCxjQUFRLENBQUNRLGtEQUFBLENBQW1CWixpQkFBaUIsQ0FBQ1csR0FBRyxDQUFDRSxJQUFMLENBQXBDLENBQUQsQ0FBUjtBQUNELEtBSEgsRUFJR0MsS0FKSCxDQUlVQyxHQUFELElBQVM7QUFDZDtBQUNBaUIsYUFBTyxDQUFDYixLQUFSLENBQWNKLEdBQUcsQ0FBQ0UsUUFBbEI7QUFDRCxLQVBIO0FBUUQsR0FURDtBQVVEIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvYXJ0aWNsZS9zZXJ2aWNlLmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiaW1wb3J0IEh0dHAgZnJvbSAnLi4vLi4vdXRpbHMvSHR0cCdcbmltcG9ydCBUcmFuc2Zvcm1lciBmcm9tICcuLi8uLi91dGlscy9UcmFuc2Zvcm1lcidcbmltcG9ydCAqIGFzIGFydGljbGVBY3Rpb25zIGZyb20gJy4vc3RvcmUvYWN0aW9ucydcblxuZnVuY3Rpb24gdHJhbnNmb3JtUmVxdWVzdChwYXJtcykge1xuICByZXR1cm4gVHJhbnNmb3JtZXIuc2VuZChwYXJtcylcbn1cblxuZnVuY3Rpb24gdHJhbnNmb3JtUmVzcG9uc2UocGFyYW1zKSB7XG4gIHJldHVybiBUcmFuc2Zvcm1lci5mZXRjaChwYXJhbXMpXG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlQWRkUmVxdWVzdChwYXJhbXMpIHtcbiAgcmV0dXJuIGRpc3BhdGNoID0+IChcbiAgICBuZXcgUHJvbWlzZSgocmVzb2x2ZSwgcmVqZWN0KSA9PiB7XG4gICAgICBIdHRwLnBvc3QoJy9hcnRpY2xlcycsIHRyYW5zZm9ybVJlcXVlc3QocGFyYW1zKSlcbiAgICAgICAgLnRoZW4ocmVzID0+IHtcbiAgICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5hZGQodHJhbnNmb3JtUmVzcG9uc2UocmVzLmRhdGEpKSlcbiAgICAgICAgICByZXR1cm4gcmVzb2x2ZSgpXG4gICAgICAgIH0pXG4gICAgICAgIC5jYXRjaCgoZXJyKSA9PiB7XG4gICAgICAgICAgY29uc3Qgc3RhdHVzQ29kZSA9IGVyci5yZXNwb25zZS5zdGF0dXM7XG4gICAgICAgICAgY29uc3QgZGF0YSA9IHtcbiAgICAgICAgICAgIGVycm9yOiBudWxsLFxuICAgICAgICAgICAgc3RhdHVzQ29kZSxcbiAgICAgICAgICB9O1xuXG4gICAgICAgICAgaWYgKHN0YXR1c0NvZGUgPT09IDQyMikge1xuICAgICAgICAgICAgY29uc3QgcmVzZXRFcnJvcnMgPSB7XG4gICAgICAgICAgICAgIGVycm9yczogZXJyLnJlc3BvbnNlLmRhdGEsXG4gICAgICAgICAgICAgIHJlcGxhY2U6IGZhbHNlLFxuICAgICAgICAgICAgICBzZWFyY2hTdHI6ICcnLFxuICAgICAgICAgICAgICByZXBsYWNlU3RyOiAnJyxcbiAgICAgICAgICAgIH07XG4gICAgICAgICAgICBkYXRhLmVycm9yID0gVHJhbnNmb3JtZXIucmVzZXRWYWxpZGF0aW9uRmllbGRzKHJlc2V0RXJyb3JzKTtcbiAgICAgICAgICB9IGVsc2UgaWYgKHN0YXR1c0NvZGUgPT09IDQwMSkge1xuICAgICAgICAgICAgZGF0YS5lcnJvciA9IGVyci5yZXNwb25zZS5kYXRhLm1lc3NhZ2U7XG4gICAgICAgICAgfVxuICAgICAgICAgIHJldHVybiByZWplY3QoZGF0YSk7XG4gICAgICAgIH0pXG4gICAgfSlcbiAgKVxufVxuXG5leHBvcnQgZnVuY3Rpb24gYXJ0aWNsZVVwZGF0ZVJlcXVlc3QocGFyYW1zKSB7XG4gIHJldHVybiBkaXNwYXRjaCA9PiAoXG4gICAgbmV3IFByb21pc2UoKHJlc29sdmUsIHJlamVjdCkgPT4ge1xuICAgICAgSHR0cC5wYXRjaChgYXJ0aWNsZXMvJHtwYXJhbXMuaWR9YCwgdHJhbnNmb3JtUmVxdWVzdChwYXJhbXMpKVxuICAgICAgICAudGhlbihyZXMgPT4ge1xuICAgICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLmFkZCh0cmFuc2Zvcm1SZXNwb25zZShyZXMuZGF0YSkpKVxuICAgICAgICAgIHJldHVybiByZXNvbHZlKClcbiAgICAgICAgfSlcbiAgICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgICBjb25zdCBzdGF0dXNDb2RlID0gZXJyLnJlc3BvbnNlLnN0YXR1cztcbiAgICAgICAgICBjb25zdCBkYXRhID0ge1xuICAgICAgICAgICAgZXJyb3I6IG51bGwsXG4gICAgICAgICAgICBzdGF0dXNDb2RlLFxuICAgICAgICAgIH07XG5cbiAgICAgICAgICBpZiAoc3RhdHVzQ29kZSA9PT0gNDIyKSB7XG4gICAgICAgICAgICBjb25zdCByZXNldEVycm9ycyA9IHtcbiAgICAgICAgICAgICAgZXJyb3JzOiBlcnIucmVzcG9uc2UuZGF0YSxcbiAgICAgICAgICAgICAgcmVwbGFjZTogZmFsc2UsXG4gICAgICAgICAgICAgIHNlYXJjaFN0cjogJycsXG4gICAgICAgICAgICAgIHJlcGxhY2VTdHI6ICcnLFxuICAgICAgICAgICAgfTtcbiAgICAgICAgICAgIGRhdGEuZXJyb3IgPSBUcmFuc2Zvcm1lci5yZXNldFZhbGlkYXRpb25GaWVsZHMocmVzZXRFcnJvcnMpO1xuICAgICAgICAgIH0gZWxzZSBpZiAoc3RhdHVzQ29kZSA9PT0gNDAxKSB7XG4gICAgICAgICAgICBkYXRhLmVycm9yID0gZXJyLnJlc3BvbnNlLmRhdGEubWVzc2FnZTtcbiAgICAgICAgICB9XG4gICAgICAgICAgcmV0dXJuIHJlamVjdChkYXRhKTtcbiAgICAgICAgfSlcbiAgICB9KVxuICApXG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlUmVtb3ZlUmVxdWVzdChpZCkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIEh0dHAuZGVsZXRlKGBhcnRpY2xlcy8ke2lkfWApXG4gICAgICAudGhlbigoKSA9PiB7XG4gICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLnJlbW92ZShpZCkpXG4gICAgICB9KVxuICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgLy8gVE9ETzogaGFuZGxlIGVyclxuICAgICAgICBjb25zb2xlLmVycm9yKGVyci5yZXNwb25zZSlcbiAgICAgIH0pXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIGFydGljbGVMaXN0UmVxdWVzdCh7cGFnZU51bWJlciA9IDEsIHVybCA9ICcvYXJ0aWNsZXMnfSkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIGlmIChwYWdlTnVtYmVyID4gMSkge1xuICAgICAgdXJsID0gdXJsICsgYD9wYWdlPSR7cGFnZU51bWJlcn1gXG4gICAgfVxuXG4gICAgSHR0cC5nZXQodXJsKVxuICAgICAgLnRoZW4oKHJlcykgPT4ge1xuICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5saXN0KHRyYW5zZm9ybVJlc3BvbnNlKHJlcy5kYXRhKSkpXG4gICAgICB9KVxuICAgICAgLmNhdGNoKChlcnIpID0+IHtcbiAgICAgICAgLy8gVE9ETzogaGFuZGxlIGVyclxuICAgICAgICBjb25zb2xlLmVycm9yKGVyci5yZXNwb25zZSlcbiAgICAgIH0pXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIGFydGljbGVFZGl0UmVxdWVzdChpZCkge1xuICByZXR1cm4gZGlzcGF0Y2ggPT4ge1xuICAgIEh0dHAuZ2V0KGBhcnRpY2xlcy8ke2lkfWApXG4gICAgICAudGhlbigocmVzKSA9PiB7XG4gICAgICAgIGRpc3BhdGNoKGFydGljbGVBY3Rpb25zLmFkZCh0cmFuc2Zvcm1SZXNwb25zZShyZXMuZGF0YSkpKVxuICAgICAgfSlcbiAgICAgIC5jYXRjaCgoZXJyKSA9PiB7XG4gICAgICAgIC8vIFRPRE86IGhhbmRsZSBlcnJcbiAgICAgICAgY29uc29sZS5lcnJvcihlcnIucmVzcG9uc2UpXG4gICAgICB9KVxuICB9XG59XG5cbmV4cG9ydCBmdW5jdGlvbiBhcnRpY2xlRmV0Y2hSZXF1ZXN0KHNsdWcpIHtcbiAgcmV0dXJuIGRpc3BhdGNoID0+IHtcbiAgICBIdHRwLmdldChgYXJ0aWNsZXMvcHVibGlzaGVkLyR7c2x1Z31gKVxuICAgICAgLnRoZW4oKHJlcykgPT4ge1xuICAgICAgICBkaXNwYXRjaChhcnRpY2xlQWN0aW9ucy5hZGQodHJhbnNmb3JtUmVzcG9uc2UocmVzLmRhdGEpKSlcbiAgICAgIH0pXG4gICAgICAuY2F0Y2goKGVycikgPT4ge1xuICAgICAgICAvLyBUT0RPOiBoYW5kbGUgZXJyXG4gICAgICAgIGNvbnNvbGUuZXJyb3IoZXJyLnJlc3BvbnNlKVxuICAgICAgfSlcbiAgfVxufSJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/article/service.js\n");
 
 /***/ }),
 
@@ -44,7 +67,43 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"add\", function() { return add; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"update\", function() { return update; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"remove\", function() { return remove; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"list\", function() { return list; });\n/* harmony import */ var _action_types__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./action-types */ \"./resources/js/modules/article/store/action-types.js\");\n/* ============\n * Actions for the article module\n * ============\n *\n * The actions that are available on the\n * article module.\n */\n\nfunction add(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_ADD\"],\n    payload: payload\n  };\n}\nfunction update(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_UPDATE\"],\n    payload: payload\n  };\n}\nfunction remove(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_REMOVE\"],\n    payload: payload\n  };\n}\nfunction list(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_LIST\"],\n    payload: payload\n  };\n}//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL3N0b3JlL2FjdGlvbnMuanM/ZjY0ZCJdLCJuYW1lcyI6WyJhZGQiLCJwYXlsb2FkIiwidHlwZSIsIkFSVElDTEVfQUREIiwidXBkYXRlIiwiQVJUSUNMRV9VUERBVEUiLCJyZW1vdmUiLCJBUlRJQ0xFX1JFTU9WRSIsImxpc3QiLCJBUlRJQ0xFX0xJU1QiXSwibWFwcGluZ3MiOiJBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBOzs7Ozs7O0FBUUE7QUFPTyxTQUFTQSxHQUFULENBQWFDLE9BQWIsRUFBc0I7QUFDM0IsU0FBTztBQUNMQyxRQUFJLEVBQUVDLHlEQUREO0FBRUxGLFdBQU8sRUFBUEE7QUFGSyxHQUFQO0FBSUQ7QUFFTSxTQUFTRyxNQUFULENBQWdCSCxPQUFoQixFQUF5QjtBQUM5QixTQUFPO0FBQ0xDLFFBQUksRUFBRUcsNERBREQ7QUFFTEosV0FBTyxFQUFQQTtBQUZLLEdBQVA7QUFJRDtBQUVNLFNBQVNLLE1BQVQsQ0FBZ0JMLE9BQWhCLEVBQXlCO0FBQzlCLFNBQU87QUFDTEMsUUFBSSxFQUFFSyw0REFERDtBQUVMTixXQUFPLEVBQVBBO0FBRkssR0FBUDtBQUlEO0FBRU0sU0FBU08sSUFBVCxDQUFjUCxPQUFkLEVBQXVCO0FBQzVCLFNBQU87QUFDTEMsUUFBSSxFQUFFTywwREFERDtBQUVMUixXQUFPLEVBQVBBO0FBRkssR0FBUDtBQUlEIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvYXJ0aWNsZS9zdG9yZS9hY3Rpb25zLmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiLyogPT09PT09PT09PT09XG4gKiBBY3Rpb25zIGZvciB0aGUgYXJ0aWNsZSBtb2R1bGVcbiAqID09PT09PT09PT09PVxuICpcbiAqIFRoZSBhY3Rpb25zIHRoYXQgYXJlIGF2YWlsYWJsZSBvbiB0aGVcbiAqIGFydGljbGUgbW9kdWxlLlxuICovXG5cbmltcG9ydCB7XG4gIEFSVElDTEVfQURELFxuICBBUlRJQ0xFX1VQREFURSxcbiAgQVJUSUNMRV9SRU1PVkUsXG4gIEFSVElDTEVfTElTVCxcbn0gZnJvbSAnLi9hY3Rpb24tdHlwZXMnO1xuXG5leHBvcnQgZnVuY3Rpb24gYWRkKHBheWxvYWQpIHtcbiAgcmV0dXJuIHtcbiAgICB0eXBlOiBBUlRJQ0xFX0FERCxcbiAgICBwYXlsb2FkXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIHVwZGF0ZShwYXlsb2FkKSB7XG4gIHJldHVybiB7XG4gICAgdHlwZTogQVJUSUNMRV9VUERBVEUsXG4gICAgcGF5bG9hZFxuICB9XG59XG5cbmV4cG9ydCBmdW5jdGlvbiByZW1vdmUocGF5bG9hZCkge1xuICByZXR1cm4ge1xuICAgIHR5cGU6IEFSVElDTEVfUkVNT1ZFLFxuICAgIHBheWxvYWRcbiAgfVxufVxuXG5leHBvcnQgZnVuY3Rpb24gbGlzdChwYXlsb2FkKSB7XG4gIHJldHVybiB7XG4gICAgdHlwZTogQVJUSUNMRV9MSVNULFxuICAgIHBheWxvYWRcbiAgfVxufSJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/article/store/actions.js\n");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"add\", function() { return add; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"update\", function() { return update; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"remove\", function() { return remove; });\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"list\", function() { return list; });\n/* harmony import */ var _action_types__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./action-types */ \"./resources/js/modules/article/store/action-types.js\");\n/* ============\n * Actions for the article module\n * ============\n *\n * The actions that are available on the\n * article module.\n */\n\nfunction add(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_ADD\"],\n    payload\n  };\n}\nfunction update(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_UPDATE\"],\n    payload\n  };\n}\nfunction remove(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_REMOVE\"],\n    payload\n  };\n}\nfunction list(payload) {\n  return {\n    type: _action_types__WEBPACK_IMPORTED_MODULE_0__[\"ARTICLE_LIST\"],\n    payload\n  };\n}//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy9hcnRpY2xlL3N0b3JlL2FjdGlvbnMuanM/ZjY0ZCJdLCJuYW1lcyI6WyJhZGQiLCJwYXlsb2FkIiwidHlwZSIsIkFSVElDTEVfQUREIiwidXBkYXRlIiwiQVJUSUNMRV9VUERBVEUiLCJyZW1vdmUiLCJBUlRJQ0xFX1JFTU9WRSIsImxpc3QiLCJBUlRJQ0xFX0xJU1QiXSwibWFwcGluZ3MiOiJBQUFBO0FBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtBQUFBOzs7Ozs7O0FBUUE7QUFPTyxTQUFTQSxHQUFULENBQWFDLE9BQWIsRUFBc0I7QUFDM0IsU0FBTztBQUNMQyxRQUFJLEVBQUVDLHlEQUREO0FBRUxGO0FBRkssR0FBUDtBQUlEO0FBRU0sU0FBU0csTUFBVCxDQUFnQkgsT0FBaEIsRUFBeUI7QUFDOUIsU0FBTztBQUNMQyxRQUFJLEVBQUVHLDREQUREO0FBRUxKO0FBRkssR0FBUDtBQUlEO0FBRU0sU0FBU0ssTUFBVCxDQUFnQkwsT0FBaEIsRUFBeUI7QUFDOUIsU0FBTztBQUNMQyxRQUFJLEVBQUVLLDREQUREO0FBRUxOO0FBRkssR0FBUDtBQUlEO0FBRU0sU0FBU08sSUFBVCxDQUFjUCxPQUFkLEVBQXVCO0FBQzVCLFNBQU87QUFDTEMsUUFBSSxFQUFFTywwREFERDtBQUVMUjtBQUZLLEdBQVA7QUFJRCIsImZpbGUiOiIuL3Jlc291cmNlcy9qcy9tb2R1bGVzL2FydGljbGUvc3RvcmUvYWN0aW9ucy5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbIi8qID09PT09PT09PT09PVxuICogQWN0aW9ucyBmb3IgdGhlIGFydGljbGUgbW9kdWxlXG4gKiA9PT09PT09PT09PT1cbiAqXG4gKiBUaGUgYWN0aW9ucyB0aGF0IGFyZSBhdmFpbGFibGUgb24gdGhlXG4gKiBhcnRpY2xlIG1vZHVsZS5cbiAqL1xuXG5pbXBvcnQge1xuICBBUlRJQ0xFX0FERCxcbiAgQVJUSUNMRV9VUERBVEUsXG4gIEFSVElDTEVfUkVNT1ZFLFxuICBBUlRJQ0xFX0xJU1QsXG59IGZyb20gJy4vYWN0aW9uLXR5cGVzJztcblxuZXhwb3J0IGZ1bmN0aW9uIGFkZChwYXlsb2FkKSB7XG4gIHJldHVybiB7XG4gICAgdHlwZTogQVJUSUNMRV9BREQsXG4gICAgcGF5bG9hZFxuICB9XG59XG5cbmV4cG9ydCBmdW5jdGlvbiB1cGRhdGUocGF5bG9hZCkge1xuICByZXR1cm4ge1xuICAgIHR5cGU6IEFSVElDTEVfVVBEQVRFLFxuICAgIHBheWxvYWRcbiAgfVxufVxuXG5leHBvcnQgZnVuY3Rpb24gcmVtb3ZlKHBheWxvYWQpIHtcbiAgcmV0dXJuIHtcbiAgICB0eXBlOiBBUlRJQ0xFX1JFTU9WRSxcbiAgICBwYXlsb2FkXG4gIH1cbn1cblxuZXhwb3J0IGZ1bmN0aW9uIGxpc3QocGF5bG9hZCkge1xuICByZXR1cm4ge1xuICAgIHR5cGU6IEFSVElDTEVfTElTVCxcbiAgICBwYXlsb2FkXG4gIH1cbn0iXSwic291cmNlUm9vdCI6IiJ9\n//# sourceURL=webpack-internal:///./resources/js/modules/article/store/actions.js\n");
+
+/***/ }),
+
+/***/ "./resources/js/modules/web/pages/blog/details/Page.js":
+/*!*************************************************************!*\
+  !*** ./resources/js/modules/web/pages/blog/details/Page.js ***!
+  \*************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(react__WEBPACK_IMPORTED_MODULE_0__);\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! prop-types */ \"./node_modules/prop-types/index.js\");\n/* harmony import */ var prop_types__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(prop_types__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var react_document_title__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! react-document-title */ \"./node_modules/react-document-title/index.js\");\n/* harmony import */ var react_document_title__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(react_document_title__WEBPACK_IMPORTED_MODULE_2__);\n/* harmony import */ var _article_service__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../../../article/service */ \"./resources/js/modules/article/service.js\");\n/* harmony import */ var _values_index__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../../../../values/index */ \"./resources/js/values/index.js\");\nfunction _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }\n\n// import libs\n\n\n\n\n\n\nclass Page extends react__WEBPACK_IMPORTED_MODULE_0__[\"Component\"] {\n  constructor(props) {\n    super(props);\n    this.state = {//\n    };\n  }\n\n  UNSAFE_componentWillMount() {\n    this.loadArticle();\n  }\n\n  loadArticle() {\n    const _this$props = this.props,\n          match = _this$props.match,\n          article = _this$props.article,\n          dispatch = _this$props.dispatch;\n\n    if (!article.slug) {\n      dispatch(Object(_article_service__WEBPACK_IMPORTED_MODULE_3__[\"articleFetchRequest\"])(match.params.slug));\n    }\n  }\n\n  renderPublishedDate() {\n    const publishedAt = this.props.article.publishedAt;\n\n    if (publishedAt) {\n      return \"at \".concat(publishedAt.format('MMMM d, YYYY'));\n    }\n  }\n\n  renderAuthor() {\n    const user = this.props.article.user;\n\n    if (user) {\n      return \"by \".concat(user.name);\n    }\n  }\n\n  createMarkup() {\n    return {\n      __html: this.props.article.content\n    };\n  }\n\n  renderArticle() {\n    const article = this.props.article;\n    return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"col-12 col-sm-9 mb-5 mx-auto\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"h2\", null, article.title), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"small\", {\n      className: \"text-muted mb-5\"\n    }, this.renderPublishedDate(), \" \", this.renderAuthor()), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"p\", {\n      className: \"text-muted mb-5\"\n    }, article.description), react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      dangerouslySetInnerHTML: this.createMarkup()\n    }));\n  }\n\n  render() {\n    return react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(react_document_title__WEBPACK_IMPORTED_MODULE_2___default.a, {\n      title: \"\".concat(this.props.article.title, \" - \").concat(_values_index__WEBPACK_IMPORTED_MODULE_4__[\"APP_TITLE\"])\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"section\", {\n      id: \"components-articles\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"container\"\n    }, react__WEBPACK_IMPORTED_MODULE_0___default.a.createElement(\"div\", {\n      className: \"row\"\n    }, this.renderArticle()))));\n  }\n\n}\n\n_defineProperty(Page, \"displayName\", 'ArticleShowPage');\n\n_defineProperty(Page, \"propTypes\", {\n  match: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.object.isRequired,\n  article: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.object.isRequired,\n  dispatch: prop_types__WEBPACK_IMPORTED_MODULE_1___default.a.func.isRequired\n});\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Page);//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy93ZWIvcGFnZXMvYmxvZy9kZXRhaWxzL1BhZ2UuanM/NmU2MiJdLCJuYW1lcyI6WyJQYWdlIiwiQ29tcG9uZW50IiwiY29uc3RydWN0b3IiLCJwcm9wcyIsInN0YXRlIiwiVU5TQUZFX2NvbXBvbmVudFdpbGxNb3VudCIsImxvYWRBcnRpY2xlIiwibWF0Y2giLCJhcnRpY2xlIiwiZGlzcGF0Y2giLCJzbHVnIiwiYXJ0aWNsZUZldGNoUmVxdWVzdCIsInBhcmFtcyIsInJlbmRlclB1Ymxpc2hlZERhdGUiLCJwdWJsaXNoZWRBdCIsImZvcm1hdCIsInJlbmRlckF1dGhvciIsInVzZXIiLCJuYW1lIiwiY3JlYXRlTWFya3VwIiwiX19odG1sIiwiY29udGVudCIsInJlbmRlckFydGljbGUiLCJ0aXRsZSIsImRlc2NyaXB0aW9uIiwicmVuZGVyIiwiQVBQX1RJVExFIiwiUHJvcFR5cGVzIiwib2JqZWN0IiwiaXNSZXF1aXJlZCIsImZ1bmMiXSwibWFwcGluZ3MiOiI7Ozs7Ozs7Ozs7O0FBQUE7QUFDQTtBQUNBO0FBQ0E7QUFDQTtBQUNBOztBQUVBLE1BQU1BLElBQU4sU0FBbUJDLCtDQUFuQixDQUE2QjtBQVEzQkMsYUFBVyxDQUFDQyxLQUFELEVBQVE7QUFDakIsVUFBTUEsS0FBTjtBQUVBLFNBQUtDLEtBQUwsR0FBYSxDQUNYO0FBRFcsS0FBYjtBQUdEOztBQUVEQywyQkFBeUIsR0FBRztBQUMxQixTQUFLQyxXQUFMO0FBQ0Q7O0FBRURBLGFBQVcsR0FBRztBQUFBLHdCQUN1QixLQUFLSCxLQUQ1QjtBQUFBLFVBQ0xJLEtBREssZUFDTEEsS0FESztBQUFBLFVBQ0VDLE9BREYsZUFDRUEsT0FERjtBQUFBLFVBQ1dDLFFBRFgsZUFDV0EsUUFEWDs7QUFHWixRQUFJLENBQUNELE9BQU8sQ0FBQ0UsSUFBYixFQUFtQjtBQUNqQkQsY0FBUSxDQUFDRSw0RUFBbUIsQ0FBQ0osS0FBSyxDQUFDSyxNQUFOLENBQWFGLElBQWQsQ0FBcEIsQ0FBUjtBQUNEO0FBQ0Y7O0FBRURHLHFCQUFtQixHQUFHO0FBQUEsVUFDYkMsV0FEYSxHQUNFLEtBQUtYLEtBQUwsQ0FBV0ssT0FEYixDQUNiTSxXQURhOztBQUdwQixRQUFJQSxXQUFKLEVBQWlCO0FBQ2YsMEJBQWFBLFdBQVcsQ0FBQ0MsTUFBWixDQUFtQixjQUFuQixDQUFiO0FBQ0Q7QUFDRjs7QUFFREMsY0FBWSxHQUFHO0FBQUEsVUFDTkMsSUFETSxHQUNFLEtBQUtkLEtBQUwsQ0FBV0ssT0FEYixDQUNOUyxJQURNOztBQUdiLFFBQUlBLElBQUosRUFBVTtBQUNSLDBCQUFhQSxJQUFJLENBQUNDLElBQWxCO0FBQ0Q7QUFFRjs7QUFFREMsY0FBWSxHQUFHO0FBQ2IsV0FBTztBQUFDQyxZQUFNLEVBQUUsS0FBS2pCLEtBQUwsQ0FBV0ssT0FBWCxDQUFtQmE7QUFBNUIsS0FBUDtBQUNEOztBQUVEQyxlQUFhLEdBQUc7QUFBQSxVQUNQZCxPQURPLEdBQ0ksS0FBS0wsS0FEVCxDQUNQSyxPQURPO0FBRWQsV0FBUTtBQUFLLGVBQVMsRUFBQztBQUFmLE9BQ04sdUVBQUtBLE9BQU8sQ0FBQ2UsS0FBYixDQURNLEVBRU47QUFBTyxlQUFTLEVBQUM7QUFBakIsT0FBb0MsS0FBS1YsbUJBQUwsRUFBcEMsT0FBaUUsS0FBS0csWUFBTCxFQUFqRSxDQUZNLEVBR047QUFBRyxlQUFTLEVBQUM7QUFBYixPQUFnQ1IsT0FBTyxDQUFDZ0IsV0FBeEMsQ0FITSxFQUlOO0FBQUssNkJBQXVCLEVBQUUsS0FBS0wsWUFBTDtBQUE5QixNQUpNLENBQVI7QUFNRDs7QUFFRE0sUUFBTSxHQUFHO0FBQ1AsV0FDRSwyREFBQywyREFBRDtBQUFlLFdBQUssWUFBSyxLQUFLdEIsS0FBTCxDQUFXSyxPQUFYLENBQW1CZSxLQUF4QixnQkFBbUNHLHVEQUFuQztBQUFwQixPQUNFO0FBQVMsUUFBRSxFQUFDO0FBQVosT0FDRTtBQUFLLGVBQVMsRUFBQztBQUFmLE9BQ0U7QUFBSyxlQUFTLEVBQUM7QUFBZixPQUNHLEtBQUtKLGFBQUwsRUFESCxDQURGLENBREYsQ0FERixDQURGO0FBV0Q7O0FBdkUwQjs7Z0JBQXZCdEIsSSxpQkFDaUIsaUI7O2dCQURqQkEsSSxlQUVlO0FBQ2pCTyxPQUFLLEVBQUVvQixpREFBUyxDQUFDQyxNQUFWLENBQWlCQyxVQURQO0FBRWpCckIsU0FBTyxFQUFFbUIsaURBQVMsQ0FBQ0MsTUFBVixDQUFpQkMsVUFGVDtBQUdqQnBCLFVBQVEsRUFBRWtCLGlEQUFTLENBQUNHLElBQVYsQ0FBZUQ7QUFIUixDOztBQXdFTjdCLG1FQUFmIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvd2ViL3BhZ2VzL2Jsb2cvZGV0YWlscy9QYWdlLmpzLmpzIiwic291cmNlc0NvbnRlbnQiOlsiLy8gaW1wb3J0IGxpYnNcbmltcG9ydCBSZWFjdCwge0NvbXBvbmVudH0gZnJvbSAncmVhY3QnXG5pbXBvcnQgUHJvcFR5cGVzIGZyb20gJ3Byb3AtdHlwZXMnXG5pbXBvcnQgRG9jdW1lbnRUaXRsZSBmcm9tICdyZWFjdC1kb2N1bWVudC10aXRsZSc7XG5pbXBvcnQge2FydGljbGVGZXRjaFJlcXVlc3R9IGZyb20gJy4uLy4uLy4uLy4uL2FydGljbGUvc2VydmljZSdcbmltcG9ydCB7QVBQX1RJVExFfSBmcm9tICcuLi8uLi8uLi8uLi8uLi92YWx1ZXMvaW5kZXgnXG5cbmNsYXNzIFBhZ2UgZXh0ZW5kcyBDb21wb25lbnQge1xuICBzdGF0aWMgZGlzcGxheU5hbWUgPSAnQXJ0aWNsZVNob3dQYWdlJ1xuICBzdGF0aWMgcHJvcFR5cGVzID0ge1xuICAgIG1hdGNoOiBQcm9wVHlwZXMub2JqZWN0LmlzUmVxdWlyZWQsXG4gICAgYXJ0aWNsZTogUHJvcFR5cGVzLm9iamVjdC5pc1JlcXVpcmVkLFxuICAgIGRpc3BhdGNoOiBQcm9wVHlwZXMuZnVuYy5pc1JlcXVpcmVkXG4gIH1cblxuICBjb25zdHJ1Y3Rvcihwcm9wcykge1xuICAgIHN1cGVyKHByb3BzKVxuXG4gICAgdGhpcy5zdGF0ZSA9IHtcbiAgICAgIC8vXG4gICAgfVxuICB9XG5cbiAgVU5TQUZFX2NvbXBvbmVudFdpbGxNb3VudCgpIHtcbiAgICB0aGlzLmxvYWRBcnRpY2xlKClcbiAgfVxuXG4gIGxvYWRBcnRpY2xlKCkge1xuICAgIGNvbnN0IHttYXRjaCwgYXJ0aWNsZSwgZGlzcGF0Y2h9ID0gdGhpcy5wcm9wc1xuXG4gICAgaWYgKCFhcnRpY2xlLnNsdWcpIHtcbiAgICAgIGRpc3BhdGNoKGFydGljbGVGZXRjaFJlcXVlc3QobWF0Y2gucGFyYW1zLnNsdWcpKVxuICAgIH1cbiAgfVxuXG4gIHJlbmRlclB1Ymxpc2hlZERhdGUoKSB7XG4gICAgY29uc3Qge3B1Ymxpc2hlZEF0fSA9IHRoaXMucHJvcHMuYXJ0aWNsZVxuXG4gICAgaWYgKHB1Ymxpc2hlZEF0KSB7XG4gICAgICByZXR1cm4gYGF0ICR7cHVibGlzaGVkQXQuZm9ybWF0KCdNTU1NIGQsIFlZWVknKX1gXG4gICAgfVxuICB9XG5cbiAgcmVuZGVyQXV0aG9yKCkge1xuICAgIGNvbnN0IHt1c2VyfSA9IHRoaXMucHJvcHMuYXJ0aWNsZVxuXG4gICAgaWYgKHVzZXIpIHtcbiAgICAgIHJldHVybiBgYnkgJHt1c2VyLm5hbWV9YFxuICAgIH1cblxuICB9XG5cbiAgY3JlYXRlTWFya3VwKCkge1xuICAgIHJldHVybiB7X19odG1sOiB0aGlzLnByb3BzLmFydGljbGUuY29udGVudH07XG4gIH1cblxuICByZW5kZXJBcnRpY2xlKCkge1xuICAgIGNvbnN0IHthcnRpY2xlfSA9IHRoaXMucHJvcHNcbiAgICByZXR1cm4gKDxkaXYgY2xhc3NOYW1lPVwiY29sLTEyIGNvbC1zbS05IG1iLTUgbXgtYXV0b1wiPlxuICAgICAgPGgyPnthcnRpY2xlLnRpdGxlfTwvaDI+XG4gICAgICA8c21hbGwgY2xhc3NOYW1lPVwidGV4dC1tdXRlZCBtYi01XCI+e3RoaXMucmVuZGVyUHVibGlzaGVkRGF0ZSgpfSB7dGhpcy5yZW5kZXJBdXRob3IoKX08L3NtYWxsPlxuICAgICAgPHAgY2xhc3NOYW1lPVwidGV4dC1tdXRlZCBtYi01XCI+e2FydGljbGUuZGVzY3JpcHRpb259PC9wPlxuICAgICAgPGRpdiBkYW5nZXJvdXNseVNldElubmVySFRNTD17dGhpcy5jcmVhdGVNYXJrdXAoKX0vPlxuICAgIDwvZGl2PilcbiAgfVxuXG4gIHJlbmRlcigpIHtcbiAgICByZXR1cm4gKFxuICAgICAgPERvY3VtZW50VGl0bGUgdGl0bGU9e2Ake3RoaXMucHJvcHMuYXJ0aWNsZS50aXRsZX0gLSAke0FQUF9USVRMRX1gfT5cbiAgICAgICAgPHNlY3Rpb24gaWQ9XCJjb21wb25lbnRzLWFydGljbGVzXCI+XG4gICAgICAgICAgPGRpdiBjbGFzc05hbWU9XCJjb250YWluZXJcIj5cbiAgICAgICAgICAgIDxkaXYgY2xhc3NOYW1lPVwicm93XCI+XG4gICAgICAgICAgICAgIHt0aGlzLnJlbmRlckFydGljbGUoKX1cbiAgICAgICAgICAgIDwvZGl2PlxuICAgICAgICAgIDwvZGl2PlxuICAgICAgICA8L3NlY3Rpb24+XG4gICAgICA8L0RvY3VtZW50VGl0bGU+XG4gICAgKVxuICB9XG59XG5cbmV4cG9ydCBkZWZhdWx0IFBhZ2VcbiJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/web/pages/blog/details/Page.js\n");
+
+/***/ }),
+
+/***/ "./resources/js/modules/web/pages/blog/details/index.js":
+/*!**************************************************************!*\
+  !*** ./resources/js/modules/web/pages/blog/details/index.js ***!
+  \**************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var react_redux__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react-redux */ \"./node_modules/react-redux/es/index.js\");\n/* harmony import */ var _article_Article__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../../../../article/Article */ \"./resources/js/modules/article/Article.js\");\n/* harmony import */ var _Page__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Page */ \"./resources/js/modules/web/pages/blog/details/Page.js\");\n\n // import components\n\n\n\nconst mapStateToProps = (state, router) => {\n  const params = router.match.params;\n  const article = state.articles.data.find(article => article.slug === params.slug);\n  return {\n    article: article ? new _article_Article__WEBPACK_IMPORTED_MODULE_1__[\"default\"](article) : new _article_Article__WEBPACK_IMPORTED_MODULE_1__[\"default\"]({})\n  };\n};\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (Object(react_redux__WEBPACK_IMPORTED_MODULE_0__[\"connect\"])(mapStateToProps)(_Page__WEBPACK_IMPORTED_MODULE_2__[\"default\"]));//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvbW9kdWxlcy93ZWIvcGFnZXMvYmxvZy9kZXRhaWxzL2luZGV4LmpzP2RkMzAiXSwibmFtZXMiOlsibWFwU3RhdGVUb1Byb3BzIiwic3RhdGUiLCJyb3V0ZXIiLCJwYXJhbXMiLCJtYXRjaCIsImFydGljbGUiLCJhcnRpY2xlcyIsImRhdGEiLCJmaW5kIiwic2x1ZyIsIkFydGljbGUiLCJjb25uZWN0IiwiUGFnZSJdLCJtYXBwaW5ncyI6IkFBQUE7QUFBQTtBQUFBO0FBQUE7QUFBQTtDQUdBOztBQUNBOztBQUVBLE1BQU1BLGVBQWUsR0FBRyxDQUFDQyxLQUFELEVBQVFDLE1BQVIsS0FBbUI7QUFBQSxRQUNqQ0MsTUFEaUMsR0FDdEJELE1BQU0sQ0FBQ0UsS0FEZSxDQUNqQ0QsTUFEaUM7QUFFekMsUUFBTUUsT0FBTyxHQUFHSixLQUFLLENBQUNLLFFBQU4sQ0FBZUMsSUFBZixDQUFvQkMsSUFBcEIsQ0FBeUJILE9BQU8sSUFBSUEsT0FBTyxDQUFDSSxJQUFSLEtBQWlCTixNQUFNLENBQUNNLElBQTVELENBQWhCO0FBQ0EsU0FBTztBQUNMSixXQUFPLEVBQUVBLE9BQU8sR0FBRyxJQUFJSyx3REFBSixDQUFZTCxPQUFaLENBQUgsR0FBMEIsSUFBSUssd0RBQUosQ0FBWSxFQUFaO0FBRHJDLEdBQVA7QUFHRCxDQU5EOztBQVFlQywwSEFBTyxDQUFDWCxlQUFELENBQVAsQ0FBeUJZLDZDQUF6QixDQUFmIiwiZmlsZSI6Ii4vcmVzb3VyY2VzL2pzL21vZHVsZXMvd2ViL3BhZ2VzL2Jsb2cvZGV0YWlscy9pbmRleC5qcy5qcyIsInNvdXJjZXNDb250ZW50IjpbImltcG9ydCB7IGNvbm5lY3QgfSBmcm9tICdyZWFjdC1yZWR1eCdcbmltcG9ydCBBcnRpY2xlIGZyb20gJy4uLy4uLy4uLy4uL2FydGljbGUvQXJ0aWNsZSdcblxuLy8gaW1wb3J0IGNvbXBvbmVudHNcbmltcG9ydCBQYWdlIGZyb20gJy4vUGFnZSdcblxuY29uc3QgbWFwU3RhdGVUb1Byb3BzID0gKHN0YXRlLCByb3V0ZXIpID0+IHtcbiAgY29uc3QgeyBwYXJhbXMgfSA9IHJvdXRlci5tYXRjaFxuICBjb25zdCBhcnRpY2xlID0gc3RhdGUuYXJ0aWNsZXMuZGF0YS5maW5kKGFydGljbGUgPT4gYXJ0aWNsZS5zbHVnID09PSBwYXJhbXMuc2x1ZylcbiAgcmV0dXJuIHtcbiAgICBhcnRpY2xlOiBhcnRpY2xlID8gbmV3IEFydGljbGUoYXJ0aWNsZSkgOiBuZXcgQXJ0aWNsZSh7fSlcbiAgfVxufVxuXG5leHBvcnQgZGVmYXVsdCBjb25uZWN0KG1hcFN0YXRlVG9Qcm9wcykoUGFnZSlcbiJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/modules/web/pages/blog/details/index.js\n");
+
+/***/ }),
+
+/***/ "./resources/js/values/index.js":
+/*!**************************************!*\
+  !*** ./resources/js/values/index.js ***!
+  \**************************************/
+/*! exports provided: APP_TITLE */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"APP_TITLE\", function() { return APP_TITLE; });\nconst APP_TITLE = 'Your Blog Name';\n//# sourceURL=[module]\n//# sourceMappingURL=data:application/json;charset=utf-8;base64,eyJ2ZXJzaW9uIjozLCJzb3VyY2VzIjpbIndlYnBhY2s6Ly8vLi9yZXNvdXJjZXMvanMvdmFsdWVzL2luZGV4LmpzP2NhYzQiXSwibmFtZXMiOlsiQVBQX1RJVExFIl0sIm1hcHBpbmdzIjoiQUFBQTtBQUFBO0FBQUEsTUFBTUEsU0FBUyxHQUFHLGdCQUFsQiIsImZpbGUiOiIuL3Jlc291cmNlcy9qcy92YWx1ZXMvaW5kZXguanMuanMiLCJzb3VyY2VzQ29udGVudCI6WyJjb25zdCBBUFBfVElUTEUgPSAnWW91ciBCbG9nIE5hbWUnXG5cbmV4cG9ydCB7XG4gICAgQVBQX1RJVExFLFxufSJdLCJzb3VyY2VSb290IjoiIn0=\n//# sourceURL=webpack-internal:///./resources/js/values/index.js\n");
 
 /***/ })
 
