@@ -11,7 +11,7 @@ const PublicRoutes = ({ component: Component, ...rest }) => {
 }
 
 PublicRoutes.propTypes = {
-  // component: PropTypes.func.isRequired,
+  component: PropTypes.object.isRequired,
   location: PropTypes.object,
 };
 
